% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bangor.R
\docType{data}
\name{bangor_1}
\alias{bangor_1}
\title{Primary corporate colours of the University of Bangor, Wales}
\format{
An object of class \code{character} of length 4.
}
\source{
Colour definitions are based on \href{https://www.bangor.ac.uk}{Bangor}'s 
Brand guidelines (PDF).
}
\usage{
bangor_1
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{bangor_1} provides the four primary corporate colours 
of the \href{https://www.bangor.ac.uk}{University of Bangor}, Wales, UK.
}
\details{
The 4 primary colours of the University of Bangor are 
etc. 

\code{bangor_1} uses the HEX color definitions.
}
\examples{
bangor_1
unikn::seecol(bangor_1, col_bg = "grey90",  
              main = "Primary colours of the University of Bangor, Wales")  # view palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other Welsh university color palettes: 
\code{\link{aberystwyth_1}},
\code{\link{aberystwyth_2}},
\code{\link{cardiff_1}},
\code{\link{cardiff_2}}
}
\author{
\strong{unicol}, 2023-10-09.
}
\concept{Welsh university color palettes}
\keyword{datasets}
