% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Birmingham.R
\docType{data}
\name{birmingham_1}
\alias{birmingham_1}
\title{colours of the University of Birmingham}
\format{
An object of class \code{character} of length 6.
}
\source{
colour definitions are based on the
\href{https://intranet.birmingham.ac.uk/social-sciences/schools/business/documents/public/brand-templates/11687-REVISED-BBS-BRAND-GUIDELINES-FINAL-FEB-2016.pdf}{CD manual (PDF)}.
}
\usage{
birmingham_1
}
\value{
A named vector of colours (HEX/HTML codes of type character).
}
\description{
\code{birmingham_1} provides the three primary colours 
of the \href{https://www.birmingham.ac.uk/index.aspx}{University of Birmingham}, UK.
}
\details{
The three primary colours are
\code{"undergraduate"} (defined as HEX #c1d82f),
\code{"postgraduate"} (defined as HEX #c1d82f),
\code{"MBA"} (defined as HEX #7D0049), and
\code{"Research/Business"} (defined as HEX #605270) .


The two secondary colours are
 \code{"Secondary colour 1"} (defined as HEX #bfb6ad) and
\code{"Secondary colour 2"} (defined as HEX #757477) .

\code{birmingham_1} uses the HEX colour definition.
}
\examples{
birmingham_1
unikn::seecol(birmingham_1, main = "University of Birmingham") 
# view colour palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing colour palettes;
\code{\link{usecol}} for using colour palettes;
\code{\link{simcol}} for finding similar colours;
\code{\link{newpal}} for defining new colour palettes;
\code{\link{grepal}} for finding named colours.

Other university colour palettes: 
\code{\link{dundee_background}},
\code{\link{dundee_block}},
\code{\link{dundee_core}},
\code{\link{dundee_highlight}},
\code{\link{mcgill_brights}},
\code{\link{mcgill_darks}},
\code{\link{mcgill_grey}},
\code{\link{mcgill_muted}},
\code{\link{mcgill_pastels}},
\code{\link{mcgill_red}},
\code{\link{mcmaster_brighterworld}},
\code{\link{mcmaster_heritage}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{queens_primary}},
\code{\link{queens_secondary}},
\code{\link{sfu_brand}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_primary}},
\code{\link{ucalgary_secondary}},
\code{\link{ucalgary_warmgreys}},
\code{\link{uvic_complementary}},
\code{\link{uvic_core}}
}
\author{
\strong{unicol}, 2023-06-20.
}
\concept{university colour palettes}
\keyword{datasets}
