% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BU.R
\docType{data}
\name{bu_1}
\alias{bu_1}
\title{Primary colours of Boston University}
\format{
An object of class \code{character} of length 2.
}
\source{
colour definitions are based on the
\href{https://www.bu.edu/brand/guidelines/design/colors/}{BU color guideline}.
}
\usage{
bu_1
}
\value{
A named vector of colours (HEX/HTML codes of type character).
}
\description{
\code{bu_1} provides the three Primary colours 
of the \href{https://www.bu.edu}{Boston University}, USA.
}
\details{
The two Primary colours are
\code{"BU Red"} (defined as RGB - 204 0 0, CMYK - 0 100 75 4, HEX - CC0000) and
\code{"Black"} (defined as RGB - 45 41 38, CMYK - 63 62 59 94, HEX - 2D2926).


\code{bu_1} uses the HEX colour definition.
}
\examples{
bu_1
unikn::seecol(bu_1, main = "Primary colours of Boston University", col_bg = "white") 
# view colour palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing colour palettes;
\code{\link{usecol}} for using colour palettes;
\code{\link{simcol}} for finding similar colours;
\code{\link{newpal}} for defining new colour palettes;
\code{\link{grepal}} for finding named colours.

Other university colour palettes: 
\code{\link{birmingham_1}},
\code{\link{dundee_background}},
\code{\link{dundee_block}},
\code{\link{dundee_core}},
\code{\link{dundee_highlight}},
\code{\link{lancaster_1}},
\code{\link{lancaster_2}},
\code{\link{mcgill_brights}},
\code{\link{mcgill_darks}},
\code{\link{mcgill_grey}},
\code{\link{mcgill_muted}},
\code{\link{mcgill_pastels}},
\code{\link{mcgill_red}},
\code{\link{mcmaster_brighterworld}},
\code{\link{mcmaster_heritage}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{queens_primary}},
\code{\link{queens_secondary}},
\code{\link{sfu_brand}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_primary}},
\code{\link{ucalgary_secondary}},
\code{\link{ucalgary_warmgreys}},
\code{\link{uni_heidelberg_1}},
\code{\link{uvic_complementary}},
\code{\link{uvic_core}}
}
\author{
\strong{unicol}, 2023-07-04.
}
\concept{university colour palettes}
\keyword{datasets}
