% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carleton.R
\docType{data}
\name{carleton_1}
\alias{carleton_1}
\title{Colors of the University of Carleton, Canada}
\format{
An object of class \code{character} of length 2.
}
\source{
Color definitions are based on the
\href{https://www.carleton.edu/communications/resources/college-wordmark-identity-graphics/college-colors/}{University of Carleton's website}.
}
\usage{
carleton_1
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{carleton_1} provides the two official colors 
of the \href{https://carleton.ca/}{University of Carleton}, Canada.
}
\details{
The two official colors are
\code{maize} (defined as Pantone 122, 
and approximated by RGB 255/223/79, or HEX #FFDF4F), and
\code{blue_carleton} (defined as Pantone 294, 
and approximated by RGB 0/48/105, or HEX #003069).

\code{carleton_1} uses the HEX color definitions.
}
\examples{
carleton_1
unikn::seecol(carleton_1, 
              main = "Colors of the University of Carleton") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.
}
\author{
\strong{unicol}, 2023-07-12.
}
\concept{Candadian university color palettes}
\keyword{datasets}
