% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Columbia.R
\docType{data}
\name{columbia}
\alias{columbia}
\title{Primary and secondary blue colors of Columbia University, USA}
\format{
An object of class \code{character} of length 5.
}
\source{
Color definitions are based on the
\href{https://visualidentity.columbia.edu/content/colors-1}{Color manual}.
}
\usage{
columbia
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{columbia} provides one primary blue and four secondary blue colors
of the \href{http://www.columbia.edu}{Columbia University}, USA.
}
\details{
The official and primary blue color is 
\code{"columbia_blue"} (PMS 290C, RGB 185/217/235, CMYK 23/0/1/0).
It is based on the color of the Philolexian Society, Columbia’s oldest student organization, founded in 1802.

The Columbia University trademark can also be rendered in these approved blues:
\code{"sec_blue1"} (defined as PMS 284C, RGB 108/172/228, CMYK 59/17/0/0),
\code{"sec_blue2"} (defined as PMS 285C, RGB 0/114/206, CMYK 90/48/0/0),
\code{"sec_blue3"} (defined as PMS 286C, RGB 0/51/160, CMYK 100/75/0/0), and
\code{"sec_blue4"} (defined as PMS 280C, RGB 2/33/105, CMYK 100/85/5/22).

\code{columbia} uses the RGB color definitions.
}
\examples{
columbia
unikn::seecol(columbia, main = "Primary colors of Columbia University") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other U.S. university color palettes: 
\code{\link{CMU_1_core}},
\code{\link{CMU_2_tartan}},
\code{\link{CMU_3_campus}},
\code{\link{asu_1}},
\code{\link{berkeley_1}},
\code{\link{berkeley_2}},
\code{\link{brown_1}},
\code{\link{brown_2}},
\code{\link{bu_1}},
\code{\link{caltech_1}},
\code{\link{caltech_2}},
\code{\link{caltech_3}},
\code{\link{chicago_1}},
\code{\link{chicago_2}},
\code{\link{cornell_1}},
\code{\link{cornell_2}},
\code{\link{cornell_3_accent}},
\code{\link{monash_1}},
\code{\link{monash_2}}
}
\author{
\strong{unicol}, 2023-06-14.
}
\concept{U.S. university color palettes}
\keyword{datasets}
