% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concordia.R
\docType{data}
\name{conc_2}
\alias{conc_2}
\title{Secondary colors of Concordia University, Canada}
\format{
An object of class \code{character} of length 4.
}
\source{
Color definitions are based on 
\href{https://www.concordia.ca/web/design/ui-kit-style-guide-accessibility/web-palette-new.html}{Concordia's NEW web colour palette}.
}
\usage{
conc_2
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{conc_2} provides the four secondary colors 
of the \href{https://www.concordia.ca/}{Concordia University}, Canada.
}
\details{
The four secondary colors are
\code{"yellow_concordia"} (defined as HEX #e5a712),
\code{"gold_concordia"}  (defined as HEX #cbb576), 
\code{"cyan_concordia"}  (defined as HEX #00adef), and
\code{"lime"}            (defined as HEX #8cc63e).

\code{conc_2} uses the HEX color definitions.
}
\examples{
conc_2
unikn::seecol(conc_2, main = "Concordia University") # view color palette

}
\seealso{
\code{\link{conc_1}} for primary colors of Concordia University;
\code{\link{conc_3}} for background colors of Concordia University;
\code{\link{conc_4}} for neutral colors of Concordia University;
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other Canadian university color palettes: 
\code{\link{conc_1}},
\code{\link{conc_3}},
\code{\link{conc_4}},
\code{\link{guelph}},
\code{\link{laval}},
\code{\link{manitoba_1}},
\code{\link{manitoba_2}},
\code{\link{mcgill_brights}},
\code{\link{mcgill_darks}},
\code{\link{mcgill_grey}},
\code{\link{mcgill_muted}},
\code{\link{mcgill_pastels}},
\code{\link{mcgill_red}},
\code{\link{mcmaster_brighterworld}},
\code{\link{mcmaster_heritage}},
\code{\link{mun_1}},
\code{\link{mun_2}},
\code{\link{ottawa_1}},
\code{\link{toronto_1}},
\code{\link{ubc_1}},
\code{\link{uvic_complementary}},
\code{\link{uvic_core}}
}
\author{
\strong{unicol}, 2023-06-26.
}
\concept{Canadian university color palettes}
\keyword{datasets}
