% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concordia.R
\docType{data}
\name{concordia_1}
\alias{concordia_1}
\title{Primary colours of Concordia University, Canada}
\format{
An object of class \code{character} of length 8.
}
\source{
Color definitions are based on \href{https://www.concordia.ca}{Concordia}'s 
\href{https://www.concordia.ca/web/design/ui-kit-style-guide-accessibility/web-palette-new.html}{web colour palette}.
}
\usage{
concordia_1
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{concordia_1} provides the eight primary colours 
of \href{https://www.concordia.ca}{Concordia University}, Canada.
}
\details{
The 8 primary colours are
\code{"Burgundy"} (defined as HEX #912338), 
\code{"Magenta"} (defined as HEX #db0272), 
\code{"Orange"} (defined as HEX #da3a16),
\code{"Mauve"} (defined as HEX #573996),
\code{"Dark"} (defined as HEX #004085),
\code{"Blue"} (defined as HEX #0072a8),
\code{"Turquoise"} (defined as HEX #057d78), and
\code{"Grenn"} (defined as HEX #508212).

\code{concordia_1} uses the HEX color definitions.
}
\examples{
concordia_1
unikn::seecol(concordia_1, main = "Primary colours of Concordia University, CA")  # view palette

}
\seealso{
\code{\link{concordia_2}} for secondary colours of Concordia University;
\code{\link{concordia_3}} for background tints of Concordia University;
\code{\link{concordia_4}} for neutral colours of Concordia University;
\code{\link[unikn]{seecol}} for viewing and comparing color palettes;
\code{\link[unikn]{usecol}} for using color palettes;
\code{\link[unikn]{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link[unikn]{grepal}} for finding named colors.

Other Canadian university color palettes: 
\code{\link{carleton}},
\code{\link{concordia_2}},
\code{\link{concordia_3}},
\code{\link{concordia_4}},
\code{\link{guelph}},
\code{\link{laval}},
\code{\link{manitoba_1}},
\code{\link{manitoba_2}},
\code{\link{mcgill_brights}},
\code{\link{mcgill_darks}},
\code{\link{mcgill_grey}},
\code{\link{mcgill_muted}},
\code{\link{mcgill_pastels}},
\code{\link{mcgill_red}},
\code{\link{mcmaster_brighterworld}},
\code{\link{mcmaster_heritage}},
\code{\link{mun_1}},
\code{\link{mun_2}},
\code{\link{queensu_1}},
\code{\link{queensu_2}},
\code{\link{sfu_brand}},
\code{\link{ubc}},
\code{\link{ucalgary_1}},
\code{\link{ucalgary_2}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_warmgreys}},
\code{\link{uottawa}},
\code{\link{utoronto}},
\code{\link{uvic_1}},
\code{\link{uvic_2}},
\code{\link{uwaterloo_arts}},
\code{\link{uwaterloo_engineering}},
\code{\link{uwaterloo_environment}},
\code{\link{uwaterloo_health}},
\code{\link{uwaterloo_main}},
\code{\link{uwaterloo_math}},
\code{\link{uwaterloo_science}},
\code{\link{western_uni}},
\code{\link{yorku_1}},
\code{\link{yorku_2}}
}
\author{
\strong{unicol}, 2023-06-26.
}
\concept{Canadian university color palettes}
\keyword{datasets}
