% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dartmouth.R
\docType{data}
\name{dartmouth_2}
\alias{dartmouth_2}
\title{Secondary colors of Dartmouth College, USA}
\format{
An object of class \code{character} of length 8.
}
\source{
Color definitions are based on \href{https://home.dartmouth.edu}{Dartmouth}'s 
\href{https://communications.dartmouth.edu/guides-and-tools/design-guidelines/dartmouth-colors}{visual identity guide}.
}
\usage{
dartmouth_2
}
\value{
A names vector of colors (HEX/HTML codes of type character).
}
\description{
\code{dartmouth_2} provides the eight secondary colors
of \href{https://home.dartmouth.edu}{Dartmouth College}, USA.
}
\details{
The 8 secondary colors are 
\code{"forest green"} (defined as HEX #12312b),
\code{"rich forest green"} (defined as HEX #0D1E1C),
\code{"snow white"} (identical to \code{"white"} and defined as HEX #ffffff),
\code{"midnight black"} (defined as HEX #000000)
\code{"web gray 1"} (defined as HEX #f7f7f7)
\code{"web gray 2"} (defined as HEX #e2e2e2)
\code{"web gray 3"} (defined as HEX #707070), and 
\code{"granite gray"} (defined as HEX #424141). 

\code{dartmouth_2} uses the HEX color definitions.
}
\examples{
dartmouth_2
unikn::seecol(dartmouth_2, main = "Secondary colors of Dartmouth College")  # view palette

}
\seealso{
\code{\link{dartmouth_1}} for primary color of Dartmouth College;
\code{\link{dartmouth_3}} for tertiary colors of Dartmouth College; 
\code{\link[unikn]{seecol}} for viewing and comparing color palettes;
\code{\link[unikn]{usecol}} for using color palettes;
\code{\link[unikn]{simcol}} for finding similar colors;
\code{\link[unikn]{newpal}} for defining new color palettes;
\code{\link[unikn]{grepal}} for finding named colors.

Other U.S. university color palettes: 
\code{\link{IU_1}},
\code{\link{IU_2}},
\code{\link{IU_2_dp}},
\code{\link{asu_1}},
\code{\link{asu_2}},
\code{\link{asu_3}},
\code{\link{berkeley_1}},
\code{\link{berkeley_2}},
\code{\link{brown_1}},
\code{\link{brown_2}},
\code{\link{bu}},
\code{\link{caltech_1}},
\code{\link{caltech_2}},
\code{\link{caltech_3}},
\code{\link{cmu_1}},
\code{\link{cmu_2}},
\code{\link{cmu_3}},
\code{\link{columbia_1}},
\code{\link{columbia_2}},
\code{\link{columbia_3}},
\code{\link{cornell_1}},
\code{\link{cornell_2}},
\code{\link{cornell_3_accent}},
\code{\link{dartmouth_1}},
\code{\link{dartmouth_3}},
\code{\link{duke_1}},
\code{\link{duke_2}},
\code{\link{harvard_1}},
\code{\link{harvard_2}},
\code{\link{harvard_3}},
\code{\link{jhu_0}},
\code{\link{jhu_1}},
\code{\link{jhu_2}},
\code{\link{jhu_3_accent}},
\code{\link{jhu_4_gray}},
\code{\link{manchester_uni_1}},
\code{\link{manchester_uni_2}},
\code{\link{michigan_1}},
\code{\link{michigan_2}},
\code{\link{minnesotatwin_1}},
\code{\link{minnesotatwin_2}},
\code{\link{mit}},
\code{\link{monash_1}},
\code{\link{monash_2}},
\code{\link{msu}},
\code{\link{northwestern_1}},
\code{\link{northwestern_2}},
\code{\link{notredame_1}},
\code{\link{notredame_2}},
\code{\link{nyu_1}},
\code{\link{nyu_2}},
\code{\link{nyu_accent}},
\code{\link{nyu_neutral}},
\code{\link{ohio_uni_1}},
\code{\link{ohio_uni_2}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{princeton_0}},
\code{\link{princeton_1}},
\code{\link{princeton_2}},
\code{\link{rpi_1}},
\code{\link{rpi_2}},
\code{\link{rpi_3}},
\code{\link{stanford_1}},
\code{\link{stanford_2}},
\code{\link{stanford_3}},
\code{\link{uchicago_1}},
\code{\link{uchicago_2}},
\code{\link{uci_1}},
\code{\link{uci_2}},
\code{\link{uci_3}},
\code{\link{uci_4}},
\code{\link{ucla_1}},
\code{\link{ucla_2}},
\code{\link{ucla_3}},
\code{\link{ucla_4}},
\code{\link{ucsd_1}},
\code{\link{ucsd_2}},
\code{\link{ucsd_3}},
\code{\link{uflorida_1}},
\code{\link{uflorida_2}},
\code{\link{uflorida_3}},
\code{\link{umass_1}},
\code{\link{umass_2}},
\code{\link{umass_neutrals}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{willamette}},
\code{\link{wm_1}},
\code{\link{wm_2}},
\code{\link{wm_3}},
\code{\link{yale}},
\code{\link{yeshiva}}
}
\author{
\strong{unicol}, 2320-07-13.
}
\concept{U.S. university color palettes}
\keyword{datasets}
