% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duke.R
\docType{data}
\name{duke_2}
\alias{duke_2}
\title{Secondary colors of Duke University, USA}
\format{
An object of class \code{character} of length 18.
}
\source{
Color definitions are based on \href{https://duke.edu}{Duke}'s 
\href{https://brand.duke.edu/colors/}{Color manual}.
}
\usage{
duke_2
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{duke_2} provides the 18 secondary colors
of \href{https://duke.edu}{Duke University}, USA.
}
\details{
These 18 colors are
\code{"Copper"} (defined as CMYK 0/76/100/0, RGB 200/78/0, HEX #C84E00),
\code{"Persimmon"} (defined as CMYK 0/45/95/0, RGB 232/153/35, HEX #E89923),
\code{"Dandelion"} (defined as CMYK 0/8/70/0, RGB 255/217/96, HEX #FFD960),
\code{"Piedmont"} (defined as CMYK 54/0/100/0, RGB 161/183/13, HEX #A1B70D),
\code{"Eno"} (defined as CMYK 81/0/39/0, RGB 51/152/152, HEX #339898),
\code{"Magnolia"} (defined as CMYK 96/16/42/57, RGB 29/99/99, HEX #1D6363),
\code{"Prussian blue"} (defined as CMYK 100/45/0/45, RGB 0/85/135, HEX #005587),
\code{"Shale blue"} (defined as CMYK 100/0/1/3, RGB 5/119/177, HEX #0577B1),
\code{"Ironweed"} (defined as CMYK 35/95/0/0, RGB 153/51/153, HEX #993399),
\code{"Hatteras"} (defined as CMYK 10/2/0/0, RGB 226/230/237, HEX #E2E6ED),
\code{"Whisper gray"} (defined as CMYK 4/2/4/8, RGB 243/242/241, HEX #F3F2F1),
\code{"Ginger beer"} (defined as CMYK 0/2/15/0, RGB 252/247/229, HEX #FCF7E5),
\code{"Dogwood"} (defined as CMYK 10/18/25/32, RGB 152/134/117, HEX #988675),
\code{"Shackleford"} (defined as CMYK 3/4/14/8, RGB 218/208/198, HEX #DAD0C6),
\code{"Cast iron"} (defined as CMYK 67/44/67/95, RGB 38/38/38, HEX #262626),
\code{"Graphite"} (defined as CMYK 40/30/20/66, RGB 102/102/102, HEX #666666),
\code{"Granite"} (defined as CMYK 13/8/11/26, RGB 181/181/181, HEX #B5B5B5), and
\code{"Limestone"} (defined as CMYK 5/3/5/11, RGB 229/229/229, HEX #E5E5E5).


The colors in Duke’s extended palette are intended for use as secondary and tertiary colors in design projects. 

They were selected to complement \code{Duke navy blue} and should be used  for a range of elements, 
including graphic accents, typography, backgrounds, call-to-action buttons and more.
}
\examples{
duke_2
unikn::seecol(duke_2, main = "The secondary colors of Duke University") # view palette

}
\seealso{
\code{\link{duke_1}} for primary colors of Duke University;
\code{\link[unikn]{seecol}} for viewing and comparing color palettes;
\code{\link[unikn]{usecol}} for using color palettes;
\code{\link[unikn]{simcol}} for finding similar colors;
\code{\link[unikn]{newpal}} for defining new color palettes;
\code{\link[unikn]{grepal}} for finding named colors.

Other U.S. university color palettes: 
\code{\link{IU_1}},
\code{\link{IU_2}},
\code{\link{IU_2_dp}},
\code{\link{asu_1}},
\code{\link{asu_2}},
\code{\link{asu_3}},
\code{\link{berkeley_1}},
\code{\link{berkeley_2}},
\code{\link{brown_1}},
\code{\link{brown_2}},
\code{\link{bu}},
\code{\link{caltech_1}},
\code{\link{caltech_2}},
\code{\link{caltech_3}},
\code{\link{cmu_1}},
\code{\link{cmu_2}},
\code{\link{cmu_3}},
\code{\link{columbia_1}},
\code{\link{columbia_2}},
\code{\link{columbia_3}},
\code{\link{cornell_1}},
\code{\link{cornell_2}},
\code{\link{cornell_3_accent}},
\code{\link{dartmouth_1}},
\code{\link{dartmouth_2}},
\code{\link{dartmouth_3}},
\code{\link{duke_1}},
\code{\link{harvard_1}},
\code{\link{harvard_2}},
\code{\link{harvard_3}},
\code{\link{jhu_0}},
\code{\link{jhu_1}},
\code{\link{jhu_2}},
\code{\link{jhu_3_accent}},
\code{\link{jhu_4_gray}},
\code{\link{manchester_uni_1}},
\code{\link{manchester_uni_2}},
\code{\link{michigan_1}},
\code{\link{michigan_2}},
\code{\link{minnesotatwin_1}},
\code{\link{minnesotatwin_2}},
\code{\link{mit}},
\code{\link{monash_1}},
\code{\link{monash_2}},
\code{\link{msu}},
\code{\link{northwestern_1}},
\code{\link{northwestern_2}},
\code{\link{notredame_1}},
\code{\link{notredame_2}},
\code{\link{nyu_1}},
\code{\link{nyu_2}},
\code{\link{nyu_accent}},
\code{\link{nyu_neutral}},
\code{\link{ohio_uni_1}},
\code{\link{ohio_uni_2}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{princeton_0}},
\code{\link{princeton_1}},
\code{\link{princeton_2}},
\code{\link{rpi_1}},
\code{\link{rpi_2}},
\code{\link{rpi_3}},
\code{\link{stanford_1}},
\code{\link{stanford_2}},
\code{\link{stanford_3}},
\code{\link{uchicago_1}},
\code{\link{uchicago_2}},
\code{\link{uci_1}},
\code{\link{uci_2}},
\code{\link{uci_3}},
\code{\link{uci_4}},
\code{\link{ucla_1}},
\code{\link{ucla_2}},
\code{\link{ucla_3}},
\code{\link{ucla_4}},
\code{\link{ucsd_1}},
\code{\link{ucsd_2}},
\code{\link{ucsd_3}},
\code{\link{uflorida_1}},
\code{\link{uflorida_2}},
\code{\link{uflorida_3}},
\code{\link{umass_1}},
\code{\link{umass_2}},
\code{\link{umass_neutrals}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{willamette}},
\code{\link{wm_1}},
\code{\link{wm_2}},
\code{\link{wm_3}},
\code{\link{yale}},
\code{\link{yeshiva}}
}
\author{
\strong{unicol}, 2023-06-16.
}
\concept{U.S. university color palettes}
\keyword{datasets}
