% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dundee.R
\docType{data}
\name{dundee_highlight}
\alias{dundee_highlight}
\title{Colours of the University of Dundee}
\format{
An object of class \code{character} of length 3.
}
\source{
Colour definitions are based on the
\href{https://www.dundee.ac.uk/brand/colour-palette}{Brand colour palette}.
}
\usage{
dundee_highlight
}
\value{
A named vector of colours (HEX/HTML codes of type character).
}
\description{
\code{dundee_highlight} provides the highlight colours 
of the \href{https://www.dundee.ac.uk/}{University of Dundee}, Scotland.
}
\details{
Three complementary highlight colours have been specified to add vibrancy to the identity. 
These colours can only be used for typography, icons or delicate line work.

The highlight colours are 
\code{"highlight_1"} (defined as PANTONE 710C, CMYK 0/85/52/0, RGB 255/98/100, HEX #ff6264),
\code{"highlight_2"} (defined as PANTONE 7480C, CMYK 82/0/72/0, RGB 1/209/124, HEX #01d17c) and
\code{"highlight_3"} (defined as 88\% Black, CMYK 0/0/0/88, RGB 70/70/70, HEX #464646).

\code{dundee_highlight} uses the HEX colour definition. 
Pantone and CMYK colours differ from RGB and HEX definition.
}
\examples{
dundee_highlight
unikn::seecol(dundee_highlight, main = "Dundee highlight colours") # view colour palette

}
\seealso{
\code{\link{dundee_core}} for the core colour of the University of Dundee;
\code{\link{dundee_block}} for the block colours of the University of Dundee;
\code{\link{dundee_background}} for the background colour of the University of Dundee;
\code{\link{seecol}} for viewing and comparing colour palettes;
\code{\link{usecol}} for using colour palettes;
\code{\link{simcol}} for finding similar colours;
\code{\link{newpal}} for defining new colour palettes;
\code{\link{grepal}} for finding named colours.

Other university colour palettes: 
\code{\link{birmingham_1}},
\code{\link{dundee_background}},
\code{\link{dundee_block}},
\code{\link{dundee_core}},
\code{\link{mcgill_brights}},
\code{\link{mcgill_darks}},
\code{\link{mcgill_grey}},
\code{\link{mcgill_muted}},
\code{\link{mcgill_pastels}},
\code{\link{mcgill_red}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_primary}},
\code{\link{ucalgary_secondary}},
\code{\link{ucalgary_warmgreys}}
}
\author{
\strong{unicol}, 2023-06-22.
}
\concept{university colour palettes}
\keyword{datasets}
