% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_example_pal.R
\docType{data}
\name{example}
\alias{example}
\title{Primary colors of the University of Example, Country}
\format{
An object of class \code{character} of length 4.
}
\source{
Color definitions are based on the 
\href{https://example.org}{doc title} (HTML or PDF).
}
\usage{
example
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{example} provides the four primary colors 
of the \href{https://example.org}{University of Example}, Country.
}
\details{
The 4 primary colors of the University of Example are 
\code{"Uni pink"} (defined as HEX \code{"#E0607E"}),
\code{"bright grey"} (defined as HEX \code{"#CCCCCC"}), 
\code{"dark gray"} (defined as HEX \code{"#4C4C4C"}), and
\code{"black"} (defined as the R color \code{"black"} or HEX \code{"#000000"}).

\code{example} uses the HEX color definitions.
}
\examples{
example
unikn::seecol(example, main = "Colors of the University of Example")  # view color palette
unikn::demopal(example, main = "Illustrating the colors of Example")  # demo color palette 

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.
}
\author{
\strong{unicol}, 2023-10-12.
}
\concept{university color palettes}
\keyword{datasets}
