% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_example_pal.R
\docType{data}
\name{example_1}
\alias{example_1}
\title{Primary colors of the University of Example, Country}
\format{
An object of class \code{character} of length 4.
}
\source{
Color definitions are based on the 
\href{https://example.org}{color source (HTML or PDF)}.
}
\usage{
example_1
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{example_1} provides the four primary colors 
of the \href{https://example.org}{University of Example}, Country.
}
\details{
The 4 primary colors are 
\code{"pinky pink"} (defined as HEX \code{"#E0607E"}),
\code{"bright grey"} (defined as HEX \code{"#CCCCCC"}), 
\code{"dark grey"} (defined as HEX \code{"#4C4C4C"}), and
\code{"dark"} (defined as R color \code{"black"} or HEX \code{"#000000"}).

\code{example_1} uses the HEX color definitions.
}
\examples{
example_1
unikn::seecol(example_1, main = "Colors of the University of Example 1") # view color palette
unikn::demopal(example_1, main = "Illustrating the colors of Example 1") # demo color palette 

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{oxford_error}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{uni_manchester}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_print}},
\code{\link{yale_1}},
\code{\link{yeshiva}},
\code{\link{york_1}},
\code{\link{york_2}}
}
\author{
\strong{unicol}, 2023-08-23.
}
\concept{university color palettes}
\keyword{datasets}
