% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Guelph.R
\docType{data}
\name{guelph}
\alias{guelph}
\title{Colors of the University of Guelph, Canada}
\format{
An object of class \code{character} of length 4.
}
\source{
Color definitions are based on the
\href{https://news.uoguelph.ca/guides/brand-guide/colour-palette/}{Website of the University of Guelph}.
}
\usage{
guelph
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{guelph} provides the two primary colors and the two secondary colors
of the \href{https://www.uoguelph.ca/}{University of Guelph}, Canada.
}
\details{
The two primary colors are
\code{"PMS_200"} (defined as RGB 194/4/48, HEX #C20430) and
\code{"black"} (defined as RGB 0/0/0, HEX #000000).

The two secondary color are
\code{"PMS_123"} (defined as RGB 255/199/42, HEX #FFC72A) and
\code{"PMS_549"} (defined as RGB 105/163/185, HEX ##69A3B9).

\code{guelph} uses the HEX color definitions.
}
\examples{
guelph
unikn::seecol(guelph, main = "Colors of the University of Guelph") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other Canadian university color palettes: 
\code{\link{conc_1}},
\code{\link{conc_2}},
\code{\link{conc_3}},
\code{\link{conc_4}},
\code{\link{laval}},
\code{\link{manitoba_1}},
\code{\link{manitoba_2}},
\code{\link{mcgill_brights}},
\code{\link{mcgill_darks}},
\code{\link{mcgill_grey}},
\code{\link{mcgill_muted}},
\code{\link{mcgill_pastels}},
\code{\link{mcgill_red}},
\code{\link{mcmaster_brighterworld}},
\code{\link{mcmaster_heritage}},
\code{\link{mun_1}},
\code{\link{mun_2}},
\code{\link{ottawa_1}},
\code{\link{toronto_1}},
\code{\link{ubc_1}},
\code{\link{uvic_complementary}},
\code{\link{uvic_core}}
}
\author{
\strong{unicol}, 2023-07-12.
}
\concept{Canadian university color palettes}
\keyword{datasets}
