% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Harvard.R
\docType{data}
\name{harvard_2}
\alias{harvard_2}
\title{Accent colors of Harvard University, USA}
\format{
An object of class \code{character} of length 4.
}
\source{
Color definitions are based on the
\href{https://identityguide.hms.harvard.edu/color/}{Color manual}.
}
\usage{
harvard_2
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{harvard_2} provides the four accent colors
of the \href{https://www.harvard.edu/}{Harvard University}, USA.
}
\details{
The secondary palette provides a neutral field against which the primary colors can live:
\code{"mortar"} (defined as RGB 140/129/121, HEX #8C8179),
\code{"parchment"} (defined as RGB 243/243/241, HEX #F3F3F1),
\code{"slate"} (defined as RGB 137/150/160 , HEX #8996A0) and
\code{"shade"} (defined as RGB 186/197/198, HEX #BAC5C6).
}
\examples{
harvard_2
unikn::seecol(harvard_2, main = "Accent colors Harvard Uni") # view color palette

}
\seealso{
\code{\link{harvard_1}} for primary colors of Harvard University;
\code{\link{harvard_3}} for web colors of Harvard University;
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other U.S. university color palettes: 
\code{\link{CMU_1_core}},
\code{\link{CMU_2_tartan}},
\code{\link{CMU_3_campus}},
\code{\link{asu_1}},
\code{\link{asu_2}},
\code{\link{asu_3}},
\code{\link{berkeley_1}},
\code{\link{berkeley_2}},
\code{\link{brown_1}},
\code{\link{brown_2}},
\code{\link{bu_1}},
\code{\link{caltech_1}},
\code{\link{caltech_2}},
\code{\link{caltech_3}},
\code{\link{chicago_1}},
\code{\link{chicago_2}},
\code{\link{columbia}},
\code{\link{cornell_1}},
\code{\link{cornell_2}},
\code{\link{cornell_3_accent}},
\code{\link{dartmouth_1}},
\code{\link{dartmouth_2}},
\code{\link{dartmouth_3}},
\code{\link{duke_1}},
\code{\link{duke_2}},
\code{\link{harvard_1}},
\code{\link{harvard_3}},
\code{\link{jhu_0}},
\code{\link{jhu_1}},
\code{\link{jhu_2}},
\code{\link{jhu_3_accent}},
\code{\link{jhu_4_gray}},
\code{\link{michigan_1}},
\code{\link{michigan_2}},
\code{\link{minnesotatwin_1}},
\code{\link{minnesotatwin_2}},
\code{\link{mit}},
\code{\link{monash_1}},
\code{\link{monash_2}},
\code{\link{msu}},
\code{\link{northwestern_1}},
\code{\link{northwestern_2}},
\code{\link{notredame_1}},
\code{\link{notredame_2}},
\code{\link{nyu_1}},
\code{\link{nyu_2}},
\code{\link{nyu_accent}},
\code{\link{nyu_neutral}},
\code{\link{princeton_0}},
\code{\link{princeton_1}},
\code{\link{princeton_2}},
\code{\link{rpi_1}},
\code{\link{rpi_2}},
\code{\link{rpi_3}},
\code{\link{uci_1}},
\code{\link{uci_2}},
\code{\link{uci_3}},
\code{\link{uci_4}},
\code{\link{uflorida_1}},
\code{\link{uflorida_2}},
\code{\link{uflorida_3}}
}
\author{
\strong{unicol}, 2023-06-14.
}
\concept{U.S. university color palettes}
\keyword{datasets}
