% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HU_Berlin.R
\docType{data}
\name{hu_1_digital}
\alias{hu_1_digital}
\title{Digital colors of the Humboldt University (HU) Berlin, Germany}
\format{
An object of class \code{character} of length 2.
}
\source{
Color definitions are based on HU's 
\href{https://www.hu-berlin.de/de/service/design/basiselemente/farbe}{Corporate Design guide}.
}
\usage{
hu_1_digital
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{hu_1_digital} provides the two digital colors 
of the \href{https://www.hu-berlin.de/en}{Humboldt University Berlin}, Germany, 
updated in 2023.
}
\details{
The 2 primary colors of HU Berlin are 
\code{HU blau} (defined as CMYK 100 62 0 46, RGB 0 55 108, or HEX \code{"#00376c"}), and 
\code{HU schwarz} (defined as CMYK 70 30 30 90, RGB 10 10 25, or HEX \code{"#0a0a19"}). 

The alternative blue color 
\code{HU digitalblau} (NOT defined as CMYK, but as RGB 0 0 255 or HEX \code{"#0000ff"}) 
is reserved for digital applications (e.g., online or slide presentations) 
and is combined here with plain \code{"black"} (\code{"schwarz"} or HEX \code{"#000000"}).

\code{hu_1_digital} uses the HEX color definitions.
}
\examples{
hu_1_digital
unikn::seecol(hu_1_digital, n = 5, main = "5 digital shades of HU Berlin")  # view color palette
unikn::demopal(hu_1_digital, type = 1, 
               main = "Digital colors of Humboldt University Berlin")

}
\seealso{
\code{\link{hu_1}} for primary colors of the HU Berlin;
\code{\link{hu_2}} for secondary colors of the HU Berlin; 
\code{\link{hu_3_accent}} for possible accent colors of the HU Berlin;
\code{\link{hu_1_2022}} for primary colors of the HU Berlin (2022); 
\code{\link{hu_2_2022}} for secondary colors of the HU Berlin (2022);  
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other German university color palettes: 
\code{\link{fu_0}},
\code{\link{fu_1}},
\code{\link{fu_2}},
\code{\link{fu_3}},
\code{\link{hu_1_2022}},
\code{\link{hu_1}},
\code{\link{hu_2_2022}},
\code{\link{hu_2}},
\code{\link{hu_3_accent}},
\code{\link{lmu_1}},
\code{\link{lmu_2}},
\code{\link{lmu_3}},
\code{\link{rptu}},
\code{\link{uni_bonn_1}},
\code{\link{uni_bonn_2}},
\code{\link{uni_freiburg_0}},
\code{\link{uni_freiburg_1}},
\code{\link{uni_freiburg_2}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_halle}},
\code{\link{uni_hamburg_1}},
\code{\link{uni_hamburg_2}},
\code{\link{uni_heidelberg}},
\code{\link{uni_jena_1}},
\code{\link{uni_jena_2}},
\code{\link{uni_kassel}},
\code{\link{uni_kiel_1}},
\code{\link{uni_kiel_2}},
\code{\link{uni_koeln_1}},
\code{\link{uni_koeln_2}},
\code{\link{uni_konstanz_1}},
\code{\link{uni_konstanz_2}},
\code{\link{uni_magdeburg_1}},
\code{\link{uni_magdeburg_2}},
\code{\link{uni_magdeburg_3}},
\code{\link{uni_mannheim_1}},
\code{\link{uni_mannheim_2}},
\code{\link{uni_potsdam}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_2}},
\code{\link{uni_regensburg_3}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_print}},
\code{\link{uni_ulm_1}},
\code{\link{uni_ulm_2}}
}
\author{
\strong{unicol}, 2023-08-31.
}
\concept{German university color palettes}
\keyword{datasets}
