% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indiana.R
\docType{data}
\name{IU_1}
\alias{IU_1}
\title{Primary colors of Indiana University Bloomington (IU), USA}
\format{
An object of class \code{character} of length 2.
}
\source{
Color definitions are based on \href{https://bloomington.iu.edu/index.html}{IU}'s 
\strong{Brand and Uniform Guidelines} (2018).
}
\usage{
IU_1
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{IU_1} provides the two primary colors
of \href{https://bloomington.iu.edu/index.html}{Indiana University Bloomington} (IU), USA.
}
\details{
The 2 primary colors are  
\code{"cream"} (defined as HEX \code{"#EDEBEB"}) and 
\code{"crimson"} (defined as HEX \code{"#990000"}). 

Cream and crimson are the key components of the IU brand and the IU Athletics look. 
Crimson is the visual anchor and should be dominant in all marketing materials. 
For marketing purposes, “cream” is treated as white.

Note that the members of the IU community are collectively known as Hoosiers  
and the logo of Indiana University is the IU trident. 

\code{IU_1} uses the HEX color definitions.
}
\examples{
IU_1
unikn::seecol(IU_1, main = "Primary colors of IU Bloomington (IU)") # view color palette

}
\seealso{
\code{\link{IU_2}} for IU's secondary colors;
\code{\link{IU_2_dp}} a digital print version of IU's secondary colors;
\code{\link[unikn]{seecol}} for viewing and comparing color palettes;
\code{\link[unikn]{usecol}} for using color palettes;
\code{\link[unikn]{simcol}} for finding similar colors;
\code{\link[unikn]{newpal}} for defining new color palettes;
\code{\link[unikn]{grepal}} for finding named colors.

Other U.S. university color palettes: 
\code{\link{IU_2}},
\code{\link{IU_2_dp}},
\code{\link{asu_1}},
\code{\link{asu_2}},
\code{\link{asu_3}},
\code{\link{berkeley_1}},
\code{\link{berkeley_2}},
\code{\link{brown_1}},
\code{\link{brown_2}},
\code{\link{bu}},
\code{\link{caltech_1}},
\code{\link{caltech_2}},
\code{\link{caltech_3}},
\code{\link{cmu_1}},
\code{\link{cmu_2}},
\code{\link{cmu_3}},
\code{\link{columbia_1}},
\code{\link{columbia_2}},
\code{\link{columbia_3}},
\code{\link{cornell_1}},
\code{\link{cornell_2}},
\code{\link{cornell_3_accent}},
\code{\link{dartmouth_1}},
\code{\link{dartmouth_2}},
\code{\link{dartmouth_3}},
\code{\link{duke_1}},
\code{\link{duke_2}},
\code{\link{harvard_1}},
\code{\link{harvard_2}},
\code{\link{harvard_3}},
\code{\link{jhu_0}},
\code{\link{jhu_1}},
\code{\link{jhu_2}},
\code{\link{jhu_3_accent}},
\code{\link{jhu_4_gray}},
\code{\link{manchester_uni_1}},
\code{\link{manchester_uni_2}},
\code{\link{michigan_1}},
\code{\link{michigan_2}},
\code{\link{minnesotatwin_1}},
\code{\link{minnesotatwin_2}},
\code{\link{mit}},
\code{\link{monash_1}},
\code{\link{monash_2}},
\code{\link{msu}},
\code{\link{northwestern_1}},
\code{\link{northwestern_2}},
\code{\link{notredame_1}},
\code{\link{notredame_2}},
\code{\link{nyu_1}},
\code{\link{nyu_2}},
\code{\link{nyu_accent}},
\code{\link{nyu_neutral}},
\code{\link{ohio_uni_1}},
\code{\link{ohio_uni_2}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{princeton_0}},
\code{\link{princeton_1}},
\code{\link{princeton_2}},
\code{\link{rpi_1}},
\code{\link{rpi_2}},
\code{\link{rpi_3}},
\code{\link{stanford_1}},
\code{\link{stanford_2}},
\code{\link{stanford_3}},
\code{\link{uchicago_1}},
\code{\link{uchicago_2}},
\code{\link{uci_1}},
\code{\link{uci_2}},
\code{\link{uci_3}},
\code{\link{uci_4}},
\code{\link{ucla_1}},
\code{\link{ucla_2}},
\code{\link{ucla_3}},
\code{\link{ucla_4}},
\code{\link{ucsd_1}},
\code{\link{ucsd_2}},
\code{\link{ucsd_3}},
\code{\link{uflorida_1}},
\code{\link{uflorida_2}},
\code{\link{uflorida_3}},
\code{\link{umass_1}},
\code{\link{umass_2}},
\code{\link{umass_neutrals}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{willamette}},
\code{\link{wm_1}},
\code{\link{wm_2}},
\code{\link{wm_3}},
\code{\link{yale}},
\code{\link{yeshiva}}
}
\author{
\strong{unicol}, 2026-01-16.
}
\concept{U.S. university color palettes}
\keyword{datasets}
