% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lund.R
\docType{data}
\name{lund_1}
\alias{lund_1}
\title{Primary colours of the University of Lund, Sweden}
\format{
An object of class \code{character} of length 2.
}
\source{
Color definitions are based on \href{https://www.lu.se}{LU}'s staff guidelines on 
\href{https://www.staff.lu.se/support-and-tools/communication-and-graphic-profile/graphic-profile-and-logotype/colours}{Colour} (HTML).
}
\usage{
lund_1
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{lund_1} provides the two primary colors 
of the \href{https://www.lu.se}{Lund University}, Sweden.
}
\details{
\code{lund_1} is based on RGB color definitions.
}
\examples{
lund_1
unikn::seecol(lund_1, main = "Primary colours of Lund University, Sweden")

}
\seealso{
\code{\link{lund_2}} for pastel colours of Lund University; 
\code{\link{lund_3}} for darker pastel colours of Lund University; 
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other Swedish university color palettes: 
\code{\link{gothenburg}},
\code{\link{lund_2}},
\code{\link{lund_3}},
\code{\link{stockholm_1}},
\code{\link{stockholm_2}},
\code{\link{uppsala}}
}
\author{
\strong{unicol}, 2023-10-12.
}
\concept{Swedish university color palettes}
\keyword{datasets}
