% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manchester_Uni.R
\docType{data}
\name{manchester_uni_1}
\alias{manchester_uni_1}
\title{Primary colours of Manchester University, UK}
\format{
An object of class \code{character} of length 3.
}
\source{
Color definitions are based on the
\href{https://www.manchester.edu/about-manchester/office-directory/marketing/brand-toolkit/color-palette}{Manchester's marketing brand toolkit}.
}
\usage{
manchester_uni_1
}
\value{
A named vector of colours (HEX/HTML codes of type character).
}
\description{
\code{manchester_uni_1} provides the three primary colours 
of the \href{https://www.manchester.edu}{Manchester University}, England/UK.
}
\details{
The three primary colours are
\code{"Manchester gold"} (defined as C=0 M=28 Y=100 K=6, R=238 G=177 B=17, #EEB111),
\code{"Manchester black"} (defined as C=30 M=30 Y=30 K=100, R=10 G=2 B=3, #0A0203), and
\code{"Manchester gray"} (defined as C=31 M=25 Y=31 K=6, R=150 G=152 B=147, #969893).

\code{manchester_uni_1} uses the HEX colour definitions.
}
\examples{
manchester_uni_1
unikn::seecol(manchester_uni_1, main = "Primary colours of Manchester University") 
# view color palette

}
\seealso{
\code{\link{manchester_uni_accent}} for accent colours of Manchester University;
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colours;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colours.

Other U.K. university color palettes: 
\code{\link{birmingham_1}},
\code{\link{lancaster_1}},
\code{\link{lancaster_2}},
\code{\link{manchester_uni_accent}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}}
}
\author{
\strong{unicol}, 2023-06-27.
}
\concept{U.K. university color palettes}
\keyword{datasets}
