% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manchester_Uni.R
\docType{data}
\name{manchester_uni_accent}
\alias{manchester_uni_accent}
\title{Accent colours of Manchester University, UK}
\format{
An object of class \code{character} of length 4.
}
\source{
Color definitions are based on the
\href{https://www.manchester.edu/about-manchester/office-directory/marketing/brand-toolkit/color-palette}{Manchester's marketing brand toolkit}.
}
\usage{
manchester_uni_accent
}
\value{
A named vector of colours (HEX/HTML codes of type character).
}
\description{
\code{manchester_uni_accent} provides the three accent primary colours and 
white as suggested to use in minority ratio (3:1) against the primary colors on a canvas
of the \href{https://www.manchester.edu}{Manchester University}, England/UK.
}
\details{
The three accent colours are 
\code{"Manchester sunrise"} (defined as C=0 M=71 Y=100 K=0, R=255 G=108 B=12, #FF6C0C),
\code{"Manchester blue"} (defined as C=54 M=4 Y=9 K=0, R=107 G=196 B=222, #6BC4DE),
\code{"Manchester green"} (defined as C=55 M=3 Y=100 K=0, R=130 G=188 B=0, #82BC00), and
\code{"white"}.

\code{manchester_uni_accent} uses the HEX color definitions.
}
\examples{
manchester_uni_accent
unikn::seecol(manchester_uni_accent, main = "Accent colors of Manchester University") 
# view color palette

}
\seealso{
\code{\link{manchester_uni_1}} for primary colours of Manchester University;
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colours;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colours.

Other U.K. university color palettes: 
\code{\link{birmingham_1}},
\code{\link{lancaster_1}},
\code{\link{lancaster_2}},
\code{\link{manchester_uni_1}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}}
}
\author{
\strong{unicol}, 2023-06-27.
}
\concept{U.K. university color palettes}
\keyword{datasets}
