% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIT.R
\docType{data}
\name{mit}
\alias{mit}
\title{Primary colors of the MIT (Massachusetts Institute of Technology), USA}
\format{
An object of class \code{character} of length 4.
}
\source{
Color definitions are based on the
\href{https://web.mit.edu/graphicidentity/colors.html}{Color manual (HTML)}.
}
\usage{
mit
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{mit} provides the three primary colors
of the \href{https://web.mit.edu}{Massachusetts Institute of Technology}, USA, 
plus a secondary light gray color.
}
\details{
The three primary colors are
\code{"black"} (corresponding to RGB 0 0 0 and HEX #000000), 
\code{"red"} (defined as PMS 201, RGB 163 31 52, HEX #A31F34), and 
\code{"gray"} (defined as PMS 424, RGB 138 139 140, HEX #8A8B8C). 

The secondary color 
\code{"lightgray"} (defined as PMS 420, RGB 194 192 191, HEX #C2C0BF)
is used for contrast reasons.
}
\examples{
mit
unikn::seecol(mit, main = "MIT, Massachusetts Institute of Technology") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other U.S. university color palettes: 
\code{\link{CMU_1_core}},
\code{\link{CMU_2_tartan}},
\code{\link{CMU_3_campus}},
\code{\link{UCLA_1}},
\code{\link{UCLA_2}},
\code{\link{UCLA_3}},
\code{\link{UCSD_1_core}},
\code{\link{UCSD_2_accent}},
\code{\link{UCSD_3_neutral}},
\code{\link{asu_1}},
\code{\link{asu_2}},
\code{\link{asu_3}},
\code{\link{berkeley_1}},
\code{\link{berkeley_2}},
\code{\link{brown_1}},
\code{\link{brown_2}},
\code{\link{bu_1}},
\code{\link{caltech_1}},
\code{\link{caltech_2}},
\code{\link{caltech_3}},
\code{\link{chicago_1}},
\code{\link{chicago_2}},
\code{\link{columbia}},
\code{\link{cornell_1}},
\code{\link{cornell_2}},
\code{\link{cornell_3_accent}},
\code{\link{dartmouth_1}},
\code{\link{dartmouth_2}},
\code{\link{dartmouth_3}},
\code{\link{duke_1}},
\code{\link{duke_2}},
\code{\link{harvard_1}},
\code{\link{harvard_2}},
\code{\link{harvard_3}},
\code{\link{jhu_0}},
\code{\link{jhu_1}},
\code{\link{jhu_2}},
\code{\link{jhu_3_accent}},
\code{\link{jhu_4_gray}},
\code{\link{michigan_1}},
\code{\link{michigan_2}},
\code{\link{minnesotatwin_1}},
\code{\link{minnesotatwin_2}},
\code{\link{monash_1}},
\code{\link{monash_2}},
\code{\link{msu}},
\code{\link{northwestern_1}},
\code{\link{northwestern_2}},
\code{\link{notredame_1}},
\code{\link{notredame_2}},
\code{\link{nyu_1}},
\code{\link{nyu_2}},
\code{\link{nyu_accent}},
\code{\link{nyu_neutral}},
\code{\link{princeton_0}},
\code{\link{princeton_1}},
\code{\link{princeton_2}},
\code{\link{rpi_1}},
\code{\link{rpi_2}},
\code{\link{rpi_3}},
\code{\link{stanford_1}},
\code{\link{stanford_2_accent}},
\code{\link{stanford_3_web}},
\code{\link{uci_1}},
\code{\link{uci_2}},
\code{\link{uci_3}},
\code{\link{uci_4}},
\code{\link{uflorida_1}},
\code{\link{uflorida_2}},
\code{\link{uflorida_3}},
\code{\link{umass_1}},
\code{\link{umass_2}},
\code{\link{umass_neutrals}}
}
\author{
\strong{unicol}, 2023-06-11.
}
\concept{U.S. university color palettes}
\keyword{datasets}
