% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIT.R
\docType{data}
\name{mit_pal}
\alias{mit_pal}
\title{Colors of MIT, Massachusetts Institute of Technology, USA}
\format{
An object of class \code{character} of length 4.
}
\source{
Color definitions are based on the
\href{https://web.mit.edu/graphicidentity/colors.html}{Color manual (HTML)}.
}
\usage{
mit_pal
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{mit_pal} provides the three primary colors
of the \href{https://web.mit.edu}{Massachusetts Institute of Technology}, USA.
}
\details{
The three primary colors are
\code{"black"}, 
\code{"red"} (defined as PMS 201, RGB 163/31/52, HEX #A31F34), and 
\code{"gray"} (defined as PMS 424, RGB 138/139/140, HEX #8A8B8C). 

The secondary color is
\code{"lightgray"} (defined as PMS 420, RGB 194/192/191, HEX #C2C0BF)
and used for contrast reasons.
}
\examples{
mit_pal
unikn::seecol(mit_pal, main = "MIT, Massachusetts Institute of Technology") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{caltech_pal_1}},
\code{\link{caltech_pal_2}},
\code{\link{caltech_pal_3}},
\code{\link{eth_pal_1}},
\code{\link{eth_pal_2}},
\code{\link{eth_pal_3}},
\code{\link{fu_pal_0}},
\code{\link{fu_pal_1}},
\code{\link{fu_pal_2}},
\code{\link{fu_pal_3}},
\code{\link{hu_pal_1}},
\code{\link{hu_pal_2}},
\code{\link{lmu_pal_1}},
\code{\link{lmu_pal_2}},
\code{\link{lmu_pal_3}},
\code{\link{mpg_pal}},
\code{\link{nyu_1}},
\code{\link{nyu_2}},
\code{\link{nyu_accent}},
\code{\link{nyu_neutral}},
\code{\link{rpi_pal_1}},
\code{\link{rpi_pal_2}},
\code{\link{rpi_pal_3}},
\code{\link{rptu_pal}},
\code{\link{uni_berkeley_1}},
\code{\link{uni_berkeley_2}},
\code{\link{uni_bonn_1}},
\code{\link{uni_bonn_2}},
\code{\link{uni_chicago_1}},
\code{\link{uni_chicago_2}},
\code{\link{uni_columbia}},
\code{\link{uni_cornell_1}},
\code{\link{uni_cornell_2}},
\code{\link{uni_cornell_3_accent}},
\code{\link{uni_freiburg_0}},
\code{\link{uni_freiburg_1}},
\code{\link{uni_freiburg_2}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_hamburg_1}},
\code{\link{uni_hamburg_2}},
\code{\link{uni_harvard_accent_3}},
\code{\link{uni_harvard_primary_1}},
\code{\link{uni_harvard_secondary_2}},
\code{\link{uni_hopkins_accent_4}},
\code{\link{uni_hopkins_grayscale_5}},
\code{\link{uni_hopkins_identity_1}},
\code{\link{uni_hopkins_primary_2}},
\code{\link{uni_hopkins_secondary_3}},
\code{\link{uni_jena_1}},
\code{\link{uni_jena_2}},
\code{\link{uni_kiel_1}},
\code{\link{uni_kiel_2}},
\code{\link{uni_koeln_1}},
\code{\link{uni_koeln_2}},
\code{\link{uni_konstanz_1}},
\code{\link{uni_konstanz_2}},
\code{\link{uni_mannheim_1}},
\code{\link{uni_mannheim_2}},
\code{\link{uni_michigan_primary_1}},
\code{\link{uni_michigan_secondary_2}},
\code{\link{uni_northwestern_primary_1}},
\code{\link{uni_northwestern_secondary_2}},
\code{\link{uni_pal_1}},
\code{\link{uni_penn_logo_1}},
\code{\link{uni_penn_web_2}},
\code{\link{uni_princeton_0}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_2}},
\code{\link{uni_regensburg_3}},
\code{\link{uni_stanford_1}},
\code{\link{uni_stanford_accent}},
\code{\link{uni_stanford_web}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_2_print}},
\code{\link{uni_ulm_1}},
\code{\link{uni_ulm_2}},
\code{\link{uni_yale_primary}}
}
\author{
\strong{unicol}, 2023-06-11.
}
\concept{university color palettes}
\keyword{datasets}
