% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monash.R
\docType{data}
\name{monash_2}
\alias{monash_2}
\title{Secondary colors of Monash University, USA}
\format{
An object of class \code{character} of length 9.
}
\source{
Color definitions are based on the
\href{https://www.monash.edu/__data/assets/pdf_file/0004/1656508/9.-Digital-Guidelines.pdf}{Color manual}.
}
\usage{
monash_2
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{monash_2} provides the nine secondary colors
of \href{https://www.monash.edu/}{Monash University}, USA.
}
\details{
In accordance with the Web Content Accessibility Guide (WCAG) AA standards, 
the secondary colors have been approved for online use with white text overlay.

Secondary colors are 
\code{"blue"} (defined as HEX #00739d),
\code{"purple"} (defined as HEX #6f64a9),
\code{"fuchsia"} (defined as HEX #c90095),
\code{"ruby"} (defined as HEX #c800d9),
\code{"red"} (defined as HEX #df0021),
\code{"orange"} (defined as HEX #c33603),
\code{"umber"} (defined as HEX #795548),
\code{"olive"} (defined as HEX #616d3a), and
\code{"green"} (defined as HEX #006f29).
}
\examples{
monash_2
unikn::seecol(monash_2, main = "Secondary colors of Monash University") # view color palette

}
\seealso{
\code{\link{monash_1}} for the primary colors of Monash University; 
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other U.S. university color palettes: 
\code{\link{CMU_1_core}},
\code{\link{CMU_2_tartan}},
\code{\link{CMU_3_campus}},
\code{\link{asu_1}},
\code{\link{berkeley_1}},
\code{\link{berkeley_2}},
\code{\link{brown_1}},
\code{\link{brown_2}},
\code{\link{bu_1}},
\code{\link{caltech_1}},
\code{\link{caltech_2}},
\code{\link{caltech_3}},
\code{\link{chicago_1}},
\code{\link{chicago_2}},
\code{\link{columbia}},
\code{\link{cornell_1}},
\code{\link{cornell_2}},
\code{\link{cornell_3_accent}},
\code{\link{monash_1}}
}
\author{
\strong{unicol}, 2023-06-16.
}
\concept{U.S. university color palettes}
\keyword{datasets}
