% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oxford.R
\docType{data}
\name{oxford_general}
\alias{oxford_general}
\title{Coulors of the University of Oxford}
\format{
An object of class \code{character} of length 11.
}
\source{
Colour definitions are based on the
\href{https://www.ox.ac.uk/public-affairs/style-guide/digital-style-guide}{Digital style guide}.
}
\usage{
oxford_general
}
\value{
A named vector of colours (HEX/HTML codes of type character).
}
\description{
\code{oxford_general} provides the general/ secondary colours 
of the \href{https://www.ox.ac.uk/}{University of Oxford}, England.
}
\details{
The wide secondary palette is used to add visual distinction and texture 
to design elements as well as signifying page links, 
call-to-actions (e.g buttons) and active pages in navigational elements.

The colours are
\code{"dark_blue"} (defined as HEX #001c3d),
\code{"washed_out_blue"} (defined as HEX #193658),
\code{"copyright_grey"} (defined as HEX #c7c2bc),
\code{"beige_grey"} (defined as HEX #f3f1ec),
\code{"brown_grey"} (defined as HEX #001c3d),
\code{"filter_form_green"} (defined as HEX #043946),
\code{"blue_grey"} (defined as HEX #a1c4d0),
\code{"teal"} (defined as HEX #003947),
\code{"very_light_blue"} (defined as HEX #f0f5f8),
\code{"cookie_bar"} (defined as HEX #00152e) and
\code{"blue_grey"} (defined as HEX #353c47).
}
\examples{
oxford_general
unikn::seecol(oxford_general, main = "Uni Oxford general colours") # view colour palette

}
\seealso{
\code{\link{oxford_brand}} for the brand colours of the University of Oxford;
\code{\link{oxford_link}} for the link colours of the University of Oxford;
\code{\link{oxford_shades}} for the shade colours of the University of Oxford;
\code{\link{oxford_blog}} for the blog colours of the University of Oxford;
\code{\link{oxford_graduate}} for the graduate colours of the University of Oxford;
\code{\link{oxford_error}} for the error colours of the University of Oxford;
\code{\link{oxford_socialmedia}} for the social media colours of the University of Oxford;
\code{\link{seecol}} for viewing and comparing colour palettes;
\code{\link{usecol}} for using colour palettes;
\code{\link{simcol}} for finding similar colours;
\code{\link{newpal}} for defining new colour palettes;
\code{\link{grepal}} for finding named colours.

Other university colour palettes: 
\code{\link{birmingham_1}},
\code{\link{mcgill_brights}},
\code{\link{mcgill_darks}},
\code{\link{mcgill_grey}},
\code{\link{mcgill_muted}},
\code{\link{mcgill_pastels}},
\code{\link{mcgill_red}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_primary}},
\code{\link{ucalgary_secondary}},
\code{\link{ucalgary_warmgreys}}
}
\author{
\strong{unicol}, 2023-06-15.
}
\concept{university colour palettes}
\keyword{datasets}
