% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arizona.R
\docType{data}
\name{pal_asu_secondary}
\alias{pal_asu_secondary}
\title{Colors of the Arizona State University}
\format{
An object of class \code{character} of length 7.
}
\source{
Color definitions are based on 
\href{https://brandguide.asu.edu/brand-elements/design/color}{ASU's brand guide}.
}
\usage{
pal_asu_secondary
}
\value{
A names vector of colors (HEX/HTML codes of type character).
}
\description{
\code{pal_asu_secondary} provides the 7 secondary colors
of the \href{https://www.asu.edu}{Arizona State University}, USA.
}
\details{
The 7 secondary colors are
\code{"asu_green"} (defined as HEX #78BE20),
\code{"asu_blue"} (defined as HEX #00A3E0),
\code{"asu_orange"} (defined as HEX #FF7F32),
\code{"asu_gray"} (defined as HEX #747474)
\code{"asu_copper"} (defined as HEX #Af674B)
\code{"asu_turquoise"} (defined as HEX #4AB7C4)
\code{"asu_pink"} (defined as HEX #E74973)

\code{pal_asu_secondary} uses the HEX color definitions.
}
\examples{
pal_asu_secondary
unikn::seecol(pal_asu_secondary, main = "Arizona State University") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes.: 
\code{\link{limerick_2}},
\code{\link{notredame_2}},
\code{\link{pal_dartmouth_2}},
\code{\link{pal_dartmouth_3}},
\code{\link{pal_trinity_2}},
\code{\link{pal_uci_secondary}},
\code{\link{pal_vanderbiltu_neutrals}},
\code{\link{pla_asu_grayscale}},
\code{\link{uflorida_2}}
}
\author{
\strong{unicol}, 2320-07-13.
}
\concept{university color palettes.}
\keyword{datasets}
