% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dartmouth.R
\docType{data}
\name{pal_dartmouth_2}
\alias{pal_dartmouth_2}
\title{Secondary colors of Dartmouth College}
\format{
An object of class \code{character} of length 8.
}
\source{
Color definitions are based on 
\href{https://communications.dartmouth.edu/visual-identity/design-elements/color-palette}{Dartmouth's visual identity guide}.
}
\usage{
pal_dartmouth_2
}
\value{
A names vector of colors (HEX/HTML codes of type character).
}
\description{
\code{pal_dartmouth_2} provides the 8 secondary colors
of the \href{https://home.dartmouth.edu}{Dartmouth College}, USA.
}
\details{
The 4 secondary colors are
\code{"forest_green"}(defined as HEX #12312b),
\code{"rich_forest_green"}(defined as HEX #0D1E1C),
\code{"snow_white"}(defined as HEX #ffffff),
\code{"midnight_black"} (defined as HEX #000000)
\code{"web_gray_1"} (defined as HEX #f7f7f7)
\code{"web_gray_2"} (defined as HEX #e2e2e2)
\code{"web_gray_3"} (defined as HEX #707070)
\code{"granite_gray"} (defined as HEX #424141)

\code{pal_dartmouth_2} uses the HEX color definition.
}
\examples{
pal_dartmouth_2
unikn::seecol(pal_dartmouth_2, main = "Dartmouth College") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes.: 
\code{\link{pal_asu_secondary}},
\code{\link{pal_dartmouth_3}},
\code{\link{pal_galway_2}},
\code{\link{pal_limerick_2}},
\code{\link{pal_notredame_secondary}},
\code{\link{pal_trinity_2}},
\code{\link{pal_uci_secondary}},
\code{\link{pal_uflorida_secondary}},
\code{\link{pal_vanderbiltu_neutrals}},
\code{\link{pla_asu_grayscale}}
}
\author{
\strong{unicol}, 2320-07-13.
}
\concept{university color palettes.}
\keyword{datasets}
