% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Galway.R
\docType{data}
\name{pal_galway_2}
\alias{pal_galway_2}
\title{Secondary colors of the University of Galway, Ireland}
\format{
An object of class \code{character} of length 14.
}
\source{
Color definitions are based on 
\href{https://www.universityofgalway.ie/media/brand/11901_University_of_Galway_Brand_Guidelines_v13.pdf}(UGalway's brand guidelines).
}
\usage{
pal_galway_2
}
\value{
A names vector of colors (HEX/HTML codes of type character).
}
\description{
\code{pal_galway_2} provides the 14 secondary colors
of the \href{https://www.universityofgalway.ie}{University of Galway}, Ireland.
}
\details{
The 14 secondary colors are
\code{"aqua_marine"} (defined as HEX #00b1ba),
\code{"ocean_blue"} (defined as HEX #0080a3),
\code{"spring_green"} (defined as HEX #00bf6f),
\code{"lime_green"} (defined as HEX #c5e86c), 
\code{"orange"} (defined as HEX #f58e2f), 
\code{"rose"} (defined as HEX #e30044), 
\code{"yellow"} (defined as HEX #ffc72b), 
\code{"deep_orchid"} (defined as HEX #c4299b), 
\code{"lavender_pink"} (defined as HEX #e66eb2), 
\code{"violet"} (defined as HEX #753abd), 
\code{"ash"} (defined as HEX #5e514d), 
\code{"warm_stone"} (defined as HEX #94795d), 
\code{"sandstone"} (defined as HEX #c4c0b7), and 
\code{"black"} (defined as HEX #000000). 

\code{pal_galway_2} uses the HEX color definition.
}
\examples{
pal_galway_2
unikn::seecol(pal_galway_2, main = "University of Galway 2") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes.: 
\code{\link{pal_asu_secondary}},
\code{\link{pal_dartmouth_2}},
\code{\link{pal_dartmouth_3}},
\code{\link{pal_limerick_2}},
\code{\link{pal_notredame_secondary}},
\code{\link{pal_trinity_2}},
\code{\link{pal_uci_secondary}},
\code{\link{pal_uflorida_secondary}},
\code{\link{pal_vanderbiltu_neutrals}},
\code{\link{pla_asu_grayscale}}
}
\author{
\strong{unicol}, 2320-07-16.
}
\concept{university color palettes.}
\keyword{datasets}
