% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glasgow.R
\docType{data}
\name{pal_glasgow_2}
\alias{pal_glasgow_2}
\title{Secondary colors of the University of Glasgow, Scotland}
\format{
An object of class \code{character} of length 5.
}
\source{
Color definitions are based on 
\href{https://www.gla.ac.uk/myglasgow/staff/brandtoolkit/brandelements/colours/}{UGlasgow's website}.
}
\usage{
pal_glasgow_2
}
\value{
A named vector of colors (RGB/HTML codes of type character).
}
\description{
\code{pal_glasgow_2} provides the 5 secondary colors 
of the \href{https://www.gla.ac.uk/}{University of Glasgow}, Scotland.
}
\details{
The 5 secondary colors are 
\code{"glasgow_turquoise"} (defined as R0 G181 B209),
\code{"glasgow_rose"} (defined as R176 G108 B150),
\code{"glasgow_pumpkin"} (defined as R255 G185 B72),
\code{"glasgow_sunshine"} (defined as R255 G220 B54), and
\code{"glasgow_mocha"} (defined as R170 G128 B102).

\code{pal_glasgow_2} uses the RGB color definition.
}
\examples{
pal_glasgow_2
unikn::seecol(pal_glasgow_2, main = "University of Glasgow 2") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university of color palettes: 
\code{\link{pal_glasgow_1}},
\code{\link{pal_minnesotatwin_1}}
}
\author{
\strong{unicol}, 2023-07-05.
}
\concept{university of color palettes}
\keyword{datasets}
