% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limerick.R
\docType{data}
\name{pal_limerick_2}
\alias{pal_limerick_2}
\title{Colors of the University of Limerick}
\format{
An object of class \code{character} of length 16.
}
\source{
Color definitions are based on the
\href{https://www.ul.ie/media/31805/download?inline}
}
\usage{
pal_limerick_2
}
\value{
A names vector of colors (HEX/HTML codes of type character).
}
\description{
\code{pal_limerick_2} provides the 17 secondary colors
of the \href{https://www.ul.ie}{University of Limerick}, Ireland.
}
\details{
The 14 secondary colors are
\code{"shannon"}(defined as HEX #00A3E0),
\code{"sky"}(defined as HEX #007DBA),
\code{"stone"}(defined as HEX #2D5980),
\code{"marine"} (defined as HEX #13294B)
\code{"marigold"} (defined as HEX #FFC72C)
\code{"golden_leaf"} (defined as HEX #FFA300)
\code{"pumpkin"} (defined as HEX #D45D00)
\code{"munster"} (defined as HEX #CB333B)
\code{"salmon"} (defined as HEX #E31C79)
\code{"autumnal_leaf"} (defined as HEX #6F263D)
\code{"copper"} (defined as HEX #89532F)
\code{"bronze"} (defined as HEX #7e5c4d)
\code{"white"} (defined as HEX #FFFFFF)
\code{"limestone"} (defined as HEX #919D9D)
\code{"pebble"} (defined as HEX #707372)
\code{"granite"} (defined as HEX #505759)
\code{"slate"} (defined as HEX #373A36)

\code{pal_limerick_2} uses the HEX color definition.
}
\examples{
pal_limerick_2
unikn::seecol(pal_limerick_2, main = "University of Limerick") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes.: 
\code{\link{pal_asu_secondary}},
\code{\link{pal_dartmouth_2}},
\code{\link{pal_dartmouth_3}},
\code{\link{pal_notredame_secondary}},
\code{\link{pal_trinity_2}},
\code{\link{pal_uci_secondary}},
\code{\link{pal_uflorida_secondary}},
\code{\link{pal_vanderbiltu_neutrals}},
\code{\link{pla_asu_grayscale}}
}
\author{
\strong{unicol}, 2320-07-16.
}
\concept{university color palettes.}
\keyword{datasets}
