% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Minnesota Twin Cities.R
\docType{data}
\name{pal_minnesotatwin_primary}
\alias{pal_minnesotatwin_primary}
\title{Colors of the University of Minnesota}
\format{
An object of class \code{character} of length 2.
}
\source{
Color definitions are based on 
\href{https://university-relations.umn.edu/resources/colors-and-type}{UMN's website}.
}
\usage{
pal_minnesotatwin_primary
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{pal_minnesotatwin_primary} provides the primary colors
of the \href{https://twin-cities.umn.edu/}{University of Minnesota}, USA.
}
\details{
The 2 primary colors are
\code{"minnesotatwin_maroon"}(defined as HEX #7A0019) and
\code{"minnesotatwin_gold"} (defined as HEX #FFCC33).

\code {pal_minnesotatwin_primary} uses the HEX color definition.
}
\examples{
pal_minnesotatwin_primary
unikn::seecol(pal_minnesotatwin_primary, main="University of Minnesota Twin Cities") #view color palette


}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university of color palettes: 
\code{\link{pal_glasgow_1}},
\code{\link{pal_glasgow_2}},
\code{\link{pal_minnesotatwin_1}}
}
\author{
\strong{unicol}, 2023-07-01.
}
\concept{university of color palettes}
\keyword{datasets}
