% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Michigan_State.R
\docType{data}
\name{pal_msu}
\alias{pal_msu}
\title{\code {pal_msu} provides the 6 colors of the \href {https://msu.edu/}, USA.}
\format{
An object of class \code{character} of length 6.
}
\source{
Color definitions are based on the
\href{https://brand.msu.edu/visual/color-palette}.
}
\usage{
pal_msu
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
The 6 colors are
\code{"Spartan_Green_msu"} (defined as HEX #18453B),
\code{"White"} (defined as HEX #FFFFFF),
\code{"Black"} (defined as HEX #000000),
\code{"kelly_green_msu"} (defined as HEX #008208),
\code{"lime_green_msu"} (defined as HEX #7BBD00) and
\code{"excellence_green_msu"} (defined as HEX #0B9A6D).
}
\details{
\code{pal_msu} uses the HEX color definition
}
\examples{
pal_msu
unikn::seecol(pal_msu, main="Michigan State University") #view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.
}
\author{
\strong{unicol}, 2023-07-01
}
\keyword{datasets}
