% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pittsburgh.R
\docType{data}
\name{pal_pitt_1}
\alias{pal_pitt_1}
\title{Contemporary colors of the University of Pittsburgh}
\format{
An object of class \code{character} of length 2.
}
\source{
Color definitions are based on 
\href{https://www.brand.pitt.edu/visual-identity/color-palettes/primary-palette}{Pitt's visual identity guide}.
}
\usage{
pal_pitt_1
}
\value{
A named vector of colors (HEX/HTML codes of type character)
}
\description{
\code{pal_pitt_1} provides the 2 contemporary colors of the 
\href{https://www.pitt.edu/}{University of Pittsburgh},USA
}
\details{
The 2 contemporary colors are
\code {"pittsburgh_royal"} (defined as HEX #003594),
\code {"pittsburgh_gold"} (defined as HEX #FFB81C)
}
\examples{
pal_pitt_1
uniln::seecol(pal_pitt_1, main "Contemporary colors") #view color palette

}
\seealso{
\code{\link{pal_pitt_2}} for the University of Pittsburgh's legacy colors; 
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colours;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.
}
\author{
\strong{unicol}, 2023-07-01
}
\concept{university colour palettes}
\keyword{datasets}
