% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pittsburgh.R
\docType{data}
\name{pal_pittsburgh_contemporary}
\alias{pal_pittsburgh_contemporary}
\title{Colors of the university of Pittyburgh}
\format{
An object of class \code{character} of length 2.
}
\source{
color definitions are based in the
\href{https://www.brand.pitt.edu/visual-identity/color-palettes/primary-palette}.
}
\usage{
pal_pittsburgh_contemporary
}
\value{
A named vector of colors (HEX/HTML codes of type character)
}
\description{
\code{pal_pittsburgh_primary} provides the 2 contemporary colors of the \href{https://www.pitt.edu/}{University of Pittsburgh},USA
}
\details{
The 2 contemporary colors are
\code {"pittsburgh_royal"} (defined as HEX #003594),
\code {"pittsburgh_gold"} (defined as HEX #FFB81C)
}
\seealso{
\code{\link{seecol}} for viewing and comparing colour palettes;
\code{\link{usecol}} for using colour palettes;
\code{\link{simcol}} for finding similar colours;
\code{\link{newpal}} for defining new colour palettes;
\code{\link{grepal}} for finding named colours.
}
\author{
\strong{unicol}, 2023-07-01
}
\concept{university colour palettes}
\keyword{datasets}
