% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Irvine.R
\docType{data}
\name{pal_ucihealth_2}
\alias{pal_ucihealth_2}
\title{Neutral colors of the University of California, Irvine}
\format{
An object of class \code{character} of length 5.
}
\source{
Color definitions are based on 
\href{https://brand.uci.edu/master-branding/color-palette/}{UCI's brand guide}.
}
\usage{
pal_ucihealth_2
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{pal_ucihealth_2} provides the 5 secondary health colors
of the \href{https://www.uci.edu}{University of California Health, Irvine}, USA.
}
\details{
The 5 neutral colors are
\code{"brick_red"} (defined as HEX #981e32),
\code{"purple"} (defined as HEX #631d76),
\code{"teal_blue"} (defined as HEX #0083b3),
\code{"green"} (defined as HEX #3f9c35), and 
\code{"dark_blue"} (defined as HEX #1b3d6d).

\code{pal_ucihealth_2} uses the HEX color definitions.
}
\examples{
pal_ucihealth_2
unikn::seecol(pal_ucihealth_2, main = "University of California, Irvine") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.
}
\author{
\strong{unicol}, 2023-07-13.
}
\keyword{datasets}
