% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Florida.R
\docType{data}
\name{pal_uflorida_neutral}
\alias{pal_uflorida_neutral}
\title{Colors of the University of Florida}
\format{
An object of class \code{character} of length 5.
}
\source{
Color definitions are based on the 
\href{https://brandcenter.ufl.edu/colors/}
}
\usage{
pal_uflorida_neutral
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{pal_uflorida_neutral} provides the 5 neutra colors of the 
\href{https://www.ufl.edu/}
}
\details{
The 5 neutral colors are
\code{"black"}(defined as HEX #000000),
\code{"florida_cool_grey_11"}(defined as HEX #343741),
\code{"florida_cool_grey_3"}(defined as HEX #C7C9C8),
\code{"florida_warm_grey_1"}(defined as HEX #D8D4D7) and 
\code{"white"}(defined as HEX #FFFFFF).

\code{pal_uflorida_neutral} uses the HEX color definitions.
}
\examples{
pal_uflorida_neutral
unikn::seecol(pal_uflorida_neutral, main = "University of Florida") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.
}
\author{
\strong{unicol}, 2023-07-08.
}
\keyword{datasets}
