% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Florida.R
\docType{data}
\name{pal_uflorida_secondary}
\alias{pal_uflorida_secondary}
\title{Colors of the University of Florida}
\format{
An object of class \code{character} of length 5.
}
\source{
Color definitions are based on the
\href{https://brandcenter.ufl.edu/colors/}
}
\usage{
pal_uflorida_secondary
}
\value{
A names vector of colors (HEX/HTML codes of type character).
}
\description{
\code{pal_uflorida_secondary} provides the 5 secondary colors
of the \href{https://www.ufl.edu/}{University of Florida}, USA.
}
\details{
The 5 secondary colors are
\code{"florida_bottlebrush"}(defined as HEX #D32737),
\code{"florida_alachua"}(defined as HEX #F2A900),
\code{"florida_gator"}(defined as HEX #2284C),
\code{"florida_dark_blue"} (defined as HEX #002657) and
\code{"florida_perennial"} (defined as HEX #6A2A60).

\code{pal_uflorida_secondary} uses the HEX color definition.
}
\examples{
pal_uflorida_secondary
unikn::seecol(pal_uflorida_secondary, main = "University of Florida") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes.: 
\code{\link{pal_asu_secondary}},
\code{\link{pal_dartmouth_2}},
\code{\link{pal_dartmouth_3}},
\code{\link{pal_limerick_2}},
\code{\link{pal_notredame_secondary}},
\code{\link{pal_trinity_2}},
\code{\link{pal_uci_secondary}},
\code{\link{pal_vanderbiltu_neutrals}},
\code{\link{pla_asu_grayscale}}
}
\author{
\strong{unicol}, 2320-07-07.
}
\concept{university color palettes.}
\keyword{datasets}
