% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vanderbilt.R
\docType{data}
\name{pal_vanderbiltu_neutrals}
\alias{pal_vanderbiltu_neutrals}
\title{Neutral colors of Vanderbilt University}
\format{
An object of class \code{character} of length 4.
}
\source{
Color definitions are based on the 
\href{https://brand.vanderbilt.edu/color//}
}
\usage{
pal_vanderbiltu_neutrals
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{pal_vanderbiltu_neutrals} provides the 4 neutral colors
of the \href{https://www.vanderbilt.edu/}{Vanderbilt University}, USA.
}
\details{
The 4 neutral colors are
\code{"vanderbilt_dark_grey"}(defined as HEX #777777),
\code{"vanderbilt_light_gery"}(defines as HEX #E4E4E4),
\code{"vanderbilt_sand"}(defined as HEX #E0D5C0) and 
\code{"vanderbilt_cream"}(defined as HEX #F5F3EF).

\code{pal_vanderbiltu_neutrals} uses the HEX color definition.
}
\examples{
pal_vanderbiltu_neutrals
unikn::seecol(pal_vanderbiltu_neutrals, main = "Vanderbilt University") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes.: 
\code{\link{pal_asu_secondary}},
\code{\link{pal_dartmouth_2}},
\code{\link{pal_dartmouth_3}},
\code{\link{pal_galway_2}},
\code{\link{pal_limerick_2}},
\code{\link{pal_notredame_secondary}},
\code{\link{pal_trinity_2}},
\code{\link{pal_uci_secondary}},
\code{\link{pal_uflorida_secondary}},
\code{\link{pla_asu_grayscale}}
}
\author{
\strong{unicol}, 2023-07-12
}
\concept{university color palettes.}
\keyword{datasets}
