% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pittsburgh.R
\docType{data}
\name{pitt_1}
\alias{pitt_1}
\title{Contemporary colors of the University of Pittsburgh, USA}
\format{
An object of class \code{character} of length 2.
}
\source{
Color definitions are based on 
\href{https://www.brand.pitt.edu/visual-identity/color-palettes/primary-palette}{Pitt's visual identity guide}.
}
\usage{
pitt_1
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{pitt_1} provides the two contemporary colors of the 
\href{https://www.pitt.edu/}{University of Pittsburgh}, USA.
}
\details{
The 2 contemporary colors are
\code{"pitt_royal"} (defined as HEX #003594) and 
\code{"pitt_gold"} (defined as HEX #FFB81C).
}
\examples{
pitt_1
unikn::seecol(pitt_1, main = "Pitt's contemporary colors") # view color palette

}
\seealso{
\code{\link{pitt_2}} for the University of Pittsburgh's legacy colors; 
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{UCLA_1}},
\code{\link{UCLA_2}},
\code{\link{UCLA_3}},
\code{\link{UCSD_1_core}},
\code{\link{UCSD_2_accent}},
\code{\link{UCSD_3_neutral}},
\code{\link{example_1}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{pitt_2}},
\code{\link{queens_1}},
\code{\link{queens_2}},
\code{\link{sfu_brand}},
\code{\link{standrews_1}},
\code{\link{standrews_2}},
\code{\link{stanford_1}},
\code{\link{stanford_2_accent}},
\code{\link{stanford_3_web}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_primary}},
\code{\link{ucalgary_secondary}},
\code{\link{ucalgary_warmgreys}},
\code{\link{umass_2}},
\code{\link{umass_brand}},
\code{\link{umass_neutrals}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_heidelberg_1}},
\code{\link{uni_manchester_1}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_2_print}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{uwaterloo_arts}},
\code{\link{uwaterloo_engineering}},
\code{\link{uwaterloo_environment}},
\code{\link{uwaterloo_health}},
\code{\link{uwaterloo_main}},
\code{\link{uwaterloo_math}},
\code{\link{uwaterloo_science}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{waikato}},
\code{\link{western_uni_1}},
\code{\link{yale_1}},
\code{\link{yeshiva}},
\code{\link{york_1}},
\code{\link{york_2}}
}
\author{
\strong{unicol}, 2023-07-01.
}
\concept{university color palettes}
\keyword{datasets}
