% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueensUniversity.R
\docType{data}
\name{queens_2}
\alias{queens_2}
\title{Secondary colours of Queen's University, Canada}
\format{
An object of class \code{character} of length 11.
}
\source{
Colour definitions are based on the
\href{https://www.queensu.ca/brand-central/visual-identity/colours}{Visual Identity}.
}
\usage{
queens_2
}
\value{
A named vector of colours (HEX/HTML codes of type character).
}
\description{
\code{queens_2} provides the secondary colours 
of \href{https://www.queensu.ca/}{Queen's University}, Canada.
}
\details{
A secondary palette of neutral colours may be used with the primary palette to provide balance.

The three primary colours are
\code{"white"} (defined as CMYK 0/0/0/0, RGB 255/255/255, HEX #ffffff),
\code{"lightest_grey"} (defined as CMYK 0/0/0/5, RGB 241/242/242, HEX #f1f2f2),
\code{"light_grey"} (defined as CMYK 0/0/0/10, RGB 230/231/232, HEX #e6e7e8),
\code{"grey"} (defined as CMYK 0/0/0/20, RGB 209/211/212, HEX #d1d3d4),
\code{"medium_grey"} (defined as CMYK 0/0/0/40, RGB 167/169/172, HEX #a7a9ac),
\code{"dark_grey"} (defined as CMYK 0/0/0/60, RGB 128/130/133, HEX #808285),
\code{"darkest_grey"} (defined as CMYK 0/0/0/80, RGB 88/89/91, HEX #58595b),
\code{"black"} (defined as CMYK 0/0/0/100, RGB 33/33/33, HEX #212121),
\code{"light_limestone"} (defined as PMS 401, CMYK 28/25/28/3, RGB 180/174/168, HEX #b3aea8),
\code{"limestone"} (defined as PMS 403 CMYK 39/36/40/14, RGB 144/137/130, HEX #908982), and
\code{"dark_limestone"} (defined as PMS 405, CMYK 49/47/51/32, RGB 106/98/92, HEX #6a625c).

\code{queens_2} uses the HEX colour definition. 
PMS and CMYK colours differ from RGB and HEX definition.
}
\examples{
queens_2
unikn::seecol(queens_2, 
              main = "Secondary colours of Queen's University") # view colour palette

}
\seealso{
\code{\link{queens_1}} for the primary colours of Queen's University;
\code{\link{seecol}} for viewing and comparing colour palettes;
\code{\link{usecol}} for using colour palettes;
\code{\link{simcol}} for finding similar colours;
\code{\link{newpal}} for defining new colour palettes;
\code{\link{grepal}} for finding named colours.

Other university color palettes: 
\code{\link{UCLA_1}},
\code{\link{UCLA_2}},
\code{\link{UCLA_3}},
\code{\link{UCSD_1_core}},
\code{\link{UCSD_2_accent}},
\code{\link{UCSD_3_neutral}},
\code{\link{example_1}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{queens_1}},
\code{\link{sfu_brand}},
\code{\link{standrews_1}},
\code{\link{standrews_2}},
\code{\link{stanford_1}},
\code{\link{stanford_2_accent}},
\code{\link{stanford_3_web}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_primary}},
\code{\link{ucalgary_secondary}},
\code{\link{ucalgary_warmgreys}},
\code{\link{umass_2}},
\code{\link{umass_brand}},
\code{\link{umass_neutrals}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_heidelberg_1}},
\code{\link{uni_manchester_1}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_2_print}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{uwaterloo_arts}},
\code{\link{uwaterloo_engineering}},
\code{\link{uwaterloo_environment}},
\code{\link{uwaterloo_health}},
\code{\link{uwaterloo_main}},
\code{\link{uwaterloo_math}},
\code{\link{uwaterloo_science}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{waikato}},
\code{\link{western_uni_1}},
\code{\link{yale_1}},
\code{\link{yeshiva}},
\code{\link{york_1}},
\code{\link{york_2}}
}
\author{
\strong{unicol}, 2023-06-28.
}
\concept{university color palettes}
\keyword{datasets}
