% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueensUniversity.R
\docType{data}
\name{queens_secondary}
\alias{queens_secondary}
\title{Secondary colours of Queen's University}
\format{
An object of class \code{character} of length 11.
}
\source{
Colour definitions are based on the
\href{https://www.queensu.ca/brand-central/visual-identity/colours}{Visual Identity}.
}
\usage{
queens_secondary
}
\value{
A named vector of colours (HEX/HTML codes of type character).
}
\description{
\code{queens_secondary} provides the secondary colours 
of \href{https://www.queensu.ca/}{Queen's University}, Canada.
}
\details{
A secondary palette of neutral colours may be used with the primary palette to provide balance.

The three primary colours are
\code{"white"} (defined as CMYK 0/0/0/0, RGB 255/255/255, HEX #ffffff),
\code{"lightest_grey"} (defined as CMYK 0/0/0/5, RGB 241/242/242, HEX #f1f2f2),
\code{"light_grey"} (defined as CMYK 0/0/0/10, RGB 230/231/232, HEX #e6e7e8),
\code{"grey"} (defined as CMYK 0/0/0/20, RGB 209/211/212, HEX #d1d3d4),
\code{"medium_grey"} (defined as CMYK 0/0/0/40, RGB 167/169/172, HEX #a7a9ac),
\code{"dark_grey"} (defined as CMYK 0/0/0/60, RGB 128/130/133, HEX #808285),
\code{"darkest_grey"} (defined as CMYK 0/0/0/80, RGB 88/89/91, HEX #58595b),
\code{"black"} (defined as CMYK 0/0/0/100, RGB 33/33/33, HEX #212121),
\code{"light_limestone"} (defined as PMS 401, CMYK 28/25/28/3, RGB 180/174/168, HEX #b3aea8),
\code{"limestone"} (defined as PMS 403 CMYK 39/36/40/14, RGB 144/137/130, HEX #908982), and
\code{"dark_limestone"} (defined as PMS 405, CMYK 49/47/51/32, RGB 106/98/92, HEX #6a625c).

\code{queens_secondary} uses the HEX colour definition. 
PMS and CMYK colours differ from RGB and HEX definition.
}
\examples{
queens_secondary
unikn::seecol(queens_secondary, 
              main = "Secondary colours of Queen's University") # view colour palette

}
\seealso{
\code{\link{queens_primary}} for the primary colours of Queen's University;
\code{\link{seecol}} for viewing and comparing colour palettes;
\code{\link{usecol}} for using colour palettes;
\code{\link{simcol}} for finding similar colours;
\code{\link{newpal}} for defining new colour palettes;
\code{\link{grepal}} for finding named colours.

Other university colour palettes: 
\code{\link{birmingham_1}},
\code{\link{bu_1}},
\code{\link{dundee_background}},
\code{\link{dundee_block}},
\code{\link{dundee_core}},
\code{\link{dundee_highlight}},
\code{\link{lancaster_1}},
\code{\link{lancaster_2}},
\code{\link{mcgill_brights}},
\code{\link{mcgill_darks}},
\code{\link{mcgill_grey}},
\code{\link{mcgill_muted}},
\code{\link{mcgill_pastels}},
\code{\link{mcgill_red}},
\code{\link{mcmaster_brighterworld}},
\code{\link{mcmaster_heritage}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{queens_primary}},
\code{\link{sfu_brand}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_primary}},
\code{\link{ucalgary_secondary}},
\code{\link{ucalgary_warmgreys}},
\code{\link{uni_heidelberg_1}},
\code{\link{uvic_complementary}},
\code{\link{uvic_core}}
}
\author{
\strong{unicol}, 2023-06-28.
}
\concept{university colour palettes}
\keyword{datasets}
