% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_pals.R
\docType{data}
\name{rpi_pal_2}
\alias{rpi_pal_2}
\title{Secondary colors of RPI}
\format{
An object of class \code{character} of length 3.
}
\source{
Color definitions are based on 
\url{https://scer.rpi.edu/brand-in-action/colors}.
}
\usage{
rpi_pal_2
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{rpi_pal_2} provides the secondary use colors of the 
\href{https://www.rpi.edu/}{Rensselaer Polytechnic Institute} (RPI), Troy, NY.
}
\details{
The colors of \code{rpi_pal_2} should be used in combination with  
the primary colors of \code{\link{rpi_pal_1}} to add color accents 
to digital and print materials. 

\code{\link{rpi_pal_3}} provides tint variants of these colors.
}
\examples{
rpi_pal_2
unikn::seecol(rpi_pal_2, main = "The secondary colors of RPI")  # view color palette
unikn::demopal(rpi_pal_2, type = 5, main = "Secondary colors of RPI")

}
\seealso{
\code{\link{rpi_pal_1}} for primary colors of RPI; 
\code{\link{rpi_pal_3}} for tint colors of RPI; 
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{caltech_pal_1}},
\code{\link{caltech_pal_2}},
\code{\link{caltech_pal_3}},
\code{\link{eth_pal_1}},
\code{\link{eth_pal_2}},
\code{\link{eth_pal_3}},
\code{\link{fu_pal_0}},
\code{\link{fu_pal_1}},
\code{\link{fu_pal_2}},
\code{\link{fu_pal_3}},
\code{\link{hu_pal_1}},
\code{\link{hu_pal_2}},
\code{\link{lmu_pal_1}},
\code{\link{lmu_pal_2}},
\code{\link{lmu_pal_3}},
\code{\link{mit_pal}},
\code{\link{mpg_pal}},
\code{\link{nyu_1}},
\code{\link{nyu_2}},
\code{\link{nyu_accent}},
\code{\link{nyu_neutral}},
\code{\link{rpi_pal_1}},
\code{\link{rpi_pal_3}},
\code{\link{rptu_pal}},
\code{\link{uni_berkeley_primary_1}},
\code{\link{uni_berkeley_secondary_2}},
\code{\link{uni_bonn_1}},
\code{\link{uni_bonn_2}},
\code{\link{uni_chicago_primary_1}},
\code{\link{uni_chicago_secondary_2}},
\code{\link{uni_columbia}},
\code{\link{uni_cornell_accent_web_3}},
\code{\link{uni_cornell_primary_web_1}},
\code{\link{uni_cornell_secondary_web_2}},
\code{\link{uni_freiburg_0}},
\code{\link{uni_freiburg_1}},
\code{\link{uni_freiburg_2}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_hamburg_1}},
\code{\link{uni_hamburg_2}},
\code{\link{uni_harvard_accent_3}},
\code{\link{uni_harvard_primary_1}},
\code{\link{uni_harvard_secondary_2}},
\code{\link{uni_hopkins_accent_4}},
\code{\link{uni_hopkins_grayscale_5}},
\code{\link{uni_hopkins_identity_1}},
\code{\link{uni_hopkins_primary_2}},
\code{\link{uni_hopkins_secondary_3}},
\code{\link{uni_jena_1}},
\code{\link{uni_jena_2}},
\code{\link{uni_kiel_1}},
\code{\link{uni_kiel_2}},
\code{\link{uni_koeln_1}},
\code{\link{uni_koeln_2}},
\code{\link{uni_konstanz_1}},
\code{\link{uni_konstanz_2}},
\code{\link{uni_mannheim_1}},
\code{\link{uni_mannheim_2}},
\code{\link{uni_michigan_primary_1}},
\code{\link{uni_michigan_secondary_2}},
\code{\link{uni_northwestern_primary_1}},
\code{\link{uni_northwestern_secondary_2}},
\code{\link{uni_pal_1}},
\code{\link{uni_penn_logo_1}},
\code{\link{uni_penn_web_2}},
\code{\link{uni_princeton_0}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_2}},
\code{\link{uni_regensburg_3}},
\code{\link{uni_stanford_1}},
\code{\link{uni_stanford_accent}},
\code{\link{uni_stanford_web}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_2_print}},
\code{\link{uni_ulm_1}},
\code{\link{uni_ulm_2}},
\code{\link{uni_yale_primary}}
}
\author{
\strong{unicol}, 2022-12-01.
}
\concept{university color palettes}
\keyword{datasets}
