% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/St_Andrews.R
\docType{data}
\name{standrews_2}
\alias{standrews_2}
\title{Secondary colours of the University of St Andrews, UK}
\format{
An object of class \code{character} of length 8.
}
\source{
Colour definitions are based on 
\href{https://www.st-andrews.ac.uk/brand/design-guidelines/colours/}{St. Andrews design guidelines}.
}
\usage{
standrews_2
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{standrews_2} provides the eight secondary colours 
of the \href{https://www.st-andrews.ac.uk/brand/design-guidelines/colours/}{University of St Andrews}, UK.
}
\details{
The 8 secondary colours are 
\code{"andrews_light_blue"} (defined as HEX #00aeef), 
\code{"andrews_mid_blue"} (defined as HEX #007dc5),
\code{"andrews_burgundy"} (defined as HEX #c60c46),
\code{"andrews_purple"} (defined as HEX #7b439a)
\code{"andrews_orange"} (defined as HEX #f5842b),
\code{"andrews_green"} (defined as HEX #54b948), 
\code{"andrews_mid_green"} (defined as HEX #00853f), and 
\code{"andrews_dark_green"} (defined as HEX #005953).

\code{standrews_2} uses the HEX colour definitions.
}
\examples{
standrews_2
unikn::seecol(standrews_2, main = "University of St Andrews 2") # view color palette

}
\seealso{
\code{\link{standrews_1}} for primary colours of St. Andrews; 
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{UCLA_1}},
\code{\link{UCLA_2}},
\code{\link{UCLA_3}},
\code{\link{UCSD_1_core}},
\code{\link{UCSD_2_accent}},
\code{\link{UCSD_3_neutral}},
\code{\link{example_1}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{queens_1}},
\code{\link{queens_2}},
\code{\link{sfu_brand}},
\code{\link{standrews_1}},
\code{\link{stanford_1}},
\code{\link{stanford_2_accent}},
\code{\link{stanford_3_web}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_primary}},
\code{\link{ucalgary_secondary}},
\code{\link{ucalgary_warmgreys}},
\code{\link{umass_2}},
\code{\link{umass_brand}},
\code{\link{umass_neutrals}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_heidelberg_1}},
\code{\link{uni_manchester_1}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_2_print}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{uwaterloo_arts}},
\code{\link{uwaterloo_engineering}},
\code{\link{uwaterloo_environment}},
\code{\link{uwaterloo_health}},
\code{\link{uwaterloo_main}},
\code{\link{uwaterloo_math}},
\code{\link{uwaterloo_science}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{waikato}},
\code{\link{western_uni_1}},
\code{\link{yale_1}},
\code{\link{yeshiva}},
\code{\link{york_1}},
\code{\link{york_2}}
}
\author{
\strong{unicol}, 2023-07-16.
}
\concept{university color palettes}
\keyword{datasets}
