% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stanford.R
\docType{data}
\name{stanford_1}
\alias{stanford_1}
\title{Primary colors of Stanford University}
\format{
An object of class \code{character} of length 15.
}
\source{
Color definitions are based on the
\href{https://identity.stanford.edu/design-elements/color/}{Color manual}.
}
\usage{
stanford_1
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{stanford_1} provides the 15 primary colors
of the \href{https://www.stanford.edu/}{Stanford University}, USA.
}
\details{
Primary color palette consists of Cardinal red, white, black and cool grey.
The layouts lean heavily on these colors, which are
\code{"cardinal_red"} (defined as HEX #8C1515),
\code{"cardinal_red_light"} (defined as HEX #B83A4B),
\code{"cardinal_red_dark"} (defined as HEX #820000),
\code{"white"} (defined as HEX #FFFFFF),
\code{"black"} (defined as HEX #2E2D29),
\code{"black_90"} (defined as HEX #43423E),
\code{"black_80"} (defined as HEX #585754),
\code{"black_70"} (defined as HEX #6D6C69),
\code{"black_60"} (defined as HEX #767674),
\code{"black_50"} (defined as HEX #979694),
\code{"black_40"} (defined as HEX #ABABA9),
\code{"black_30"} (defined as HEX #C0C0BF),
\code{"black_20"} (defined as HEX #D5D5D4),
\code{"black_10"} (defined as HEX #EAEAEA) and 
\code{"cool_grey"} (defined as HEX #53565A).
}
\examples{
stanford_1
unikn::seecol(stanford_1, main = "Primary colors Stanford Uni") # view color palette

}
\seealso{
\code{\link{stanford_2_accent}} for secondary, accent colors of Stanford University;
\code{\link{stanford_3_web}} for web colors of Stanford University;
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{UCLA_1}},
\code{\link{UCLA_2}},
\code{\link{UCLA_3}},
\code{\link{UCSD_1_core}},
\code{\link{UCSD_2_accent}},
\code{\link{UCSD_3_neutral}},
\code{\link{example_1}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{queens_1}},
\code{\link{queens_2}},
\code{\link{sfu_brand}},
\code{\link{stanford_2_accent}},
\code{\link{stanford_3_web}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_primary}},
\code{\link{ucalgary_secondary}},
\code{\link{ucalgary_warmgreys}},
\code{\link{umass_2}},
\code{\link{umass_brand}},
\code{\link{umass_neutrals}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_heidelberg_1}},
\code{\link{uni_manchester_1}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_2_print}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{uwaterloo_arts}},
\code{\link{uwaterloo_engineering}},
\code{\link{uwaterloo_environment}},
\code{\link{uwaterloo_health}},
\code{\link{uwaterloo_main}},
\code{\link{uwaterloo_math}},
\code{\link{uwaterloo_science}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{waikato}},
\code{\link{western_uni_1}},
\code{\link{yale_1}},
\code{\link{yeshiva}},
\code{\link{york_1}},
\code{\link{york_2}}
}
\author{
\strong{unicol}, 2023-06-14.
}
\concept{university color palettes}
\keyword{datasets}
