% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stockholm.R
\docType{data}
\name{stockholm_2}
\alias{stockholm_2}
\title{Additional colours of the University of Stockholm, Sweden}
\format{
An object of class \code{character} of length 9.
}
\source{
Color definitions are based on 
\href{https://www.su.se}{SU's} 
\href{https://www.su.se/staff/organisation-governance/governing-documents-rules-and-regulations/communication-collaboration/rules-for-the-visual-identity-at-stockholm-university-1.17458}{visual identity guide} (HTML).
}
\usage{
stockholm_2
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{stockholm_2} provides nine additional colours 
of the \href{https://www.su.se}{University of Stockholm}, Sweden.
}
\details{
\code{stockholm_2} consists of 2 accent colours, 4 base colours, and 3 other colours. 
Note that the 3 other colours -- named "Green", "Red", and "Red 10\%" -- have limited use 
(see the Visual identity manual for details).

\code{stockholm_2} is based on the digital palette and uses HEX color definitions. 
(See the \href{https://www.su.se/staff/organisation-governance/governing-documents-rules-and-regulations/communication-collaboration/rules-for-the-visual-identity-at-stockholm-university-1.17458}{visual identity guide} 
for printing details and definitions.)
}
\examples{
stockholm_2
unikn::seecol(stockholm_2, col_bg = "grey96",  
              main = "Additional colours of the University of Stockholm, Sweden")  # view palette

}
\seealso{
\code{\link{stockholm_1}} for main colours of the University of Stockholm;
\code{\link[unikn]{seecol}} for viewing and comparing color palettes;
\code{\link[unikn]{usecol}} for using color palettes;
\code{\link[unikn]{simcol}} for finding similar colors;
\code{\link[unikn]{newpal}} for defining new color palettes;
\code{\link[unikn]{grepal}} for finding named colors.

Other Swedish university color palettes: 
\code{\link{gothenburg}},
\code{\link{lund_1}},
\code{\link{lund_2}},
\code{\link{lund_3}},
\code{\link{stockholm_1}},
\code{\link{uppsala}}
}
\author{
\strong{unicol}, 2023-10-12.
}
\concept{Swedish university color palettes}
\keyword{datasets}
