% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toronto.R
\docType{data}
\name{toronto_1}
\alias{toronto_1}
\title{Colors of the University of Toronto, Canada}
\format{
An object of class \code{character} of length 4.
}
\source{
Color definitions are based on the   
\href{https://new.brand.utoronto.ca/principles-and-fundamentals/colour/}{U of T's Brand Portal}.
}
\usage{
toronto_1
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{toronto_1} provides the four colors of the University of Toronto
of the \href{https://www.utoronto.ca/}{University of Toronto}, Canada.
}
\details{
The four colors are
\code{"u_of_t_blue"}, (defined as RGB 37/53/90, HEX #25355A)
\code{"boundless_blue"} (defined as RGB 0/127/163, HEX #007FA3), 
\code{"black"} (defined as RGB 0/0/0, HEX #000000), and
\code{"light_grey"} (defined as RGB 242/244/247, HEX #F2F4F7).


\code{uni_pal_1} uses the HEX color definitions.
}
\examples{
toronto_1
unikn::seecol(toronto_1, main = "University of Toronto") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other Canadian university color palettes: 
\code{\link{conc_1}},
\code{\link{conc_2}},
\code{\link{conc_3}},
\code{\link{conc_4}},
\code{\link{guelph}},
\code{\link{mun_1}},
\code{\link{mun_2}},
\code{\link{ottawa_1}},
\code{\link{ubc_1}},
\code{\link{uvic_complementary}},
\code{\link{uvic_core}}
}
\author{
\strong{unicol}, 2023-06-20.
}
\concept{Canadian university color palettes}
\keyword{datasets}
