% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCDublin.R
\docType{data}
\name{UCD}
\alias{UCD}
\title{Primary colours of University College Dublin, Ireland}
\format{
An object of class \code{character} of length 5.
}
\source{
Color definitions are based on the \href{https://www.ucd.ie}{UCD}'s 
\href{https://silo.tips/download/brand-identity-guidelines}{Brand identity guidelines} 
(PDF, Version 3, 2006).
}
\usage{
UCD
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{UCD} provides the five primary colours 
of the \href{https://www.ucd.ie}{University College Dublin}, Ireland, UK.
}
\details{
\code{UCD} is based on (approximated) HEX color values
}
\examples{
UCD
unikn::seecol(UCD, main = "Colours of the University College Dublin")  # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other Irish university color palettes: 
\code{\link{UCC_1}},
\code{\link{UCC_2}},
\code{\link{galway_1}},
\code{\link{galway_2}},
\code{\link{limerick_1}},
\code{\link{limerick_2}},
\code{\link{maynooth}},
\code{\link{trinity_1}},
\code{\link{trinity_2}}
}
\author{
Contributed to \strong{unicol} by EPH, 2023-09-13.
}
\concept{Irish university color palettes}
\keyword{datasets}
