% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCLA.R
\docType{data}
\name{ucla_3}
\alias{ucla_3}
\title{Tertiary colors of UCLA, University of California, Los Angeles, USA}
\format{
An object of class \code{character} of length 6.
}
\source{
Color definitions are based on \href{https://www.ucla.edu}{UCLA}'s 
\href{https://brand.ucla.edu/identity/colors}{Color manual}.
}
\usage{
ucla_3
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{ucla_3} provides six tertiary brand colors
of the \href{https://www.ucla.edu}{University of California, Los Angeles (UCLA)}, USA.
}
\details{
The 6 tertiary colors have different definitions for print and screen uses. 
See \href{https://www.ucla.edu}{UCLA}'s 
\href{https://brand.ucla.edu/identity/colors}{Color manual} for definitions and details. 

\code{ucla_3} uses the HEX color definitions.


A tertiary palette has been developed for use as an accent to the 
primary and secondary colors (of \code{\link{ucla_1}} and \code{\link{ucla_2}}). 

Please note that in order to maintain maximum vibrancy of these colors, 
they will appear slightly different between screen and print.
Due to printing limitations, the CMYK values are slightly duller than ideal. 

If your budget allows, select one tertiary color from the palette to include 
in your project and print it as a spot to bring the vibrancy fully to life in print.
For the sake of accessibility, restrict use of tertiary colors to graphics only.
}
\examples{
ucla_3
unikn::seecol(ucla_3, main = "The tertiary colors of UCLA")  # view color palette

}
\seealso{
\code{\link{ucla_1}} for primary colors of UCLA,; 
\code{\link{ucla_2}} for secondary colors of UCLA,; 
\code{\link{ucla_4}} for blue gradient colors of UCLA,; 
\code{\link[unikn]{seecol}} for viewing and comparing color palettes;
\code{\link[unikn]{usecol}} for using color palettes;
\code{\link[unikn]{simcol}} for finding similar colors;
\code{\link[unikn]{newpal}} for defining new color palettes;
\code{\link[unikn]{grepal}} for finding named colors.

Other U.S. university color palettes: 
\code{\link{asu_1}},
\code{\link{asu_2}},
\code{\link{asu_3}},
\code{\link{berkeley_1}},
\code{\link{berkeley_2}},
\code{\link{brown_1}},
\code{\link{brown_2}},
\code{\link{bu}},
\code{\link{caltech_1}},
\code{\link{caltech_2}},
\code{\link{caltech_3}},
\code{\link{cmu_1}},
\code{\link{cmu_2}},
\code{\link{cmu_3}},
\code{\link{columbia_1}},
\code{\link{columbia_2}},
\code{\link{columbia_3}},
\code{\link{cornell_1}},
\code{\link{cornell_2}},
\code{\link{cornell_3_accent}},
\code{\link{dartmouth_1}},
\code{\link{dartmouth_2}},
\code{\link{dartmouth_3}},
\code{\link{duke_1}},
\code{\link{duke_2}},
\code{\link{harvard_1}},
\code{\link{harvard_2}},
\code{\link{harvard_3}},
\code{\link{jhu_0}},
\code{\link{jhu_1}},
\code{\link{jhu_2}},
\code{\link{jhu_3_accent}},
\code{\link{jhu_4_gray}},
\code{\link{manchester_uni_1}},
\code{\link{manchester_uni_2}},
\code{\link{michigan_1}},
\code{\link{michigan_2}},
\code{\link{minnesotatwin_1}},
\code{\link{minnesotatwin_2}},
\code{\link{mit}},
\code{\link{monash_1}},
\code{\link{monash_2}},
\code{\link{msu}},
\code{\link{northwestern_1}},
\code{\link{northwestern_2}},
\code{\link{notredame_1}},
\code{\link{notredame_2}},
\code{\link{nyu_1}},
\code{\link{nyu_2}},
\code{\link{nyu_accent}},
\code{\link{nyu_neutral}},
\code{\link{ohio_uni_1}},
\code{\link{ohio_uni_2}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{princeton_0}},
\code{\link{princeton_1}},
\code{\link{princeton_2}},
\code{\link{rpi_1}},
\code{\link{rpi_2}},
\code{\link{rpi_3}},
\code{\link{stanford_1}},
\code{\link{stanford_2}},
\code{\link{stanford_3}},
\code{\link{uchicago_1}},
\code{\link{uchicago_2}},
\code{\link{uci_1}},
\code{\link{uci_2}},
\code{\link{uci_3}},
\code{\link{uci_4}},
\code{\link{ucla_1}},
\code{\link{ucla_2}},
\code{\link{ucla_4}},
\code{\link{ucsd_1}},
\code{\link{ucsd_2}},
\code{\link{ucsd_3}},
\code{\link{uflorida_1}},
\code{\link{uflorida_2}},
\code{\link{uflorida_3}},
\code{\link{umass_1}},
\code{\link{umass_2}},
\code{\link{umass_neutrals}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{willamette}},
\code{\link{wm_1}},
\code{\link{wm_2}},
\code{\link{wm_3}},
\code{\link{yale}},
\code{\link{yeshiva}}
}
\author{
\strong{unicol}, 2023-06-16.
}
\concept{U.S. university color palettes}
\keyword{datasets}
