% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Florida.R
\docType{data}
\name{uflorida_2}
\alias{uflorida_2}
\title{Secondary colors of the University of Florida}
\format{
An object of class \code{character} of length 5.
}
\source{
Color definitions are based on the
\href{https://brandcenter.ufl.edu/colors/}{UFlorida's brand center website}.
}
\usage{
uflorida_2
}
\value{
A names vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uflorida_2} provides the 5 secondary colors
of the \href{https://www.ufl.edu/}{University of Florida}, USA.
}
\details{
The 5 secondary colors are
\code{"florida_bottlebrush"}(defined as HEX #D32737),
\code{"florida_alachua"}(defined as HEX #F2A900),
\code{"florida_gator"}(defined as HEX #2284C),
\code{"florida_dark_blue"} (defined as HEX #002657) and
\code{"florida_perennial"} (defined as HEX #6A2A60).

\code{uflorida_2} uses the HEX color definition.
}
\examples{
uflorida_2
unikn::seecol(uflorida_2, 
              main = "Secondary colors of UFlorida") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.
}
\author{
\strong{unicol}, 2320-07-07.
}
\concept{university color palettes.}
\keyword{datasets}
