% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMass.R
\docType{data}
\name{umass_brand}
\alias{umass_brand}
\title{Brand colors of the University of Massachusetts Amherst}
\format{
An object of class \code{character} of length 3.
}
\source{
Color definitions are based on the
\href{https://www.umass.edu/brand/elements/color}{UMass Brand guide}.
}
\usage{
umass_brand
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{umass_brand} provides the two brand colors 
of the \href{https://www.umass.edu}{University of Massachusetts Amherst}, USA.
}
\details{
The brand color is
\code{"UMass Maroon"} (defined as C9 M100 Y64 K48, R136 G28 B28, Hex 881c1c)
 and should be implemented consistently across all media paired with the complementary 
\code{"Black"} (defined as C67 M44 Y67 K95, R33 G39 B33, Hex 212721) and
\code{"White"} (not defined).


\code{umass_brand} uses the HEX color definitions.
}
\examples{
umass_brand
unikn::seecol(umass_brand, main = "Brand colors of the University of Massachusetts Amherst") 
# view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colours;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colours.

Other university color palettes: 
\code{\link{UCLA_1}},
\code{\link{UCLA_2}},
\code{\link{UCLA_3}},
\code{\link{UCSD_1_core}},
\code{\link{UCSD_2_accent}},
\code{\link{UCSD_3_neutral}},
\code{\link{example_1}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{queens_1}},
\code{\link{queens_2}},
\code{\link{sfu_brand}},
\code{\link{standrews_1}},
\code{\link{standrews_2}},
\code{\link{stanford_1}},
\code{\link{stanford_2_accent}},
\code{\link{stanford_3_web}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_primary}},
\code{\link{ucalgary_secondary}},
\code{\link{ucalgary_warmgreys}},
\code{\link{umass_2}},
\code{\link{umass_neutrals}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_heidelberg_1}},
\code{\link{uni_manchester_1}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_2_print}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{uwaterloo_arts}},
\code{\link{uwaterloo_engineering}},
\code{\link{uwaterloo_environment}},
\code{\link{uwaterloo_health}},
\code{\link{uwaterloo_main}},
\code{\link{uwaterloo_math}},
\code{\link{uwaterloo_science}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{waikato}},
\code{\link{western_uni_1}},
\code{\link{yale_1}},
\code{\link{yeshiva}},
\code{\link{york_1}},
\code{\link{york_2}}
}
\author{
\strong{unicol}, 2023-06-27.
}
\concept{university color palettes}
\keyword{datasets}
