% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMass.R
\docType{data}
\name{umass_neutrals}
\alias{umass_neutrals}
\title{Neutral colors of the University of Massachusetts Amherst}
\format{
An object of class \code{character} of length 3.
}
\source{
Color definitions are based on the
\href{https://www.umass.edu/brand/elements/color}{UMass Brand guide}.
}
\usage{
umass_neutrals
}
\value{
A named vector of colours (HEX/HTML codes of type character).
}
\description{
\code{umass_brand} provides the three neutral colors 
of the \href{https://www.umass.edu}{University of Massachusetts Amherst}, USA.
}
\details{
The neutral colors are
\code{"PMS 447"} (defined as C50 M30 Y40 K90, R55 G58 B54, Hex 373a36), 
\code{"PMS 445"} (defined as C52 M23 Y30 K74, R80 G87 B89, Hex 505759), and
\code{"PMS 429"} (defined as C21 M11 Y9 K23, R162 G170 B173, Hex a2aaad).


\code{umass_neutrals} uses the HEX color definitions.
}
\examples{
umass_neutrals
unikn::seecol(umass_neutrals, main = "Neutral colors of UMass") 
# view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colours;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colours.

Other university color palettes: 
\code{\link{UCLA_1}},
\code{\link{UCLA_2}},
\code{\link{UCLA_3}},
\code{\link{UCSD_1_core}},
\code{\link{UCSD_2_accent}},
\code{\link{UCSD_3_neutral}},
\code{\link{example_1}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_primary}},
\code{\link{ucalgary_secondary}},
\code{\link{ucalgary_warmgreys}},
\code{\link{umass_2}},
\code{\link{umass_brand}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_heidelberg_1}},
\code{\link{uni_manchester_1}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_2_print}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{uwaterloo_arts}},
\code{\link{uwaterloo_engineering}},
\code{\link{uwaterloo_environment}},
\code{\link{uwaterloo_health}},
\code{\link{uwaterloo_main}},
\code{\link{uwaterloo_math}},
\code{\link{uwaterloo_science}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{waikato}},
\code{\link{western_uni_1}},
\code{\link{yale_1}},
\code{\link{yeshiva}},
\code{\link{york_1}},
\code{\link{york_2}}
}
\author{
\strong{unicol}, 2023-06-27.
}
\concept{university color palettes}
\keyword{datasets}
