% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_cornell.R
\docType{data}
\name{uni_cornell_accent_web_3}
\alias{uni_cornell_accent_web_3}
\title{Accent colors of Cornell University}
\format{
An object of class \code{character} of length 9.
}
\source{
Color definitions are based on the
\href{https://brand.cornell.edu/design-center/colors/}{Color manual}.
}
\usage{
uni_cornell_accent_web_3
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_cornell_accent_web_3} provides the nine primary colors
of the \href{https://www.cornell.edu/}{Cornell University}, USA.
}
\details{
Although our primary and secondary palettes should guide most layouts, in certain instances other colors may be needed.
For those circumstances, refer to the accent palette.
These colors should not be used as full-color bleeds and should be used periodically and in moderation.

This secondary colors are 
\code{"cornell_blue"} (defined as HEX #006699),
\code{"cornell_green_graphic"} (defined as HEX #6EB43F),
\code{"cornell_green_text"} (defined as HEX #4B7B2B),
\code{"cornell_green_text_big"} (defined as HEX #578E32),
\code{"cornell_orange_graphic"} (defined as HEX #F8981D),
\code{"cornell_orange_text"} (defined as HEX #D47500),
\code{"cornell_red_graphic"} (defined as HEX #EF4035),
\code{"cornell_red_text"} (defined as HEX #DF1E12) and
\code{"cornell_navy"} (defined as HEX #073949).
}
\examples{
uni_cornell_accent_web_3
unikn::seecol(uni_cornell_accent_web_3, main = "Accent colors Cornell Uni") # view color palette

}
\seealso{
\code{\link{uni_cornell_primary_web_1}} for primary colors of Cornell University
\code{\link{uni_cornell_secondary_web_2}} for primary colors of Cornell University
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{caltech_pal_1}},
\code{\link{caltech_pal_2}},
\code{\link{caltech_pal_3}},
\code{\link{eth_pal_1}},
\code{\link{eth_pal_2}},
\code{\link{eth_pal_3}},
\code{\link{fu_pal_0}},
\code{\link{fu_pal_1}},
\code{\link{fu_pal_2}},
\code{\link{fu_pal_3}},
\code{\link{hu_pal_1}},
\code{\link{hu_pal_2}},
\code{\link{lmu_pal_1}},
\code{\link{lmu_pal_2}},
\code{\link{lmu_pal_3}},
\code{\link{mit_1}},
\code{\link{mpg_pal}},
\code{\link{rpi_pal_1}},
\code{\link{rpi_pal_2}},
\code{\link{rpi_pal_3}},
\code{\link{rptu_pal}},
\code{\link{uni_bonn_1}},
\code{\link{uni_bonn_2}},
\code{\link{uni_california_primary_1}},
\code{\link{uni_california_secondary_2}},
\code{\link{uni_chicago_primary_1}},
\code{\link{uni_chicago_secondary_2}},
\code{\link{uni_columbia}},
\code{\link{uni_cornell_primary_web_1}},
\code{\link{uni_cornell_secondary_web_2}},
\code{\link{uni_freiburg_0}},
\code{\link{uni_freiburg_1}},
\code{\link{uni_freiburg_2}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_hamburg_1}},
\code{\link{uni_hamburg_2}},
\code{\link{uni_harvard_accent_3}},
\code{\link{uni_harvard_primary_1}},
\code{\link{uni_harvard_secondary_2}},
\code{\link{uni_hopkins_accent_4}},
\code{\link{uni_hopkins_grayscale_5}},
\code{\link{uni_hopkins_identity_1}},
\code{\link{uni_hopkins_primary_2}},
\code{\link{uni_hopkins_secondary_3}},
\code{\link{uni_jena_1}},
\code{\link{uni_jena_2}},
\code{\link{uni_kiel_1}},
\code{\link{uni_kiel_2}},
\code{\link{uni_koeln_1}},
\code{\link{uni_koeln_2}},
\code{\link{uni_konstanz_1}},
\code{\link{uni_konstanz_2}},
\code{\link{uni_mannheim_1}},
\code{\link{uni_mannheim_2}},
\code{\link{uni_michigan_primary_1}},
\code{\link{uni_michigan_secondary_2}},
\code{\link{uni_pal_1}},
\code{\link{uni_penn_logo_1}},
\code{\link{uni_penn_web_2}},
\code{\link{uni_princeton_0}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_2}},
\code{\link{uni_regensburg_3}},
\code{\link{uni_stanford_accent_2}},
\code{\link{uni_stanford_primary_1}},
\code{\link{uni_stanford_web_3}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_2_print}},
\code{\link{uni_ulm_1}},
\code{\link{uni_ulm_2}},
\code{\link{uni_yale_primary}}
}
\author{
\strong{unicol}, 2023-06-14.
}
\concept{university color palettes}
\keyword{datasets}
