% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uni_Goettingen.R
\docType{data}
\name{uni_goettingen_1}
\alias{uni_goettingen_1}
\title{Primary colors of the University of Göttingen, Germany}
\format{
An object of class \code{character} of length 4.
}
\source{
Color definitions are based on 
\url{https://www.uni-goettingen.de/de/589412.html}.
}
\usage{
uni_goettingen_1
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_goettingen_1} provides the primary blue colors 
of the \href{https://www.uni-goettingen.de/}{University of Göttingen}, Germany.
}
\details{
The primary dark blue color \code{uniblau} is defined as \code{Uni-Blau HKS 41}. 
The palette \code{uni_goettingen_1} adds some derived colors, 
while the palette \code{\link{uni_goettingen_2}} provides secondary colors.
}
\examples{
uni_goettingen_1
unikn::seecol(uni_goettingen_1, main = "The primary colors of Uni Göttingen")  # view color palette
unikn::demopal(uni_goettingen_1, type = 1, main = "Primary colors of the University of Goettingen")

}
\seealso{
\code{\link{uni_goettingen_2}} and \code{\link{uni_goettingen_3}} for alternative colors of the University of Goettingen;  
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{UCLA_1}},
\code{\link{UCLA_2}},
\code{\link{UCLA_3}},
\code{\link{UCSD_1_core}},
\code{\link{UCSD_2_accent}},
\code{\link{UCSD_3_neutral}},
\code{\link{example_1}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_primary}},
\code{\link{ucalgary_secondary}},
\code{\link{ucalgary_warmgreys}},
\code{\link{umass_2}},
\code{\link{umass_brand}},
\code{\link{umass_neutrals}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_heidelberg_1}},
\code{\link{uni_manchester_1}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_2_print}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{uwaterloo_arts}},
\code{\link{uwaterloo_engineering}},
\code{\link{uwaterloo_environment}},
\code{\link{uwaterloo_health}},
\code{\link{uwaterloo_main}},
\code{\link{uwaterloo_math}},
\code{\link{uwaterloo_science}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{waikato}},
\code{\link{western_uni_1}},
\code{\link{yale_1}},
\code{\link{yeshiva}},
\code{\link{york_1}},
\code{\link{york_2}}
}
\author{
\strong{unicol}, 2022-11-20.
}
\concept{university color palettes}
\keyword{datasets}
