% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uni_Heidelberg.R
\docType{data}
\name{uni_heidelberg_1}
\alias{uni_heidelberg_1}
\title{Official colours of the University of Heidelberg}
\format{
An object of class \code{character} of length 3.
}
\source{
Color definitions are based on the
\href{https://backend.uni-heidelberg.de/de/dokumente/corporate-design-handbuch/download}{Uni Heidelberg Corporate Design}.
}
\usage{
uni_heidelberg_1
}
\value{
A named vector of colours (HEX/HTML codes of type character).
}
\description{
\code{uni_heidelberg_1} provides the three official colours 
of the \href{https://www.uni-heidelberg.de/de}{University of Heidelberg}, Germany.
}
\details{
The three official colours are
\code{"Rot (Siegel-Hintergrund)"} (defined as  C20 / M100 / y90 / K0, r198 / g24 / b38),
\code{"Rot-Braun (Siegel-Outline)"} (defined as C30 / M100 / y90 / K65, r89 / g13 / b8), and
\code{"Sand (Zusatzfarbe)"} (defined as  C5 / M5 / y9 / K0, r244 / g241 / b234).

\code{uni_heidelberg_1} uses the RGB color definition.
}
\examples{
uni_heidelberg_1
unikn::seecol(uni_heidelberg_1, main = "Official colours of the University of Heidelberg") 
# view colour palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colours;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colours.

Other university colour palettes: 
\code{\link{birmingham_1}},
\code{\link{bu_1}},
\code{\link{dundee_background}},
\code{\link{dundee_block}},
\code{\link{dundee_core}},
\code{\link{dundee_highlight}},
\code{\link{lancaster_1}},
\code{\link{lancaster_2}},
\code{\link{mcgill_brights}},
\code{\link{mcgill_darks}},
\code{\link{mcgill_grey}},
\code{\link{mcgill_muted}},
\code{\link{mcgill_pastels}},
\code{\link{mcgill_red}},
\code{\link{mcmaster_brighterworld}},
\code{\link{mcmaster_heritage}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{queens_primary}},
\code{\link{queens_secondary}},
\code{\link{sfu_brand}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_primary}},
\code{\link{ucalgary_secondary}},
\code{\link{ucalgary_warmgreys}},
\code{\link{uvic_complementary}},
\code{\link{uvic_core}}
}
\author{
\strong{unicol}, 2023-06-27.
}
\concept{university colour palettes}
\keyword{datasets}
