% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_johns_hopkins.R
\docType{data}
\name{uni_hopkins_secondary_3}
\alias{uni_hopkins_secondary_3}
\title{Secondary colors of Johns Hopkins University}
\format{
An object of class \code{character} of length 6.
}
\source{
Color definitions are based on the
\href{https://brand.jhu.edu/color}{Color manual}.
}
\usage{
uni_hopkins_secondary_3
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_hopkins_secondary_3} provides the six secondary colors
of the \href{https://www.jhu.edu}{Johns Hopkins University}, USA.
}
\details{
The secondary color palette is designed to provide creative flexibility. 
These colors are complementary to the primary blues and were chosen to add warmth and energy to your design.

These secondary colors are
\code{"jhu_red"} (defined as PMS 173 C, CMYK 0/82/94/2, RGB 207/69/32, HEX #CF4520),
\code{"jhu_burgundy1"} (defined as PMS 188 C, CMYK 16/100/65/58, RGB 118/35/47, HEX #76232F), 
\code{"jhu_purple1"} (defined as PMS 7655 C, CMYK 33/72/0/0, RGB 161/90/149, HEX #A15A95), 
\code{"jhu_green1"} (defined as PMS 3278 C, CMYK 99/0/69/0, RGB 0/155/119, HEX #009B77), 
\code{"jhu_seablue"} (defined as PMS 285 C, CMYK 90/48/0/0, RGB 0/114/206, HEX #0072CE) and 
\code{"jhu_yellow"} (defined as PMS 7406 C, CMYK 0/20/100/2, RGB 241/196/0, HEX #F1C400).
}
\examples{
uni_hopkins_secondary_3
unikn::seecol(uni_hopkins_secondary_3, main = "Accent colors Hopkins Uni") # view color palette

}
\seealso{
\code{\link{uni_hopkins_identity_1}} for identity colors of Johns Hopkins University
\code{\link{uni_hopkins_primary_2}} for primary colors of Johns Hopkins University
\code{\link{uni_hopkins_accent_4}} for accent colors of Johns Hopkins University
\code{\link{uni_hopkins_grayscale_5}} for grey scale colors of Johns Hopkins University
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{caltech_pal_1}},
\code{\link{caltech_pal_2}},
\code{\link{caltech_pal_3}},
\code{\link{eth_pal_1}},
\code{\link{eth_pal_2}},
\code{\link{eth_pal_3}},
\code{\link{fu_pal_0}},
\code{\link{fu_pal_1}},
\code{\link{fu_pal_2}},
\code{\link{fu_pal_3}},
\code{\link{hu_pal_1}},
\code{\link{hu_pal_2}},
\code{\link{lmu_pal_1}},
\code{\link{lmu_pal_2}},
\code{\link{lmu_pal_3}},
\code{\link{mit_1}},
\code{\link{mpg_pal}},
\code{\link{rpi_pal_1}},
\code{\link{rpi_pal_2}},
\code{\link{rpi_pal_3}},
\code{\link{rptu_pal}},
\code{\link{uni_bonn_1}},
\code{\link{uni_bonn_2}},
\code{\link{uni_california_primary_1}},
\code{\link{uni_california_secondary_2}},
\code{\link{uni_chicago_primary_1}},
\code{\link{uni_chicago_secondary_2}},
\code{\link{uni_columbia}},
\code{\link{uni_cornell_accent_web_3}},
\code{\link{uni_cornell_primary_web_1}},
\code{\link{uni_cornell_secondary_web_2}},
\code{\link{uni_freiburg_0}},
\code{\link{uni_freiburg_1}},
\code{\link{uni_freiburg_2}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_hamburg_1}},
\code{\link{uni_hamburg_2}},
\code{\link{uni_harvard_accent_3}},
\code{\link{uni_harvard_primary_1}},
\code{\link{uni_harvard_secondary_2}},
\code{\link{uni_hopkins_accent_4}},
\code{\link{uni_hopkins_grayscale_5}},
\code{\link{uni_hopkins_identity_1}},
\code{\link{uni_hopkins_primary_2}},
\code{\link{uni_jena_1}},
\code{\link{uni_jena_2}},
\code{\link{uni_kiel_1}},
\code{\link{uni_kiel_2}},
\code{\link{uni_koeln_1}},
\code{\link{uni_koeln_2}},
\code{\link{uni_konstanz_1}},
\code{\link{uni_konstanz_2}},
\code{\link{uni_mannheim_1}},
\code{\link{uni_mannheim_2}},
\code{\link{uni_michigan_primary_1}},
\code{\link{uni_michigan_secondary_2}},
\code{\link{uni_pal_1}},
\code{\link{uni_penn_logo_1}},
\code{\link{uni_penn_web_2}},
\code{\link{uni_princeton_0}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_2}},
\code{\link{uni_regensburg_3}},
\code{\link{uni_stanford_accent_2}},
\code{\link{uni_stanford_primary_1}},
\code{\link{uni_stanford_web_3}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_2_print}},
\code{\link{uni_ulm_1}},
\code{\link{uni_ulm_2}},
\code{\link{uni_yale_primary}}
}
\author{
\strong{unicol}, 2023-06-14.
}
\concept{university color palettes}
\keyword{datasets}
