% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_pals.R
\docType{data}
\name{uni_kiel_1}
\alias{uni_kiel_1}
\title{Primary color of Kiel University}
\format{
An object of class \code{character} of length 3.
}
\source{
Color definitions are based on 
\url{https://www.presse.uni-kiel.de/de/erscheinungsbild/farben}.
}
\usage{
uni_kiel_1
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_kiel_1} provides the primary \code{violett} color
of \href{https://www.uni-kiel.de/en/}{Kiel University}, Germany.
}
\details{
The main color is the luminant \code{violett} of \code{uni_kiel_1[2]}, 
which alludes to the historic founding faculty of theology. 
It is alternatively defined as \code{RGB 155/10/125}, \code{Hex #9b0a7d}, 
\code{HKS 33}, or \code{CMYK 45/100/0/0}. 

The palette \code{uni_kiel_1} enframes this color as its central element 
in \code{"black"} and \code{"white"} to easily enable mixing it with 
shades of \code{"grey"}. 
See examples and the \code{\link{ac}} and \code{\link{usecol}} functions 
for creating color gradients. 

The full name of \href{https://www.uni-kiel.de/en/}{Kiel University} 
is Christian-Albrechts-Universität (CAU) zu Kiel.
}
\examples{
uni_kiel_1
unikn::seecol(uni_kiel_1, main = "The main color of Kiel University (CAU)")
unikn::demopal(uni_kiel_1, type = 3, main = "Primary colors of the University of Kiel")

# Gradient: Mixing the main color with shades of grey: 
N = 9
kiel_mix <- unikn::usecol(uni_kiel_1, n = N)[c(-1, -N)]  # remove extremes
unikn::seecol(kiel_mix, main = "A color gradient of Kiel University (CAU)")

}
\seealso{
\code{\link{uni_kiel_2}} for departmental colors of Kiel University;  
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{caltech_1}},
\code{\link{caltech_2}},
\code{\link{caltech_3}},
\code{\link{chicago_1}},
\code{\link{chicago_2}},
\code{\link{columbia}},
\code{\link{cornell_1}},
\code{\link{cornell_2}},
\code{\link{cornell_3_accent}},
\code{\link{eth_pal_1}},
\code{\link{eth_pal_2}},
\code{\link{eth_pal_3}},
\code{\link{fu_pal_0}},
\code{\link{fu_pal_1}},
\code{\link{fu_pal_2}},
\code{\link{fu_pal_3}},
\code{\link{harvard_1}},
\code{\link{harvard_2}},
\code{\link{harvard_3_accent}},
\code{\link{hu_pal_1}},
\code{\link{hu_pal_2}},
\code{\link{jhu_0}},
\code{\link{jhu_1}},
\code{\link{jhu_2}},
\code{\link{jhu_3_accent}},
\code{\link{lmu_pal_1}},
\code{\link{lmu_pal_2}},
\code{\link{lmu_pal_3}},
\code{\link{michigan_1}},
\code{\link{michigan_2}},
\code{\link{mit_pal}},
\code{\link{mpg_pal}},
\code{\link{northwestern_1}},
\code{\link{northwestern_2}},
\code{\link{nyu_1}},
\code{\link{nyu_2}},
\code{\link{nyu_accent}},
\code{\link{nyu_neutral}},
\code{\link{rpi_pal_1}},
\code{\link{rpi_pal_2}},
\code{\link{rpi_pal_3}},
\code{\link{rptu_pal}},
\code{\link{stanford_1}},
\code{\link{stanford_2_accent}},
\code{\link{stanford_3_web}},
\code{\link{uni_berkeley_1}},
\code{\link{uni_berkeley_2}},
\code{\link{uni_bonn_1}},
\code{\link{uni_bonn_2}},
\code{\link{uni_freiburg_0}},
\code{\link{uni_freiburg_1}},
\code{\link{uni_freiburg_2}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_hamburg_1}},
\code{\link{uni_hamburg_2}},
\code{\link{uni_hopkins_grayscale_5}},
\code{\link{uni_jena_1}},
\code{\link{uni_jena_2}},
\code{\link{uni_kiel_2}},
\code{\link{uni_koeln_1}},
\code{\link{uni_koeln_2}},
\code{\link{uni_konstanz_1}},
\code{\link{uni_konstanz_2}},
\code{\link{uni_mannheim_1}},
\code{\link{uni_mannheim_2}},
\code{\link{uni_pal_1}},
\code{\link{uni_princeton_0}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_2}},
\code{\link{uni_regensburg_3}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_2_print}},
\code{\link{uni_ulm_1}},
\code{\link{uni_ulm_2}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{yale_1}}
}
\author{
\strong{unicol}, 2023-01-02.
}
\concept{university color palettes}
\keyword{datasets}
