% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_old_pals.R
\docType{data}
\name{uni_konstanz_2}
\alias{uni_konstanz_2}
\title{Preferred colors of the University of Konstanz, Germany}
\format{
An object of class \code{character} of length 9.
}
\source{
R package \strong{unicol}, 
retrieved from \url{https://CRAN.R-project.org/package=unikn},    
doi \doi{10.5281/zenodo.7096191}. 

Color definitions are based on 
\url{https://www.uni-konstanz.de/en/university/news-and-media/create-online-and-print-media/corporate-design/the-university-of-konstanzs-corporate-design/}  
\url{https://www.uni-konstanz.de/en/university/news-and-media/create-online-and-print-media/corporate-design/colours-for-complex-graphics/}
}
\usage{
uni_konstanz_2
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_konstanz_2} provides the palette of preferred colors  
of the \href{https://www.uni-konstanz.de/}{University of Konstanz}, Germany.
}
\details{
The \strong{unikn} package provides many additional color palettes for 
the \href{https://www.uni-konstanz.de/}{University of Konstanz}.
}
\examples{
uni_konstanz_2
unikn::seecol(uni_konstanz_2, main = "Preferred colors of the University of Konstanz")  
unikn::demopal(uni_konstanz_2, type = 3, main = "Preferred colors of the Uni Konstanz")

}
\seealso{
\code{\link{uni_konstanz_1}} for the default colors of the University of Konstanz;  
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other German university color palettes: 
\code{\link{fu_0}},
\code{\link{fu_1}},
\code{\link{fu_2}},
\code{\link{fu_3}},
\code{\link{hu_1}},
\code{\link{hu_2}},
\code{\link{lmu_1}},
\code{\link{lmu_2}},
\code{\link{lmu_3}},
\code{\link{rptu}},
\code{\link{uni_bonn_1}},
\code{\link{uni_bonn_2}},
\code{\link{uni_freiburg_0}},
\code{\link{uni_freiburg_1}},
\code{\link{uni_freiburg_2}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_hamburg_1}},
\code{\link{uni_hamburg_2}},
\code{\link{uni_heidelberg}},
\code{\link{uni_jena_1}},
\code{\link{uni_jena_2}},
\code{\link{uni_kassel}},
\code{\link{uni_kiel_1}},
\code{\link{uni_kiel_2}},
\code{\link{uni_koeln_1}},
\code{\link{uni_koeln_2}},
\code{\link{uni_konstanz_1}},
\code{\link{uni_mannheim_1}},
\code{\link{uni_mannheim_2}},
\code{\link{uni_potsdam}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_2}},
\code{\link{uni_regensburg_3}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_print}},
\code{\link{uni_ulm_1}},
\code{\link{uni_ulm_2}}
}
\author{
\strong{unicol}, 2022-10-20.
}
\concept{German university color palettes}
\keyword{datasets}
