% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_northwestern.R
\docType{data}
\name{uni_northwestern_secondary_2}
\alias{uni_northwestern_secondary_2}
\title{Secondary colors of Northwestern University}
\format{
An object of class \code{character} of length 12.
}
\source{
Color definitions are based on the
\href{https://www.northwestern.edu/brand/visual-identity/color-palettes/}{Color manual}.
}
\usage{
uni_northwestern_secondary_2
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_northwestern_secondary_2} provides the 12 secondary colors
of the \href{https://www.northwestern.edu}{Northwestern University}, USA.
}
\details{
The secondary colors should be used rarely and sparingly.
Under no circumstances should any of them become the predominant color for a school, center, institute or department.
¨
These colors are
\code{"bright_green"} (defined as CMYK 68,00,100,00, RGB 88,185,71, HEX #58B947),
\code{"bright_blue"} (defined as CMYK 48/00/22/00, RGB 127/206/205, HEX #7FCECD),
\code{"bright_ocean"} (defined as CMYK 68/34/00/00, RGB 80/145/205, HEX #5091CD),
\code{"bright_yellow"} (defined as CMYK 10/00/89/00, RGB 237/233/59, HEX #EDE93B),
\code{"bright_orange"} (defined as CMYK 00/23/95/00, RGB 255/197/32, HEX #FFC520),
\code{"bright_red"} (defined as CMYK 01/82/80/00, RGB 239/085/063, HEX #EF553F),
\code{"dark_green"} (defined as CMYK 87/24/83/09, RGB 000/134/086, HEX #008656),
\code{"dark_blue"} (defined as CMYK 86/39/23/01, RGB 000/127/164, HEX #007FA4),
\code{"dark_ocean"} (defined as CMYK 100/85/05/36, RGB 013/045/108, HEX #0D2D6C),
\code{"dark_yellow"} (defined as CMYK 18/14/100/0, RGB 217/200/038, HEX #D9C826),
\code{"dark_orange1"} (defined as CMYK 00/48/97/21, RGB 202/124/027, HEX #CA7C1B) and
\code{"dark_orange2"} (defined as CMYK 02/77/100/09, RGB 216/088/032, HEX #D85820.
}
\examples{
uni_northwestern_secondary_2
unikn::seecol(uni_northwestern_secondary_2, main = "Secondary colors NW") # view color palette

}
\seealso{
\code{\link{uni_northwestern_primary_1}} for primary colors of Northwestern University
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{caltech_pal_1}},
\code{\link{caltech_pal_2}},
\code{\link{caltech_pal_3}},
\code{\link{eth_pal_1}},
\code{\link{eth_pal_2}},
\code{\link{eth_pal_3}},
\code{\link{fu_pal_0}},
\code{\link{fu_pal_1}},
\code{\link{fu_pal_2}},
\code{\link{fu_pal_3}},
\code{\link{hu_pal_1}},
\code{\link{hu_pal_2}},
\code{\link{lmu_pal_1}},
\code{\link{lmu_pal_2}},
\code{\link{lmu_pal_3}},
\code{\link{mit_pal}},
\code{\link{mpg_pal}},
\code{\link{nyu_1}},
\code{\link{nyu_2}},
\code{\link{nyu_accent}},
\code{\link{nyu_neutral}},
\code{\link{rpi_pal_1}},
\code{\link{rpi_pal_2}},
\code{\link{rpi_pal_3}},
\code{\link{rptu_pal}},
\code{\link{uni_berkeley_primary_1}},
\code{\link{uni_berkeley_secondary_2}},
\code{\link{uni_bonn_1}},
\code{\link{uni_bonn_2}},
\code{\link{uni_chicago_primary_1}},
\code{\link{uni_chicago_secondary_2}},
\code{\link{uni_columbia}},
\code{\link{uni_cornell_accent_web_3}},
\code{\link{uni_cornell_primary_web_1}},
\code{\link{uni_cornell_secondary_web_2}},
\code{\link{uni_freiburg_0}},
\code{\link{uni_freiburg_1}},
\code{\link{uni_freiburg_2}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_hamburg_1}},
\code{\link{uni_hamburg_2}},
\code{\link{uni_harvard_accent_3}},
\code{\link{uni_harvard_primary_1}},
\code{\link{uni_harvard_secondary_2}},
\code{\link{uni_hopkins_accent_4}},
\code{\link{uni_hopkins_grayscale_5}},
\code{\link{uni_hopkins_identity_1}},
\code{\link{uni_hopkins_primary_2}},
\code{\link{uni_hopkins_secondary_3}},
\code{\link{uni_jena_1}},
\code{\link{uni_jena_2}},
\code{\link{uni_kiel_1}},
\code{\link{uni_kiel_2}},
\code{\link{uni_koeln_1}},
\code{\link{uni_koeln_2}},
\code{\link{uni_konstanz_1}},
\code{\link{uni_konstanz_2}},
\code{\link{uni_mannheim_1}},
\code{\link{uni_mannheim_2}},
\code{\link{uni_michigan_primary_1}},
\code{\link{uni_michigan_secondary_2}},
\code{\link{uni_northwestern_primary_1}},
\code{\link{uni_pal_1}},
\code{\link{uni_penn_logo_1}},
\code{\link{uni_penn_web_2}},
\code{\link{uni_princeton_0}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_2}},
\code{\link{uni_regensburg_3}},
\code{\link{uni_stanford_1}},
\code{\link{uni_stanford_accent}},
\code{\link{uni_stanford_web}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_2_print}},
\code{\link{uni_ulm_1}},
\code{\link{uni_ulm_2}},
\code{\link{uni_yale_primary}}
}
\author{
\strong{unicol}, 2023-06-14.
}
\concept{university color palettes}
\keyword{datasets}
