% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Potsdam.R
\docType{data}
\name{uni_potsdam}
\alias{uni_potsdam}
\title{Colors of the University of Potsdam, Germany}
\format{
An object of class \code{character} of length 9.
}
\source{
Color definitions are based on the 
\href{https://www.uni-potsdam.de/fileadmin/projects/zim/files/Printmedien_Leitfaden_2020-21.pdf}{University of Potsdam's Leitfaden (pdf)}.
}
\usage{
uni_potsdam
}
\value{
A names vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_potsdam} provides the nine primary colors
of the \href{https://www.uni-potsdam.de/de/}{University of Potsdam}, Germany.
}
\details{
The 9 primary colors are 
\code{"hks_41"} (defined as HEX #00305e), 
\code{"hks_41_45"} (defined as HEX #818bac), 
\code{"web_color"} (defined as HEX #c1d3e0), 
\code{"faculty_of_law"} (defined as HEX #e4003a), 
\code{"faculty_of_philosophy"} (defined as HEX #8b2939), 
\code{"faculty_of_humanities"} (defined as HEX #f59c00), 
\code{"faculty_of_economics"} (defined as HEX #52822f), 
\code{"faculty_of_mathematics"} (defined as HEX #0080b5), and 
\code{"faculty_of_digital_engineering"} (defined as HEX #5A5F62).

\code{uni_potsdam} uses the HEX color definitions.
}
\examples{
uni_potsdam
unikn::seecol(uni_potsdam, main = "University of Potsdam") # view color palette.

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other German university color palettes: 
\code{\link{uni_kassel}}
}
\author{
\strong{unicol}, 2023-07-21.
}
\concept{German university color palettes}
\keyword{datasets}
