% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_standford.R
\docType{data}
\name{uni_stanford_web_3}
\alias{uni_stanford_web_3}
\title{Web colors of Stanford University}
\format{
An object of class \code{character} of length 9.
}
\source{
Color definitions are based on the
\href{https://identity.stanford.edu/design-elements/color/}{Color manual}.
}
\usage{
uni_stanford_web_3
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_stanford_web_3} provides the nine web colors
of the \href{https://www.stanford.edu/}{Stanford University}, USA.
}
\details{
\code{"digital_red"} is not a replacement for Cardinal red—it can be used as an accent color on digital communications to add energy and emphasis.
\code{"digital_red_light"} can be used for interactive elements over dark backgrounds for contrast. \code{"digital_red_dark"} can be used for smaller text over light backgrounds.
Use \code{"digital_blue"} for links over light backgrounds, \code{"digital_blue_dark"} for link hover/focus states, and \code{"digital_blue_light"} for links over dark backgrounds. Do not use these shades of blue in contexts other than links.
Use \code{"digital_green"} for things such as form validation, \code{"digital_green_dark"} for link hover/focus states, and \code{"digital_green_light"} for links over dark backgrounds.

These nine web colors are
\code{"digital_red"} (defined as HEX #B1040E),
\code{"digital_red_light"} (defined as HEX #E50808),
\code{"digital_red_dark"} (defined as HEX #820000),
\code{"digital_blue"} (defined as HEX #006CB8),
\code{"digital_blue_light"} (defined as HEX #6FC3FF),
\code{"digital_blue_dark"} (defined as HEX #00548f),
\code{"digital_green"} (defined as HEX #008566),
\code{"digital_green_light"} (defined as HEX #1AECBA) and
\code{"digital_green_dark"} (defined as HEX #006F54).
}
\examples{
uni_stanford_web_3
unikn::seecol(uni_stanford_web_3, main = "Web colors Stanford University") # view color palette

}
\seealso{
\code{\link{uni_stanford_primary_1}} for primary colors of Stanford University
\code{\link{uni_stanford_accent_2}} for accent colors of Stanford University
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{caltech_pal_1}},
\code{\link{caltech_pal_2}},
\code{\link{caltech_pal_3}},
\code{\link{eth_pal_1}},
\code{\link{eth_pal_2}},
\code{\link{eth_pal_3}},
\code{\link{fu_pal_0}},
\code{\link{fu_pal_1}},
\code{\link{fu_pal_2}},
\code{\link{fu_pal_3}},
\code{\link{hu_pal_1}},
\code{\link{hu_pal_2}},
\code{\link{lmu_pal_1}},
\code{\link{lmu_pal_2}},
\code{\link{lmu_pal_3}},
\code{\link{mit_1}},
\code{\link{mpg_pal}},
\code{\link{rpi_pal_1}},
\code{\link{rpi_pal_2}},
\code{\link{rpi_pal_3}},
\code{\link{rptu_pal}},
\code{\link{uni_bonn_1}},
\code{\link{uni_bonn_2}},
\code{\link{uni_california_primary_1}},
\code{\link{uni_california_secondary_2}},
\code{\link{uni_chicago_primary_1}},
\code{\link{uni_chicago_secondary_2}},
\code{\link{uni_columbia}},
\code{\link{uni_cornell_accent_web_3}},
\code{\link{uni_cornell_primary_web_1}},
\code{\link{uni_cornell_secondary_web_2}},
\code{\link{uni_freiburg_0}},
\code{\link{uni_freiburg_1}},
\code{\link{uni_freiburg_2}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_hamburg_1}},
\code{\link{uni_hamburg_2}},
\code{\link{uni_harvard_accent_3}},
\code{\link{uni_harvard_primary_1}},
\code{\link{uni_harvard_secondary_2}},
\code{\link{uni_hopkins_accent_4}},
\code{\link{uni_hopkins_grayscale_5}},
\code{\link{uni_hopkins_identity_1}},
\code{\link{uni_hopkins_primary_2}},
\code{\link{uni_hopkins_secondary_3}},
\code{\link{uni_jena_1}},
\code{\link{uni_jena_2}},
\code{\link{uni_kiel_1}},
\code{\link{uni_kiel_2}},
\code{\link{uni_koeln_1}},
\code{\link{uni_koeln_2}},
\code{\link{uni_konstanz_1}},
\code{\link{uni_konstanz_2}},
\code{\link{uni_mannheim_1}},
\code{\link{uni_mannheim_2}},
\code{\link{uni_michigan_primary_1}},
\code{\link{uni_michigan_secondary_2}},
\code{\link{uni_pal_1}},
\code{\link{uni_penn_logo_1}},
\code{\link{uni_penn_web_2}},
\code{\link{uni_princeton_0}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_2}},
\code{\link{uni_regensburg_3}},
\code{\link{uni_stanford_accent_2}},
\code{\link{uni_stanford_primary_1}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_2_print}},
\code{\link{uni_ulm_1}},
\code{\link{uni_ulm_2}},
\code{\link{uni_yale_primary}}
}
\author{
\strong{unicol}, 2023-06-14.
}
\concept{university color palettes}
\keyword{datasets}
