% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uni_Stuttgart.R
\docType{data}
\name{uni_stuttgart_2_print}
\alias{uni_stuttgart_2_print}
\title{Printing colors of the University of Stuttgart, Germany}
\format{
An object of class \code{character} of length 6.
}
\source{
Color definitions are based on the
\href{https://www.beschaeftigte.uni-stuttgart.de/uni-services/oeffentlichkeitsarbeit/corporate-design/cd-dateien/Uni_Stuttgart_CD-Manual.pdf}{CD manual (PDF)}.
}
\usage{
uni_stuttgart_2_print
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_stuttgart_2_print} provides three primary colors, one secondary color and two alternative printing colors
of the \href{https://www.uni-stuttgart.de/en}{University of Stuttgart}, Germany.
}
\details{
The three primary colors are
\code{"white"},
\code{"anthrazit"} (defined as PANTONE 432U, CMYK 50/35/25/70, RGB 62/68/76, HEX #323232), and
\code{"mittelblau"} (defined as PANTONE 293U, CMYK 100/70/0/0, RGB 0/81/158, HEX #004191).

The secondary color is
\code{"hellblau"} (defined as PANTONE 2985U, CMYK 70/0/0/0, RGB 0/190/255, HEX #00BEFF)
and used for color accents.

The alternative printing colors are provided to replace \code{"anthrazit"} for better contrast. Use
\code{"anthrazit dark"} (defined as CMYK 0/0/0/90) when printing black and white, and
\code{"black"} (defined as CMYK 0/0/0/100) when printing text block in correspondence.

\code{uni_stuttgart_2_print} uses the HEX color definitions.
CMYK color differs from RGB and HEX definition.
For \code{"anthrazit"} and \code{"mittelblau"} the RGB and HEX definitions also differ numerically,
yet are visually indistinguishable. Our color palette is based on the HEX definition.
}
\examples{
uni_stuttgart_2_print
unikn::seecol(uni_stuttgart_2_print, main = "Printing colors Uni Stuttgart") # view color palette

}
\seealso{
\code{\link{uni_stuttgart_1}} for Stuttgart's primary colors without printing colors;
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{example_1}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_heidelberg_1}},
\code{\link{uni_manchester_1}},
\code{\link{uni_stuttgart_1}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{uwaterloo_arts}},
\code{\link{uwaterloo_engineering}},
\code{\link{uwaterloo_environment}},
\code{\link{uwaterloo_health}},
\code{\link{uwaterloo_main}},
\code{\link{uwaterloo_math}},
\code{\link{uwaterloo_science}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{waikato}},
\code{\link{western_uni_1}},
\code{\link{yale_1}},
\code{\link{yeshiva}},
\code{\link{york_1}},
\code{\link{york_2}}
}
\author{
\strong{unicol}, 2023-05-31.
}
\concept{university color palettes}
\keyword{datasets}
