% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_start_unicol.R
\name{unicol.guide}
\alias{unicol.guide}
\title{Open the unicol package guides}
\usage{
unicol.guide()
}
\value{
No return value, called for side effects.
}
\description{
\code{unicol.guide} allows accessing the \strong{unicol} package vignettes, 
which illustrate the package contents and some ways in which these can be used.
}
