% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_all_pals.R
\docType{data}
\name{unicol_data}
\alias{unicol_data}
\title{The color palettes of the \strong{unicol} package}
\format{
An object of class \code{data.frame} with 211 rows and 4 columns.
}
\usage{
unicol_data
}
\value{
A data frame summarizing the unicol color palettes.
}
\description{
\code{unicol_data} summarizes the color palettes 
included in the \strong{unicol} package 
(as a data frame).
}
\examples{
dim(unicol_data)
head(unicol_data)

# Descriptives: 
length(unique(unicol_data$pal))   # number of color palettes
length(unique(unicol_data$inst))  # number of institutions

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.
}
\concept{internal}
\keyword{datasets}
