% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UVic.R
\docType{data}
\name{uvic_1}
\alias{uvic_1}
\title{Core colours of the University of Victoria, Canada}
\format{
An object of class \code{character} of length 4.
}
\source{
Colour definitions are based on \href{https://www.uvic.ca}{UVic}'s 
\href{https://www.uvic.ca/brand/look-feel/colour/palette/}{brand information}.
}
\usage{
uvic_1
}
\value{
A named vector of colours (HEX/HTML codes of type character).
}
\description{
\code{uvic_1} provides four core colours 
of the \href{https://www.uvic.ca}{University of Victoria}, Canada.
}
\details{
The 4 core colours are 
\code{"UVic blue"} (defined as HEX #005493 or RGB 0/94/184), 
\code{"UVic yellow"} (defined as HEX #F5AA1C or RGB 234/170/0),
\code{"UVic red"} (defined as HEX #C63527 or RGB 210/38/48), and
\code{"UVic dark blue"} (defined as HEX #002754 or RGB 0/46/93).

\code{uvic_1} is based on the HEX colour definitions.


The 4 core \href{https://www.uvic.ca}{UVic} colours can be used in various combinations 
depending on audience and medium. The dominant colours are blue and yellow, 
with red used as an accent colour. 
Dark blue works well as a background and type colour. 

The core colours have to be present on all external communications 
and marketing materials to establish and build recognition by association.

HEX colour definitions are used for web,  
RGB colour definitions are used for electronic displays.
}
\examples{
uvic_1
unikn::seecol(uvic_1, main = "Core colours of the University of Victoria")  # view palette

}
\seealso{
\code{\link{uvic_2}} for the complementary colour palette of University of Victoria;
\code{\link[unikn]{seecol}} for viewing and comparing colour palettes;
\code{\link[unikn]{usecol}} for using colour palettes;
\code{\link[unikn]{simcol}} for finding similar colours;
\code{\link[unikn]{newpal}} for defining new colour palettes;
\code{\link[unikn]{grepal}} for finding named colours.

Other Canadian university color palettes: 
\code{\link{carleton}},
\code{\link{concordia_1}},
\code{\link{concordia_2}},
\code{\link{concordia_3}},
\code{\link{concordia_4}},
\code{\link{guelph}},
\code{\link{laval}},
\code{\link{manitoba_1}},
\code{\link{manitoba_2}},
\code{\link{mcgill_brights}},
\code{\link{mcgill_darks}},
\code{\link{mcgill_grey}},
\code{\link{mcgill_muted}},
\code{\link{mcgill_pastels}},
\code{\link{mcgill_red}},
\code{\link{mcmaster_brighterworld}},
\code{\link{mcmaster_heritage}},
\code{\link{mun_1}},
\code{\link{mun_2}},
\code{\link{queensu_1}},
\code{\link{queensu_2}},
\code{\link{sfu_brand}},
\code{\link{ubc}},
\code{\link{ucalgary_1}},
\code{\link{ucalgary_2}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_warmgreys}},
\code{\link{uottawa}},
\code{\link{utoronto}},
\code{\link{uvic_2}},
\code{\link{uwaterloo_arts}},
\code{\link{uwaterloo_engineering}},
\code{\link{uwaterloo_environment}},
\code{\link{uwaterloo_health}},
\code{\link{uwaterloo_main}},
\code{\link{uwaterloo_math}},
\code{\link{uwaterloo_science}},
\code{\link{western_uni}},
\code{\link{yorku_1}},
\code{\link{yorku_2}}
}
\author{
\strong{unicol}, 2023-06-29.
}
\concept{Canadian university color palettes}
\keyword{datasets}
