% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UWaterloo.R
\docType{data}
\name{uwaterloo_engineering}
\alias{uwaterloo_engineering}
\title{Faculty of Engineering colors of the University of Waterloo, CA}
\format{
An object of class \code{character} of length 4.
}
\source{
Color definitions are based on the University of Waterloo's 
\href{https://uwaterloo.ca/brand/}{Brand home}.
}
\usage{
uwaterloo_engineering
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uwaterloo_engineering} provides the purple color levels
of the \href{https://uwaterloo.ca/}{University of Waterloo}, Canada.
}
\details{
Use the colour purple to identify content and communications from the Faculty of Engineering.

The four purple colors are 
\code{"Purple Level 1"} (defined as CMYK: 27 28 0 0	, RGB: 208 180 239, HEX:	#D0B4E7),
\code{"Purple Level 2"} (defined as CMYK: 51 71 0 0, RGB:	190 51 218, HEX:	#BE33DA),
\code{"Purple Level 3 (print primary color)"} (defined as CMYK: 60 72 0 0, RGB:	129 0 180, HEX:	#8100B4), and
\code{"Purple Level 4 (digital primary color)"} (defined as CMYK: 78 94 0 0, RGB:	87 5 139, HEX:	#57058B).

\code{uwaterloo_engineering} uses the HEX color definitions.
}
\examples{
uwaterloo_engineering
unikn::seecol(uwaterloo_engineering, main = "Engineering Faculty colors of UWaterloo") 
# view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{example_1}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{uni_manchester}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_print}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{uwaterloo_arts}},
\code{\link{uwaterloo_environment}},
\code{\link{uwaterloo_health}},
\code{\link{uwaterloo_main}},
\code{\link{uwaterloo_math}},
\code{\link{uwaterloo_science}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{waikato}},
\code{\link{western_uni_1}},
\code{\link{yale_1}},
\code{\link{yeshiva}},
\code{\link{york_1}},
\code{\link{york_2}}
}
\author{
\strong{unicol}, 2023-06-29.
}
\concept{university color palettes}
\keyword{datasets}
