% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UWaterloo.R
\docType{data}
\name{uwaterloo_health}
\alias{uwaterloo_health}
\title{Faculty of Health colors of the University of Waterloo, CA}
\format{
An object of class \code{character} of length 4.
}
\source{
Color definitions are based on the University of Waterloo's 
\href{https://uwaterloo.ca/brand/}{Brand home}.
}
\usage{
uwaterloo_health
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uwaterloo_health} provides the teal color levels
of the \href{https://uwaterloo.ca/}{University of Waterloo}, Canada.
}
\details{
Use the colour teal to identify content and communications from the Faculty of Health.

The four teal colors are 
\code{"Teal Level 1"} (defined as CMYK: 30 0 6 0, RGB:	151 223 239, HEX:	#97DFEF),
\code{"Teal Level 2"} (defined as CMYK: 69 0 16 0, RGB:	0 190 208, HEX:	#00BED0),
\code{"Teal Level 3 (print primary color)"} (defined as CMYK: 100 0 30 2, RGB:	0 152 165, HEX:	#0098A5), and
\code{"Teal Level 4 (digital primary color)"} (defined as CMYK: 83 14 23 50, RGB:	0 89 99, HEX:	#005963).

\code{uwaterloo_health} uses the HEX color definitions.
}
\examples{
uwaterloo_health
unikn::seecol(uwaterloo_health, main = "health Faculty colors of UWaterloo") 
# view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{example_1}},
\code{\link{oxford_error}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{uni_manchester}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_print}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{uwaterloo_arts}},
\code{\link{uwaterloo_engineering}},
\code{\link{uwaterloo_environment}},
\code{\link{uwaterloo_main}},
\code{\link{uwaterloo_math}},
\code{\link{uwaterloo_science}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{waikato}},
\code{\link{western_uni_1}},
\code{\link{yale_1}},
\code{\link{yeshiva}},
\code{\link{york_1}},
\code{\link{york_2}}
}
\author{
\strong{unicol}, 2023-06-29.
}
\concept{university color palettes}
\keyword{datasets}
