% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UWaterloo.R
\docType{data}
\name{uwaterloo_science}
\alias{uwaterloo_science}
\title{Faculty of Science colors of the University of Waterloo, CA}
\format{
An object of class \code{character} of length 4.
}
\source{
Color definitions are based on the University of Waterloo's 
\href{https://uwaterloo.ca/brand/}{Brand home}.
}
\usage{
uwaterloo_science
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uwaterloo_science} provides the teal color levels
of the \href{https://uwaterloo.ca/}{University of Waterloo}, Canada.
}
\details{
Use the colour blue to identify content and communications from the Faculty of Science.

The four blue colors are 
\code{"Blue Level 1"} (defined as CMYK: 28 9 0 0, RGB:	180 213 255, HEX:	#B4D5FF),
\code{"Blue Level 2"} (defined as CMYK: 62 28 2 3, RGB:	99 160 255, HEX:	#63A0FF),
\code{"Blue Level 3 (print primary color)"} (defined as CMYK: 90 48 0 0, RGB:	0 115 206, HEX:	#0073CE), and
\code{"Blue Level 4 (digital primary color)"} (defined as CMYK: 100 75 0 5, RGB:	0 51 190, HEX:	#0033BE).

\code{uwaterloo_science} uses the HEX color definitions.
}
\examples{
uwaterloo_science
unikn::seecol(uwaterloo_science, main = "Science Faculty colors of UWaterloo") 
# view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{UCLA_1}},
\code{\link{UCLA_2}},
\code{\link{UCLA_3}},
\code{\link{UCSD_1_core}},
\code{\link{UCSD_2_accent}},
\code{\link{UCSD_3_neutral}},
\code{\link{example_1}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{queens_1}},
\code{\link{queens_2}},
\code{\link{sfu_brand}},
\code{\link{stanford_1}},
\code{\link{stanford_2_accent}},
\code{\link{stanford_3_web}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_primary}},
\code{\link{ucalgary_secondary}},
\code{\link{ucalgary_warmgreys}},
\code{\link{umass_2}},
\code{\link{umass_brand}},
\code{\link{umass_neutrals}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_heidelberg_1}},
\code{\link{uni_manchester_1}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_2_print}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{uwaterloo_arts}},
\code{\link{uwaterloo_engineering}},
\code{\link{uwaterloo_environment}},
\code{\link{uwaterloo_health}},
\code{\link{uwaterloo_main}},
\code{\link{uwaterloo_math}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{waikato}},
\code{\link{western_uni_1}},
\code{\link{yale_1}},
\code{\link{yeshiva}},
\code{\link{york_1}},
\code{\link{york_2}}
}
\author{
\strong{unicol}, 2023-06-29.
}
\concept{university color palettes}
\keyword{datasets}
