% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vanderbilt.R
\docType{data}
\name{vanderbilt_3}
\alias{vanderbilt_3}
\title{Saturated colors of the Vanderbilt University, USA}
\format{
An object of class \code{character} of length 4.
}
\source{
Color definitions are based on 
\href{https://brand.vanderbilt.edu/color/}{Vanderbilt's brand style guide}.
}
\usage{
vanderbilt_3
}
\value{
A named vector of colors (HEX/HTML of type character).
}
\description{
\code{vanderbilt_3} provides the four saturated colors 
of the \href{https://www.vanderbilt.edu/}{Vanderbilt University}, USA.
}
\details{
The 4 saturated colors are
\code{"vanderbilt_sky"} (defined as HEX #B3C9CD),
\code{"vanderbilt_highlight"} (defined as HEX #ECB748),
\code{"vanderbilt_oak"} (defined as HEX #946E24), and 
\code{"vanderbilt_sage"} (defined as HEX #8BA18E).

\code{vanderbilt_3} uses the HEX color definitions.
}
\examples{
vanderbilt_3
unikn::seecol(vanderbilt_3, main = "Vanderbilt University 3 (saturated)") # view color palette

}
\seealso{
\code{\link{vanderbilt_1}} for core colors of Vanderbilt University;
\code{\link{vanderbilt_2}} for neutral colors of Vanderbilt University;
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{example_1}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{uni_manchester}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_print}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{uwaterloo_arts}},
\code{\link{uwaterloo_engineering}},
\code{\link{uwaterloo_environment}},
\code{\link{uwaterloo_health}},
\code{\link{uwaterloo_main}},
\code{\link{uwaterloo_math}},
\code{\link{uwaterloo_science}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{waikato}},
\code{\link{western_uni_1}},
\code{\link{yale_1}},
\code{\link{yeshiva}},
\code{\link{york_1}},
\code{\link{york_2}}
}
\author{
\strong{unicol}, 2023-07-12.
}
\concept{university color palettes}
\keyword{datasets}
