% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Waikato.R
\docType{data}
\name{waikato}
\alias{waikato}
\title{Colors of the University of Waikato, NZ}
\format{
An object of class \code{character} of length 8.
}
\source{
Color definitions are based on Waikato's 
\href{https://www.waikato.ac.nz/web-guides/brand-standards/colours}{brand standards}.
}
\usage{
waikato
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{waikato} provides the eight official web colors 
of the \href{https://www.waikato.ac.nz/}{University of Waikato}, New Zealand.
}
\details{
The 8 official web colors are
\code{"waikato_light_red"} (defined as HEX #D40100),
\code{"waikato_medium_red"} (defined as HEX #BE0403),
\code{"waikato_dark_red"} (defined as HEX #AA0000),
\code{"waikato_light_yellow"} (defined as HEX #FAA61A),
\code{"waikato_darker_yellow"} (defined as HEX #E79000),
\code{"waikato_dark_grey"} (defined as HEX #353535),
\code{"waikato_blue_grey"} (defined as HEX #32373D), and
\code{"waikato_light_grey"} (defined as HEX #F2F2F2).

\code{waikato} uses the HEX color definitions.
}
\examples{
waikato
unikn::seecol(waikato, main = "Colors of the University of Waikato") # view color palettes

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.
}
\author{
\strong{unicol}, 2023-07-05.
}
\concept{New Zealand university color palettes}
\keyword{datasets}
