% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Western_Uni.R
\docType{data}
\name{western_uni_1}
\alias{western_uni_1}
\title{Official colours of Western University}
\format{
An object of class \code{character} of length 3.
}
\source{
Color definitions are based on the
\href{https://www.communications.uwo.ca/western_brand/vis_toolkit/colours.html}{Western's colour toolkit}.
}
\usage{
western_uni_1
}
\value{
A named vector of colours (HEX/HTML codes of type character).
}
\description{
\code{western_uni_1} provides the three official colours 
of the \href{https://www.uwo.ca}{Western University}, Canada.
}
\details{
The three primary colours are
\code{"Purple"} (defined as R:79 G:38 B:131, Hex: #4F2683),
\code{"Grey (Silver)"} (defined as R:128 G:127 B:131, Hex: #807F83), and
\code{"white"} (not defined).

\code{western_uni_1} uses the HEX color definitions.
}
\examples{
western_uni_1
unikn::seecol(western_uni_1, main = "Official colours of Western University") 
# view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colours;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colours.

Other university color palettes: 
\code{\link{example_1}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{uni_manchester}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_print}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{uwaterloo_arts}},
\code{\link{uwaterloo_engineering}},
\code{\link{uwaterloo_environment}},
\code{\link{uwaterloo_health}},
\code{\link{uwaterloo_main}},
\code{\link{uwaterloo_math}},
\code{\link{uwaterloo_science}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{waikato}},
\code{\link{yale_1}},
\code{\link{yeshiva}},
\code{\link{york_1}},
\code{\link{york_2}}
}
\author{
\strong{unicol}, 2023-06-27.
}
\concept{university color palettes}
\keyword{datasets}
