% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yale.R
\docType{data}
\name{yale_1}
\alias{yale_1}
\title{Primary colors of Yale University, USA}
\format{
An object of class \code{character} of length 10.
}
\source{
Color definitions are based on the
\href{https://yaleidentity.yale.edu/colors}{Yale identity website}.
}
\usage{
yale_1
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{yale_1} provides the ten primary colors
of the \href{https://www.yale.edu}{Yale University}, USA.
}
\details{
Yale Blue is the university’s identifying color, and has the hex value #00356b. 
The standard range of blue tints is shown in the primary color palette below along with the standard range of grays and two accent colors. 
Additional colors may be used as long as they complement, rather than clash with, the colors in the primary color palette.
}
\examples{
yale_1
unikn::seecol(yale_1, main = "Primary colors Yale University") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{example_1}},
\code{\link{oxford_error}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{uni_manchester}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_print}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{uwaterloo_arts}},
\code{\link{uwaterloo_engineering}},
\code{\link{uwaterloo_environment}},
\code{\link{uwaterloo_health}},
\code{\link{uwaterloo_main}},
\code{\link{uwaterloo_math}},
\code{\link{uwaterloo_science}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{waikato}},
\code{\link{western_uni_1}},
\code{\link{yeshiva}},
\code{\link{york_1}},
\code{\link{york_2}}
}
\author{
\strong{unicol}, 2023-08-07.
}
\concept{university color palettes}
\keyword{datasets}
