% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yeshiva.R
\docType{data}
\name{yeshiva}
\alias{yeshiva}
\title{Colors of Yeshiva University, USA}
\format{
An object of class \code{character} of length 3.
}
\source{
Color definitions are based on 
\href{https://www.yu.edu/marketing/branding}{YU's website}.
}
\usage{
yeshiva
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{yeshiva} provides the 3 colors 
of the \href{https://www.yu.edu/}{Yeshiva University}, USA.
}
\details{
The 3 colors of the Yeshiva University are
\code{"yeshiva_blue"} (defined as HEX #325A89), 
\code{"yeshiva_black"} (defined as HEX #000000), and
\code{"yeshiva_gray"} (defined as HEX #85878A). 

\code{yeshiva} uses the HEX color definitions.
}
\examples{
yeshiva
unikn::seecol(yeshiva, main = "Yeshiva University") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{example_1}},
\code{\link{oxford_error}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{uni_manchester}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_print}},
\code{\link{yale_1}},
\code{\link{york_1}},
\code{\link{york_2}}
}
\author{
\strong{unicol}, 2023-07-13.
}
\concept{university color palettes}
\keyword{datasets}
