% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/York.R
\docType{data}
\name{york_1}
\alias{york_1}
\title{Primary colors of the York University, Canada}
\format{
An object of class \code{character} of length 3.
}
\source{
Color definitions are based on the
\href{https://www.yorku.ca/brand/wp-content/uploads/sites/18/2021/07/YORKU_Brand_Standards_July2021.pdf}{York Brand Standards (PDF)}.
}
\usage{
york_1
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{york_1} provides the three primary colors 
of the \href{https://www.yorku.ca/}{York University}, Canada.
}
\details{
The three primary colors are
\code{"york_red"} (defined as RGB 227/24/55, HEX #E31837),
\code{"white"} (defined as RGB 255/255/255, HEX #FFFFFF), and
\code{"black"} (defined as RGB 0/0/0, HEX #000000).

\code{york_1} uses the HEX color definitions.
}
\examples{
york_1
unikn::seecol(york_1, main = "Primary colors of York University") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{example_1}},
\code{\link{oxford_error}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{uni_manchester}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_print}},
\code{\link{yale_1}},
\code{\link{yeshiva}},
\code{\link{york_2}}
}
\author{
\strong{unicol}, 2023-06-24.
}
\concept{university color palettes}
\keyword{datasets}
