% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANU.R
\docType{data}
\name{anu_2}
\alias{anu_2}
\title{Font colours of Australian National University, Australia}
\format{
An object of class \code{character} of length 4.
}
\source{
\href{https://www.anu.edu.au}{ANU}'s 
\href{https://webpublishing.anu.edu.au/web-style-guide/colours}{Web Style Guide: Colours}.
}
\usage{
anu_2
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{anu_2} provides the four font colors 
of the \href{https://www.anu.edu.au}{Australian National University}, Canberra, Australia.
}
\details{
ANU's primary colour is \code{ANU gold}, defined as HEX \code{"#BE830E"}, 
but the default font colour is \code{"black"}, or HEX \code{"#000000"}.


ANU has several alternative font colours that are approved to use on ANU websites. 
See ANU's \href{https://webpublishing.anu.edu.au/web-style-guide/text}{Web Style Guide: Text}
for corresponding class names. 

Please ensure that any combination of font and background colours passes the WCAG 2.1 accessibility check. 
(see the \href{https://webpublishing.anu.edu.au/web-style-guide/accessibility}{Accessibility section} 
for approved colour combinations).
}
\examples{
anu_2
unikn::seecol(anu_2, col_bg = "grey95", 
              main = "Font colours of the Australian National University (ANU)")

}
\seealso{
\code{\link{anu_1}} for background colours of Australian National University;
\code{\link[unikn]{seecol}} for viewing and comparing color palettes;
\code{\link[unikn]{usecol}} for using color palettes;
\code{\link[unikn]{simcol}} for finding similar colors;
\code{\link[unikn]{newpal}} for defining new color palettes;
\code{\link[unikn]{grepal}} for finding named colors.

Other Australian university color palettes: 
\code{\link{anu_1}},
\code{\link{mq_1}},
\code{\link{mq_2}},
\code{\link{wollongong_1}},
\code{\link{wollongong_2}}
}
\author{
\strong{unicol}, 2023-09-04.
}
\concept{Australian university color palettes}
\keyword{datasets}
