% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Birmingham.R
\docType{data}
\name{birmingham_1}
\alias{birmingham_1}
\title{Colours of the University of Birmingham, UK}
\format{
An object of class \code{character} of length 6.
}
\source{
colour definitions are based on the
\href{https://intranet.birmingham.ac.uk/social-sciences/schools/business/documents/public/brand-templates/11687-REVISED-BBS-BRAND-GUIDELINES-FINAL-FEB-2016.pdf}{CD manual (PDF)}.
}
\usage{
birmingham_1
}
\value{
A named vector of colours (HEX/HTML codes of type character).
}
\description{
\code{birmingham_1} provides the three primary colours 
and two secondary colours 
of the \href{https://www.birmingham.ac.uk/index.aspx}{University of Birmingham}, UK.
}
\details{
The three primary colours are
\code{"Undergraduate"} (defined as HEX #c1d82f),
\code{"Postgraduate"} (defined as HEX #c1d82f),
\code{"MBA"} (defined as HEX #7D0049), and
\code{"Research/Business"} (defined as HEX #605270).

The two secondary colours are 
\code{"Secondary colour 1"} (defined as HEX #bfb6ad), and
\code{"Secondary colour 2"} (defined as HEX #757477).

\code{birmingham_1} uses the HEX colour definition.
}
\examples{
birmingham_1
unikn::seecol(birmingham_1, main = "University of Birmingham") # view colour palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing colour palettes;
\code{\link{usecol}} for using colour palettes;
\code{\link{simcol}} for finding similar colours;
\code{\link{newpal}} for defining new colour palettes;
\code{\link{grepal}} for finding named colours.

Other U.K. university color palettes: 
\code{\link{lancaster_1}},
\code{\link{lancaster_2}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{uni_manchester}}
}
\author{
\strong{unicol}, 2023-06-20.
}
\concept{U.K. university color palettes}
\keyword{datasets}
