% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boston_University.R
\docType{data}
\name{bu_1}
\alias{bu_1}
\title{Primary colors of Boston University, USA}
\format{
An object of class \code{character} of length 2.
}
\source{
The color definitions are based on BU's 
\href{https://www.bu.edu/brand/guidelines/design/colors/}{brand guidelines}.
}
\usage{
bu_1
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{bu_1} provides the two primary colors 
of the \href{https://www.bu.edu}{Boston University}, USA.
}
\details{
BU's two primary colors are 
\code{"BU red"} (defined as RGB 204 0 0, CMYK 0 100 75 4, HEX CC0000) and
\code{"BU black"} (defined as RGB 45 41 38, CMYK 63 62 59 94, HEX 2D2926).

\code{bu_1} uses the HEX color definition.
}
\examples{
bu_1
unikn::seecol(bu_1, main = "Primary colors of Boston University") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other U.S. university color palettes: 
\code{\link{CMU_1_core}},
\code{\link{CMU_2_tartan}},
\code{\link{CMU_3_campus}},
\code{\link{asu_1}},
\code{\link{asu_2}},
\code{\link{asu_3}},
\code{\link{berkeley_1}},
\code{\link{berkeley_2}},
\code{\link{brown_1}},
\code{\link{brown_2}},
\code{\link{caltech_1}},
\code{\link{caltech_2}},
\code{\link{caltech_3}},
\code{\link{chicago_1}},
\code{\link{chicago_2}},
\code{\link{columbia}},
\code{\link{cornell_1}},
\code{\link{cornell_2}},
\code{\link{cornell_3_accent}},
\code{\link{dartmouth_1}},
\code{\link{dartmouth_2}},
\code{\link{dartmouth_3}},
\code{\link{duke_1}},
\code{\link{duke_2}},
\code{\link{harvard_1}},
\code{\link{harvard_2}},
\code{\link{harvard_3}},
\code{\link{jhu_0}},
\code{\link{jhu_1}},
\code{\link{jhu_2}},
\code{\link{jhu_3_accent}},
\code{\link{jhu_4_gray}},
\code{\link{michigan_1}},
\code{\link{michigan_2}},
\code{\link{minnesotatwin_1}},
\code{\link{minnesotatwin_2}},
\code{\link{mit}},
\code{\link{monash_1}},
\code{\link{monash_2}},
\code{\link{msu}},
\code{\link{northwestern_1}},
\code{\link{northwestern_2}},
\code{\link{notredame_1}},
\code{\link{notredame_2}},
\code{\link{nyu_1}},
\code{\link{nyu_2}},
\code{\link{nyu_accent}},
\code{\link{nyu_neutral}},
\code{\link{uci_1}},
\code{\link{uci_2}},
\code{\link{uci_3}},
\code{\link{uci_4}},
\code{\link{uflorida_1}},
\code{\link{uflorida_2}},
\code{\link{uflorida_3}}
}
\author{
\strong{unicol}, 2023-07-04.
}
\concept{U.S. university color palettes}
\keyword{datasets}
