% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cornell.R
\docType{data}
\name{cornell_1}
\alias{cornell_1}
\title{Primary colors of Cornell University, USA}
\format{
An object of class \code{character} of length 3.
}
\source{
Color definitions are based on the
\href{https://brand.cornell.edu/design-center/colors/}{Color manual}.
}
\usage{
cornell_1
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{cornell_1} provides the two primary colors
of the \href{https://www.cornell.edu/}{Cornell University}, USA.
}
\details{
Primary colors are the officially recognized color palette for the university.
These colors are critical to leverage the Cornell brand and provide clear university consistency.
Logos may only be represented in 
\code{"carnelian"} (defined as HEX #B31B1B),  
\code{"dark_gray"} (defined as HEX #222222), or 
\code{"white"} (defined as HEX #FFFFFF).
}
\examples{
cornell_1
unikn::seecol(cornell_1, main = "Primary colors Cornell Uni") # view color palette

}
\seealso{
\code{\link{cornell_2}} for secondary colors of Cornell University;
\code{\link{cornell_3_accent}} for accent colors of Cornell University;
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other U.S. university color palettes: 
\code{\link{CMU_1_core}},
\code{\link{CMU_2_tartan}},
\code{\link{CMU_3_campus}},
\code{\link{asu_1}},
\code{\link{berkeley_1}},
\code{\link{berkeley_2}},
\code{\link{brown_1}},
\code{\link{brown_2}},
\code{\link{bu_1}},
\code{\link{caltech_1}},
\code{\link{caltech_2}},
\code{\link{caltech_3}},
\code{\link{chicago_1}},
\code{\link{chicago_2}},
\code{\link{columbia}},
\code{\link{cornell_2}},
\code{\link{cornell_3_accent}},
\code{\link{monash_1}},
\code{\link{monash_2}}
}
\author{
\strong{unicol}, 2023-06-14.
}
\concept{U.S. university color palettes}
\keyword{datasets}
