% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duke.R
\docType{data}
\name{duke_1}
\alias{duke_1}
\title{Primary colors of Duke University, USA}
\format{
An object of class \code{character} of length 2.
}
\source{
Color definitions are based on the
\href{https://brand.duke.edu/colors/}{Color manual}.
}
\usage{
duke_1
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{duke_1} provides the two primary colors
of the \href{https://duke.edu}{Duke University}, USA.
}
\details{
The official Duke blue is a shade of navy blue that has been in use for decades. Called
\code{"duke_navy_blue"} (defined as CMYK 100/85/5/22, RGB 1/33/105, HEX #012169) in our palette, it is sometimes referred to as “Academic Blue.”
\code{"duke_royal_blue"} (defined as CMYK 100/53/2/16, RGB 0/83/155, HEX #00539B) is the other shade of blue in the
palette and has been in use since 2009 for athletics, apparel and promotional materials.
}
\examples{
duke_1
unikn::seecol(duke_1, main = "Primary colors Duke University") # view color palette

}
\seealso{
\code{\link{duke_2}} for secondary colors of Duke University;
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other U.S. university color palettes: 
\code{\link{CMU_1_core}},
\code{\link{CMU_2_tartan}},
\code{\link{CMU_3_campus}},
\code{\link{asu_1}},
\code{\link{berkeley_1}},
\code{\link{berkeley_2}},
\code{\link{brown_1}},
\code{\link{brown_2}},
\code{\link{bu_1}},
\code{\link{caltech_1}},
\code{\link{caltech_2}},
\code{\link{caltech_3}},
\code{\link{chicago_1}},
\code{\link{chicago_2}},
\code{\link{columbia}},
\code{\link{cornell_1}},
\code{\link{cornell_2}},
\code{\link{cornell_3_accent}},
\code{\link{dartmouth_1}},
\code{\link{dartmouth_2}},
\code{\link{dartmouth_3}},
\code{\link{duke_2}},
\code{\link{harvard_1}},
\code{\link{harvard_2}},
\code{\link{harvard_3}},
\code{\link{jhu_0}},
\code{\link{jhu_1}},
\code{\link{jhu_2}},
\code{\link{jhu_3_accent}},
\code{\link{jhu_4_gray}},
\code{\link{monash_1}},
\code{\link{monash_2}},
\code{\link{uci_1}},
\code{\link{uci_2}},
\code{\link{uci_3}},
\code{\link{uci_4}},
\code{\link{uflorida_1}},
\code{\link{uflorida_2}},
\code{\link{uflorida_3}}
}
\author{
\strong{unicol}, 2023-06-16.
}
\concept{U.S. university color palettes}
\keyword{datasets}
