% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FU_Berlin.R
\docType{data}
\name{fu_3}
\alias{fu_3}
\title{Blue color gradient of the Free University (FU) Berlin, Germany}
\format{
An object of class \code{character} of length 10.
}
\source{
Color definitions are based on 
\url{https://www.fu-berlin.de/sites/corporate-design/grundlagen/farben/index.html}.
}
\usage{
fu_3
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{fu_3} provides a gradient of the blue primary color  
of the \href{https://www.fu-berlin.de/en/index.html}{Free University Berlin}, Germany.
}
\details{
See \code{\link{fu_0}} for the two primary colors and 
the \code{\link{ac}} and \code{\link{usecol}} functions for creating 
alternative color gradients.
}
\examples{
fu_3
unikn::seecol(fu_3, main = "Blue color gradient of the FU Berlin")  # view color palette
unikn::demopal(fu_3, type = 3, main = "Blue color gradient of the FU Berlin")

}
\seealso{
\code{\link{fu_0}} for primary colors of the FU Berlin;
\code{\link{fu_1}} for primary color gradient of the FU Berlin; 
\code{\link{fu_2}} for secondary colors of the FU Berlin; 
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other German university color palettes: 
\code{\link{fu_0}},
\code{\link{fu_1}},
\code{\link{fu_2}},
\code{\link{hu_1}},
\code{\link{hu_2}},
\code{\link{lmu_1}},
\code{\link{lmu_2}},
\code{\link{lmu_3}},
\code{\link{rptu}},
\code{\link{uni_bonn_1}},
\code{\link{uni_bonn_2}},
\code{\link{uni_freiburg_0}},
\code{\link{uni_freiburg_1}},
\code{\link{uni_freiburg_2}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_hamburg_1}},
\code{\link{uni_hamburg_2}},
\code{\link{uni_heidelberg}},
\code{\link{uni_jena_1}},
\code{\link{uni_jena_2}},
\code{\link{uni_kassel}},
\code{\link{uni_kiel_1}},
\code{\link{uni_kiel_2}},
\code{\link{uni_koeln_1}},
\code{\link{uni_koeln_2}},
\code{\link{uni_konstanz_1}},
\code{\link{uni_konstanz_2}},
\code{\link{uni_mannheim_1}},
\code{\link{uni_mannheim_2}},
\code{\link{uni_potsdam}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_2}},
\code{\link{uni_regensburg_3}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_print}},
\code{\link{uni_ulm_1}},
\code{\link{uni_ulm_2}}
}
\author{
\strong{unicol}, 2022-12-03.
}
\concept{German university color palettes}
\keyword{datasets}
