% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lancaster.R
\docType{data}
\name{lancaster_2}
\alias{lancaster_2}
\title{Secondary colours of the Lancaster University}
\format{
An object of class \code{character} of length 19.
}
\source{
colour definitions are based on the
\href{https://www.lancaster.ac.uk/media/lancaster-university/content-assets/documents/current-staff/Primary/LUPrimaryGuidelines-Staff-Nov-2018.pdf}{Lancaster's Primary guide}.
}
\usage{
lancaster_2
}
\value{
A named vector of colours (HEX/HTML codes of type character).
}
\description{
This palette can be used in combination with the primary colour palette to introduce subtlety and variety to 
certain applications. They too must be used carefully to compliment and enrich the Lancaster University identity.
}
\details{
\code{lancaster_2} provides the twelve secondary colours to complement the signature Primary colours
of the \href{https://www.lancaster.ac.uk}{Lancaster University}, UK.

The 19 secondary colours are 
\code{"Pantone® 5555"} (defined as Hex:557869),
\code{"Pantone® 327"} (defined as Hex: 008375),
\code{"Pantone® 3258"} (defined as Hex: 48B6ADB),
\code{"Pantone® 7707"} (defined as Hex: 006382),
\code{"Pantone® 2377"} (defined as Hex: 324147),

\code{"Pantone® 7493"} (defined as Hex: 869978),
\code{"Pantone® 346"} (defined as Hex: 75BF9A),
\code{"Pantone® 5503"} (defined as Hex: 81B8BB),
\code{"Pantone® 643"} (defined as Hex: C4DAE5),
\code{"Pantone® 550"} (defined as Hex: 7faabe),

\code{"Pantone® 459"} (defined as Hex: e3cb8b),
\code{"Pantone® 7661"} (defined as Hex: 64606c),
\code{"Pantone® 178"} (defined as Hex: FF7372),
\code{"Pantone® 689"} (defined as Hex: 8A3E65),
\code{"Pantone® 7535"} (defined as Hex: bab6a2),

\code{"Pantone® 7411"} (defined as Hex: e1ab6c),
\code{"Pantone® 7607"} (defined as Hex: c26763),
\code{"Pantone® 1625"} (defined as Hex: F9A98E), and
\code{"Pantone® 7654"} (defined as Hex: A4769A).


\code{lancaster_2} uses the HEX colour definition.
}
\examples{
lancaster_uni_accent
unikn::seecol(lancaster_2, main = "Secondary colours of Lancaster University") 
# view colour palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing colour palettes;
\code{\link{usecol}} for using colour palettes;
\code{\link{simcol}} for finding similar colours;
\code{\link{newpal}} for defining new colour palettes;
\code{\link{grepal}} for finding named colours.

Other university colour palettes: 
\code{\link{birmingham_1}},
\code{\link{bu_1}},
\code{\link{dundee_background}},
\code{\link{dundee_block}},
\code{\link{dundee_core}},
\code{\link{dundee_highlight}},
\code{\link{lancaster_1}},
\code{\link{mcgill_brights}},
\code{\link{mcgill_darks}},
\code{\link{mcgill_grey}},
\code{\link{mcgill_muted}},
\code{\link{mcgill_pastels}},
\code{\link{mcgill_red}},
\code{\link{mcmaster_brighterworld}},
\code{\link{mcmaster_heritage}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{queens_primary}},
\code{\link{queens_secondary}},
\code{\link{sfu_brand}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_primary}},
\code{\link{ucalgary_secondary}},
\code{\link{ucalgary_warmgreys}},
\code{\link{uni_heidelberg_1}},
\code{\link{uvic_complementary}},
\code{\link{uvic_core}}
}
\author{
\strong{unicol}, 2023-06-27.
}
\concept{university colour palettes}
\keyword{datasets}
