% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NotreDame.R
\docType{data}
\name{notredame_2}
\alias{notredame_2}
\title{Secondary colors of the University of Notre Dame, USA}
\format{
An object of class \code{character} of length 11.
}
\source{
Color definitions are based on 
\href{https://onmessage.nd.edu/university-branding/colors/}{ND's website}.
}
\usage{
notredame_2
}
\value{
A names vector of colors (HEX/HTML codes of type character).
}
\description{
\code{notredame_2} provides the 11 secondary colors
of the \href{https://www.nd.edu}{University of Notre Dame}, USA.
}
\details{
The 11 secondary colors are 
\code{"brand_blue_light"} (defined as HEX #143865),
\code{"brand_blue_bright"} (defined as HEX #1c4f8f),
\code{"brand_gold_dark"} (defined as HEX #8c7535),
\code{"brand_gold_light"} (defined as HEX #d39f10), 
\code{"sky_blue"} (defined as HEX #e1e8f2), 
\code{"sky_blue_dark"} (defined as HEX #c1cddd), 
\code{"sky_blue_light"} (defined as HEX #edf2f9), 
\code{"warm_white"} (defined as HEX #f8f4ec), 
\code{"warm_white_dark"} (defined as HEX #efe9d9), 
\code{"brand_green"} (defined as HEX #0a843d), and 
\code{"brand_green_light"} (defined as HEX #b3dac5).

\code{notredame_2} uses the HEX color definitions.
}
\examples{
notredame_2
unikn::seecol(notredame_2, main = "University of Notre Dame") # view color palette

}
\seealso{
\code{\link{notredame_1}} for the primary colors of Notre Dame;
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other U.S. university color palettes: 
\code{\link{CMU_1_core}},
\code{\link{CMU_2_tartan}},
\code{\link{CMU_3_campus}},
\code{\link{UCLA_1}},
\code{\link{UCLA_2}},
\code{\link{UCLA_3}},
\code{\link{UCSD_1_core}},
\code{\link{UCSD_2_accent}},
\code{\link{UCSD_3_neutral}},
\code{\link{asu_1}},
\code{\link{asu_2}},
\code{\link{asu_3}},
\code{\link{berkeley_1}},
\code{\link{berkeley_2}},
\code{\link{brown_1}},
\code{\link{brown_2}},
\code{\link{bu}},
\code{\link{caltech_1}},
\code{\link{caltech_2}},
\code{\link{caltech_3}},
\code{\link{chicago_1}},
\code{\link{chicago_2}},
\code{\link{columbia}},
\code{\link{cornell_1}},
\code{\link{cornell_2}},
\code{\link{cornell_3_accent}},
\code{\link{dartmouth_1}},
\code{\link{dartmouth_2}},
\code{\link{dartmouth_3}},
\code{\link{duke_1}},
\code{\link{duke_2}},
\code{\link{harvard_1}},
\code{\link{harvard_2}},
\code{\link{harvard_3}},
\code{\link{jhu_0}},
\code{\link{jhu_1}},
\code{\link{jhu_2}},
\code{\link{jhu_3_accent}},
\code{\link{jhu_4_gray}},
\code{\link{manchester_uni_1}},
\code{\link{manchester_uni_accent}},
\code{\link{michigan_1}},
\code{\link{michigan_2}},
\code{\link{minnesotatwin_1}},
\code{\link{minnesotatwin_2}},
\code{\link{mit}},
\code{\link{monash_1}},
\code{\link{monash_2}},
\code{\link{msu}},
\code{\link{northwestern_1}},
\code{\link{northwestern_2}},
\code{\link{notredame_1}},
\code{\link{nyu_1}},
\code{\link{nyu_2}},
\code{\link{nyu_accent}},
\code{\link{nyu_neutral}},
\code{\link{ohio_uni_1}},
\code{\link{ohio_uni_2}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{princeton_0}},
\code{\link{princeton_1}},
\code{\link{princeton_2}},
\code{\link{rpi_1}},
\code{\link{rpi_2}},
\code{\link{rpi_3}},
\code{\link{stanford_1}},
\code{\link{stanford_2_accent}},
\code{\link{stanford_3_web}},
\code{\link{uci_1}},
\code{\link{uci_2}},
\code{\link{uci_3}},
\code{\link{uci_4}},
\code{\link{uflorida_1}},
\code{\link{uflorida_2}},
\code{\link{uflorida_3}},
\code{\link{umass_1}},
\code{\link{umass_2}},
\code{\link{umass_neutrals}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{willamette}},
\code{\link{yale}},
\code{\link{yeshiva}}
}
\author{
\strong{unicol}, 2320-07-16.
}
\concept{U.S. university color palettes}
\keyword{datasets}
