% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oxford.R
\docType{data}
\name{oxford_shades}
\alias{oxford_shades}
\title{Colours of the University of Oxford}
\format{
An object of class \code{character} of length 7.
}
\source{
Colour definitions are based on the
\href{https://www.ox.ac.uk/public-affairs/style-guide/digital-style-guide}{Digital style guide}.
}
\usage{
oxford_shades
}
\value{
A named vector of colours (HEX/HTML codes of type character).
}
\description{
\code{oxford_shades} provides the shade colours 
of the \href{https://www.ox.ac.uk/}{University of Oxford}, England.
}
\details{
The colours are
\code{"black"} (defined as HEX #000000),
\code{"off_black"} (defined as HEX #333333),
\code{"mid_grey"} (defined as HEX #7a736e),
\code{"dark_grey"} (defined as HEX #a6a6a6),
\code{"light_grey"} (defined as HEX #e0ded9),
\code{"light_grey"} (defined as HEX #f9f8f5) and
\code{"white"} (defined as HEX #ffffff).
}
\examples{
oxford_shades
unikn::seecol(oxford_shades, main = "Shade colours") # view colour palette

}
\seealso{
\code{\link{oxford_brand}} for the brand colours of the University of Oxford;
\code{\link{oxford_general}} for the general colours of the University of Oxford;
\code{\link{oxford_link}} for the link colours of the University of Oxford;
\code{\link{oxford_blog}} for the blog colours of the University of Oxford;
\code{\link{oxford_graduate}} for the graduate colours of the University of Oxford;
\code{\link{oxford_error}} for the error colours of the University of Oxford;
\code{\link{oxford_socialmedia}} for the social media colours of the University of Oxford;
\code{\link{seecol}} for viewing and comparing colour palettes;
\code{\link{usecol}} for using colour palettes;
\code{\link{simcol}} for finding similar colours;
\code{\link{newpal}} for defining new colour palettes;
\code{\link{grepal}} for finding named colours.

Other university colour palettes: 
\code{\link{birmingham_1}},
\code{\link{mcgill_brights}},
\code{\link{mcgill_darks}},
\code{\link{mcgill_grey}},
\code{\link{mcgill_muted}},
\code{\link{mcgill_pastels}},
\code{\link{mcgill_red}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_socialmedia}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_primary}},
\code{\link{ucalgary_secondary}},
\code{\link{ucalgary_warmgreys}}
}
\author{
\strong{unicol}, 2023-06-19.
}
\concept{university colour palettes}
\keyword{datasets}
