% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stanford.R
\docType{data}
\name{stanford_1}
\alias{stanford_1}
\title{Primary colors of Stanford University}
\format{
An object of class \code{character} of length 15.
}
\source{
Color definitions are based on the
\href{https://identity.stanford.edu/design-elements/color/}{Color manual}.
}
\usage{
stanford_1
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{stanford_1} provides the 15 primary colors
of the \href{https://www.stanford.edu/}{Stanford University}, USA.
}
\details{
Primary color palette consists of Cardinal red, white, black and cool grey.
The layouts lean heavily on these colors, which are
\code{"cardinal_red"} (defined as HEX #8C1515),
\code{"cardinal_red_light"} (defined as HEX #B83A4B),
\code{"cardinal_red_dark"} (defined as HEX #820000),
\code{"white"} (defined as HEX #FFFFFF),
\code{"black"} (defined as HEX #2E2D29),
\code{"black_90"} (defined as HEX #43423E),
\code{"black_80"} (defined as HEX #585754),
\code{"black_70"} (defined as HEX #6D6C69),
\code{"black_60"} (defined as HEX #767674),
\code{"black_50"} (defined as HEX #979694),
\code{"black_40"} (defined as HEX #ABABA9),
\code{"black_30"} (defined as HEX #C0C0BF),
\code{"black_20"} (defined as HEX #D5D5D4),
\code{"black_10"} (defined as HEX #EAEAEA) and 
\code{"cool_grey"} (defined as HEX #53565A).
}
\examples{
stanford_1
unikn::seecol(stanford_1, main = "Primary colors Stanford Uni") # view color palette

}
\seealso{
\code{\link{stanford_2_accent}} for secondary, accent colors of Stanford University
\code{\link{stanford_3_web}} for web colors of Stanford University
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{berkeley_1}},
\code{\link{berkeley_2}},
\code{\link{caltech_1}},
\code{\link{caltech_2}},
\code{\link{caltech_3}},
\code{\link{chicago_1}},
\code{\link{chicago_2}},
\code{\link{columbia}},
\code{\link{cornell_1}},
\code{\link{cornell_2}},
\code{\link{cornell_3_accent}},
\code{\link{eth_pal_1}},
\code{\link{eth_pal_2}},
\code{\link{eth_pal_3}},
\code{\link{fu_pal_0}},
\code{\link{fu_pal_1}},
\code{\link{fu_pal_2}},
\code{\link{fu_pal_3}},
\code{\link{harvard_1}},
\code{\link{harvard_2}},
\code{\link{harvard_3_accent}},
\code{\link{hu_pal_1}},
\code{\link{hu_pal_2}},
\code{\link{jhu_0}},
\code{\link{jhu_1}},
\code{\link{jhu_2}},
\code{\link{jhu_3_accent}},
\code{\link{jhu_4_gray}},
\code{\link{lmu_pal_1}},
\code{\link{lmu_pal_2}},
\code{\link{lmu_pal_3}},
\code{\link{michigan_1}},
\code{\link{michigan_2}},
\code{\link{mit_pal}},
\code{\link{mpg_pal}},
\code{\link{northwestern_1}},
\code{\link{northwestern_2}},
\code{\link{nyu_1}},
\code{\link{nyu_2}},
\code{\link{nyu_accent}},
\code{\link{nyu_neutral}},
\code{\link{princeton_0}},
\code{\link{princeton_1}},
\code{\link{princeton_2}},
\code{\link{rpi_pal_1}},
\code{\link{rpi_pal_2}},
\code{\link{rpi_pal_3}},
\code{\link{rptu_pal}},
\code{\link{stanford_2_accent}},
\code{\link{stanford_3_web}},
\code{\link{uni_bonn_1}},
\code{\link{uni_bonn_2}},
\code{\link{uni_freiburg_0}},
\code{\link{uni_freiburg_1}},
\code{\link{uni_freiburg_2}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_hamburg_1}},
\code{\link{uni_hamburg_2}},
\code{\link{uni_jena_1}},
\code{\link{uni_jena_2}},
\code{\link{uni_kiel_1}},
\code{\link{uni_kiel_2}},
\code{\link{uni_koeln_1}},
\code{\link{uni_koeln_2}},
\code{\link{uni_konstanz_1}},
\code{\link{uni_konstanz_2}},
\code{\link{uni_mannheim_1}},
\code{\link{uni_mannheim_2}},
\code{\link{uni_pal_1}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_2}},
\code{\link{uni_regensburg_3}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_2_print}},
\code{\link{uni_ulm_1}},
\code{\link{uni_ulm_2}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{yale_1}}
}
\author{
\strong{unicol}, 2023-06-14.
}
\concept{university color palettes}
\keyword{datasets}
