% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stanford.R
\docType{data}
\name{stanford_3_web}
\alias{stanford_3_web}
\title{Web colors of Stanford University}
\format{
An object of class \code{character} of length 9.
}
\source{
Color definitions are based on the
\href{https://identity.stanford.edu/design-elements/color/}{Color manual}.
}
\usage{
stanford_3_web
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{stanford_3_web} provides the nine web colors
of the \href{https://www.stanford.edu/}{Stanford University}, USA.
}
\details{
\code{"digital_red"} is not a replacement for Cardinal red—it can be used as an accent color on digital communications to add energy and emphasis.
\code{"digital_red_light"} can be used for interactive elements over dark backgrounds for contrast. 
\code{"digital_red_dark"} can be used for smaller text over light backgrounds.
Use \code{"digital_blue"} for links over light backgrounds, 
\code{"digital_blue_dark"} for link hover/focus states, and 
\code{"digital_blue_light"} for links over dark backgrounds. 
Do not use these shades of blue in contexts other than links.
Use \code{"digital_green"} for things such as form validation, 
\code{"digital_green_dark"} for link hover/focus states, and 
\code{"digital_green_light"} for links over dark backgrounds.

These nine web colors are
\code{"digital_red"} (defined as HEX #B1040E),
\code{"digital_red_light"} (defined as HEX #E50808),
\code{"digital_red_dark"} (defined as HEX #820000),
\code{"digital_blue"} (defined as HEX #006CB8),
\code{"digital_blue_light"} (defined as HEX #6FC3FF),
\code{"digital_blue_dark"} (defined as HEX #00548f),
\code{"digital_green"} (defined as HEX #008566),
\code{"digital_green_light"} (defined as HEX #1AECBA) and
\code{"digital_green_dark"} (defined as HEX #006F54).
}
\examples{
stanford_3_web
unikn::seecol(stanford_3_web, main = "Web colors Stanford University") # view color palette

}
\seealso{
\code{\link{stanford_1}} for primary colors of Stanford University;
\code{\link{stanford_2_accent}} for accent colors of Stanford University;
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{UCLA_1}},
\code{\link{UCLA_2}},
\code{\link{UCLA_3}},
\code{\link{UCSD_1_core}},
\code{\link{UCSD_2_accent}},
\code{\link{UCSD_3_neutral}},
\code{\link{example_1}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{queens_1}},
\code{\link{queens_2}},
\code{\link{sfu_brand}},
\code{\link{stanford_1}},
\code{\link{stanford_2_accent}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_primary}},
\code{\link{ucalgary_secondary}},
\code{\link{ucalgary_warmgreys}},
\code{\link{umass_2}},
\code{\link{umass_brand}},
\code{\link{umass_neutrals}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_heidelberg_1}},
\code{\link{uni_manchester_1}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_2_print}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{uwaterloo_arts}},
\code{\link{uwaterloo_engineering}},
\code{\link{uwaterloo_environment}},
\code{\link{uwaterloo_health}},
\code{\link{uwaterloo_main}},
\code{\link{uwaterloo_math}},
\code{\link{uwaterloo_science}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{waikato}},
\code{\link{western_uni_1}},
\code{\link{yale_1}},
\code{\link{yeshiva}},
\code{\link{york_1}},
\code{\link{york_2}}
}
\author{
\strong{unicol}, 2023-06-14.
}
\concept{university color palettes}
\keyword{datasets}
