% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calgary.R
\docType{data}
\name{ucalgary_primary}
\alias{ucalgary_primary}
\title{Colours of the University of Calgary}
\format{
An object of class \code{character} of length 2.
}
\source{
Colour definitions are based on the
\href{https://www.ucalgary.ca/sites/default/files/teams/12/3.0-eyes-high-dec2018.pdf}{Visual Identity Standards}.
}
\usage{
ucalgary_primary
}
\value{
A named vector of colours (HEX/HTML codes of type character).
}
\description{
\code{ucalgary_primary} provides the two primary colours 
of the \href{https://www.ucalgary.ca/}{University of Calgary}, Canada.
}
\details{
The two primary colours are
\code{"red"} (defined as HEX #cf0722) and
\code{"gold"} (defined as HEX #ffcd00).
}
\examples{
ucalgary_primary
unikn::seecol(ucalgary_primary, main = "Primary colours") # view colour palette

}
\seealso{
\code{\link{ucalgary_secondary}} for the secondary colours of the University of Calgary;
\code{\link{ucalgary_warmgreys}} for the warm grey colours of the University of Calgary;
\code{\link{ucalgary_accent}} for the accent colours of the University of Calgary;
\code{\link{seecol}} for viewing and comparing colour palettes;
\code{\link{usecol}} for using colour palettes;
\code{\link{simcol}} for finding similar colours;
\code{\link{newpal}} for defining new colour palettes;
\code{\link{grepal}} for finding named colours.

Other university colour palettes: 
\code{\link{birmingham_1}},
\code{\link{dundee_background}},
\code{\link{dundee_block}},
\code{\link{dundee_core}},
\code{\link{dundee_highlight}},
\code{\link{mcgill_brights}},
\code{\link{mcgill_darks}},
\code{\link{mcgill_grey}},
\code{\link{mcgill_muted}},
\code{\link{mcgill_pastels}},
\code{\link{mcgill_red}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_secondary}},
\code{\link{ucalgary_warmgreys}}
}
\author{
\strong{unicol}, 2023-06-20.
}
\concept{university colour palettes}
\keyword{datasets}
