% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calgary.R, R/UCalgary.R
\docType{data}
\name{ucalgary_warmgreys}
\alias{ucalgary_warmgreys}
\title{Colours of the University of Calgary}
\format{
An object of class \code{character} of length 2.

An object of class \code{character} of length 2.
}
\source{
Colour definitions are based on the
\href{https://www.ucalgary.ca/sites/default/files/teams/12/3.0-eyes-high-dec2018.pdf}{Visual Identity Standards}.

Colour definitions are based on the
\href{https://www.ucalgary.ca/sites/default/files/teams/12/3.0-eyes-high-dec2018.pdf}{Visual Identity Standards}.
}
\usage{
ucalgary_warmgreys

ucalgary_warmgreys
}
\value{
A named vector of colours (HEX/HTML codes of type character).

A named vector of colours (HEX/HTML codes of type character).
}
\description{
\code{ucalgary_warmgreys} provides the warm greys 
of the \href{https://www.ucalgary.ca/}{University of Calgary}, Canada.

\code{ucalgary_warmgreys} provides the warm greys 
of \href{https://www.ucalgary.ca/}{University of Calgary}, Canada.
}
\details{
The colours are
\code{"light_grey"} (defined as HEX #c4bfb6) and
\code{"dark_grey"} (defined as HEX #8d827a).

The colours are
\code{"light_grey"} (defined as HEX #c4bfb6) and
\code{"dark_grey"} (defined as HEX #8d827a).
}
\examples{
ucalgary_secondary
unikn::seecol(ucalgary_warmgreys, main = "Warm greys") # view colour palette

ucalgary_secondary
unikn::seecol(ucalgary_warmgreys, main = "Warm greys UCalgary") # view colour palette

}
\seealso{
\code{\link{ucalgary_primary}} for the primary colours of the University of Calgary;
\code{\link{ucalgary_secondary}} for the secondary colours of the University of Calgary;
\code{\link{ucalgary_accent}} for the accent colours of the University of Calgary;
\code{\link{seecol}} for viewing and comparing colour palettes;
\code{\link{usecol}} for using colour palettes;
\code{\link{simcol}} for finding similar colours;
\code{\link{newpal}} for defining new colour palettes;
\code{\link{grepal}} for finding named colours.

\code{\link{ucalgary_primary}} for the primary colours of University of Calgary;
\code{\link{ucalgary_secondary}} for the secondary colours of University of Calgary;
\code{\link{ucalgary_accent}} for the accent colours of University of Calgary;
\code{\link{seecol}} for viewing and comparing colour palettes;
\code{\link{usecol}} for using colour palettes;
\code{\link{simcol}} for finding similar colours;
\code{\link{newpal}} for defining new colour palettes;
\code{\link{grepal}} for finding named colours.

Other university colour palettes: 
\code{\link{birmingham_1}},
\code{\link{bu_1}},
\code{\link{dundee_background}},
\code{\link{dundee_block}},
\code{\link{dundee_core}},
\code{\link{dundee_highlight}},
\code{\link{lancaster_1}},
\code{\link{lancaster_2}},
\code{\link{mcgill_brights}},
\code{\link{mcgill_darks}},
\code{\link{mcgill_grey}},
\code{\link{mcgill_muted}},
\code{\link{mcgill_pastels}},
\code{\link{mcgill_red}},
\code{\link{mcmaster_brighterworld}},
\code{\link{mcmaster_heritage}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{queens_primary}},
\code{\link{queens_secondary}},
\code{\link{sfu_brand}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_primary}},
\code{\link{ucalgary_secondary}},
\code{\link{uni_heidelberg_1}},
\code{\link{uvic_complementary}},
\code{\link{uvic_core}}

Other university colour palettes: 
\code{\link{birmingham_1}},
\code{\link{bu_1}},
\code{\link{dundee_background}},
\code{\link{dundee_block}},
\code{\link{dundee_core}},
\code{\link{dundee_highlight}},
\code{\link{lancaster_1}},
\code{\link{lancaster_2}},
\code{\link{mcgill_brights}},
\code{\link{mcgill_darks}},
\code{\link{mcgill_grey}},
\code{\link{mcgill_muted}},
\code{\link{mcgill_pastels}},
\code{\link{mcgill_red}},
\code{\link{mcmaster_brighterworld}},
\code{\link{mcmaster_heritage}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{queens_primary}},
\code{\link{queens_secondary}},
\code{\link{sfu_brand}},
\code{\link{ucalgary_accent}},
\code{\link{ucalgary_primary}},
\code{\link{ucalgary_secondary}},
\code{\link{uni_heidelberg_1}},
\code{\link{uvic_complementary}},
\code{\link{uvic_core}}
}
\author{
\strong{unicol}, 2023-06-20.
}
\concept{university colour palettes}
\keyword{datasets}
