% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCLA.R
\docType{data}
\name{UCLA_2}
\alias{UCLA_2}
\title{Secondary colors of UCLA, University of California, Los Angeles, USA}
\format{
An object of class \code{character} of length 6.
}
\source{
Color definitions are based on \href{https://www.ucla.edu}{UCLA}'s 
\href{https://brand.ucla.edu/identity/colors}{Color manual}.
}
\usage{
UCLA_2
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{UCLA_2} provides six secondary colors
of the \href{https://www.ucla.edu}{University of California, Los Angeles (UCLA)}, USA.
}
\details{
Secondary colors are
\code{"darkest blue"} (defined as  CMYK 100/48/12/58, RGB 0/59/92, HEX #003B5C),
\code{"darker blue"} (defined as CMYK 100/45/0/45, RGB 0/85/135, HEX #005587),
\code{"lighter blue"} (defined as CMYK 45/14/9/9, RGB 139/184/232, HEX #8BB8E8),
\code{"lightest blue"} (defined as CMYK 20/6/0/0, RGB 218/235/254, HEX #DAEBFE),
\code{"darkest gold"} (defined as CMYK 0/31/98/0, RGB 255/184/28, HEX #FFB81C), and
\code{"darker gold"} (defined as CMYK 0/19/89/0, RGB 255/199/44, HEX #FFC72C).

\code{UCLA_2} uses the HEX color definitions.


A secondary palette has been developed to respect and complement the tradition of blue and gold, 
while adding an additional level of brightness or darkness to the palette.
}
\examples{
UCLA_2
unikn::seecol(UCLA_2, main = "The secondary colors UCLA") # view color palette

}
\seealso{
\code{\link{UCLA_1}} for primary colors of UCLA, University of California; 
\code{\link{UCLA_3}} for tertiary colors of UCLA, University of California; 
\code{\link{UCLA_4}} for blue gradient colors of UCLA, University of California; 
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other U.S. university color palettes: 
\code{\link{CMU_1_core}},
\code{\link{CMU_2_tartan}},
\code{\link{CMU_3_campus}},
\code{\link{UCLA_1}},
\code{\link{UCLA_3}},
\code{\link{UCLA_4}},
\code{\link{UCSD_1}},
\code{\link{UCSD_2}},
\code{\link{UCSD_3}},
\code{\link{asu_1}},
\code{\link{asu_2}},
\code{\link{asu_3}},
\code{\link{berkeley_1}},
\code{\link{berkeley_2}},
\code{\link{brown_1}},
\code{\link{brown_2}},
\code{\link{bu}},
\code{\link{caltech_1}},
\code{\link{caltech_2}},
\code{\link{caltech_3}},
\code{\link{columbia}},
\code{\link{cornell_1}},
\code{\link{cornell_2}},
\code{\link{cornell_3_accent}},
\code{\link{dartmouth_1}},
\code{\link{dartmouth_2}},
\code{\link{dartmouth_3}},
\code{\link{duke_1}},
\code{\link{duke_2}},
\code{\link{harvard_1}},
\code{\link{harvard_2}},
\code{\link{harvard_3}},
\code{\link{jhu_0}},
\code{\link{jhu_1}},
\code{\link{jhu_2}},
\code{\link{jhu_3_accent}},
\code{\link{jhu_4_gray}},
\code{\link{manchester_uni_1}},
\code{\link{manchester_uni_2}},
\code{\link{michigan_1}},
\code{\link{michigan_2}},
\code{\link{minnesotatwin_1}},
\code{\link{minnesotatwin_2}},
\code{\link{mit}},
\code{\link{monash_1}},
\code{\link{monash_2}},
\code{\link{msu}},
\code{\link{northwestern_1}},
\code{\link{northwestern_2}},
\code{\link{notredame_1}},
\code{\link{notredame_2}},
\code{\link{nyu_1}},
\code{\link{nyu_2}},
\code{\link{nyu_accent}},
\code{\link{nyu_neutral}},
\code{\link{ohio_uni_1}},
\code{\link{ohio_uni_2}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{princeton_0}},
\code{\link{princeton_1}},
\code{\link{princeton_2}},
\code{\link{rpi_1}},
\code{\link{rpi_2}},
\code{\link{rpi_3}},
\code{\link{stanford_1}},
\code{\link{stanford_2_accent}},
\code{\link{stanford_3_web}},
\code{\link{uchicago_1}},
\code{\link{uchicago_2}},
\code{\link{uci_1}},
\code{\link{uci_2}},
\code{\link{uci_3}},
\code{\link{uci_4}},
\code{\link{uflorida_1}},
\code{\link{uflorida_2}},
\code{\link{uflorida_3}},
\code{\link{umass_1}},
\code{\link{umass_2}},
\code{\link{umass_neutrals}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{willamette}},
\code{\link{wm_1}},
\code{\link{wm_2}},
\code{\link{wm_3}},
\code{\link{yale}},
\code{\link{yeshiva}}
}
\author{
\strong{unicol}, 2023-06-16.
}
\concept{U.S. university color palettes}
\keyword{datasets}
