% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Florida.R
\docType{data}
\name{uflorida_2}
\alias{uflorida_2}
\title{Secondary colors of the University of Florida, USA}
\format{
An object of class \code{character} of length 5.
}
\source{
Color definitions are based on the
\href{https://brandcenter.ufl.edu/colors/}{UFL brand center}.
}
\usage{
uflorida_2
}
\value{
A names vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uflorida_2} provides the 5 secondary colors
of the \href{https://www.ufl.edu/}{University of Florida}, USA.
}
\details{
The 5 secondary colors are
\code{"ufl_bottlebrush"} (defined as HEX #D32737),
\code{"ufl_alachua"} (defined as HEX #F2A900),
\code{"ufl_gator"} (defined as HEX #2284C),
\code{"ufl_dark_blue"} (defined as HEX #002657) and
\code{"ufl_perennial"} (defined as HEX #6A2A60).

\code{uflorida_2} uses the HEX color definitions.
}
\examples{
uflorida_2
unikn::seecol(uflorida_2, main = "University of Florida 2") # view color palette

}
\seealso{
\code{\link{uflorida_1}} for primary colors of the University of Florida;
\code{\link{uflorida_3}} for neutral colors of the University of Florida; 
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other U.S. university color palettes: 
\code{\link{CMU_1_core}},
\code{\link{CMU_2_tartan}},
\code{\link{CMU_3_campus}},
\code{\link{asu_1}},
\code{\link{asu_2}},
\code{\link{asu_3}},
\code{\link{berkeley_1}},
\code{\link{berkeley_2}},
\code{\link{brown_1}},
\code{\link{brown_2}},
\code{\link{bu_1}},
\code{\link{caltech_1}},
\code{\link{caltech_2}},
\code{\link{caltech_3}},
\code{\link{chicago_1}},
\code{\link{chicago_2}},
\code{\link{columbia}},
\code{\link{cornell_1}},
\code{\link{cornell_2}},
\code{\link{cornell_3_accent}},
\code{\link{dartmouth_1}},
\code{\link{dartmouth_2}},
\code{\link{dartmouth_3}},
\code{\link{duke_1}},
\code{\link{duke_2}},
\code{\link{harvard_1}},
\code{\link{harvard_2}},
\code{\link{harvard_3}},
\code{\link{jhu_0}},
\code{\link{jhu_1}},
\code{\link{jhu_2}},
\code{\link{jhu_3_accent}},
\code{\link{jhu_4_gray}},
\code{\link{michigan_1}},
\code{\link{michigan_2}},
\code{\link{minnesotatwin_1}},
\code{\link{minnesotatwin_2}},
\code{\link{mit}},
\code{\link{monash_1}},
\code{\link{monash_2}},
\code{\link{msu}},
\code{\link{uci_1}},
\code{\link{uci_2}},
\code{\link{uci_3}},
\code{\link{uci_4}},
\code{\link{uflorida_1}},
\code{\link{uflorida_3}}
}
\author{
\strong{unicol}, 2320-07-07.
}
\concept{U.S. university color palettes}
\keyword{datasets}
