% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_pals.R
\docType{data}
\name{uni_regensburg_2}
\alias{uni_regensburg_2}
\title{Departmental colors of the University of Regensburg (for areas)}
\format{
An object of class \code{character} of length 12.
}
\source{
Color definitions are based on 
\url{https://www.uni-regensburg.de/verwaltung/corporate-design/index.html}.
}
\usage{
uni_regensburg_2
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_regensburg_2} provides the departmental (faculty) colors 
for coloring print or web areas 
at the \href{https://www.uni-regensburg.de/}{University of Regensburg}, Germany.
}
\details{
Named colors of \code{uni_regensburg_2} correspond to the following UR faculties:

\enumerate{
\item \code{urangelb}: Rechtswissenschaft
\item \code{laerchennadelgruen}: Wirtschaftswissenschaften
\item \code{vatikangelb}: Katholische Theologie
\item \code{orangerot}: Philosophie, Kunst-, Geschichts- und Gesellschaftswissenschaften
\item \code{heucherarot}: Humanwissenschaften
\item \code{heidenelkenrot}: Sprach-, Literatur- und Kulturwissenschaften
\item \code{blattgruen}: Biologie und Vorklinische Medizin
\item \code{tuerkisgruen}: Mathematik
\item \code{eisvogelblau}: Physik
\item \code{spektralblau}: Chemie und Pharmazie
\item \code{capriblau}: Medizin
\item \code{glutrot}: Informatik und Data Science
}

When coloring large areas, named colors can be used in shades of 70%.
}
\examples{
uni_regensburg_2
unikn::seecol(uni_regensburg_2, main = "Departmental colors at the University of Regensburg")
unikn::demopal(uni_regensburg_2, type = 3, main = "Department colors of Regensburg University")

}
\seealso{
\code{\link{uni_regensburg_1}} and \code{\link{uni_regensburg_3}} for alternative colors of the University of Regensburg;  
\code{\link{seecol}} for viewing and comparing color palettes; 
\code{\link{usecol}} for using color palettes; 
\code{\link{simcol}} for finding similar colors; 
\code{\link{newpal}} for defining new color palettes; 
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{caltech_pal_1}},
\code{\link{caltech_pal_2}},
\code{\link{caltech_pal_3}},
\code{\link{eth_pal_1}},
\code{\link{eth_pal_2}},
\code{\link{eth_pal_3}},
\code{\link{fu_pal_0}},
\code{\link{fu_pal_1}},
\code{\link{fu_pal_2}},
\code{\link{fu_pal_3}},
\code{\link{hu_pal_1}},
\code{\link{hu_pal_2}},
\code{\link{lmu_pal_1}},
\code{\link{lmu_pal_2}},
\code{\link{lmu_pal_3}},
\code{\link{mit_1}},
\code{\link{mpg_pal}},
\code{\link{rpi_pal_1}},
\code{\link{rpi_pal_2}},
\code{\link{rpi_pal_3}},
\code{\link{rptu_pal}},
\code{\link{uni_berkeley_primary_1}},
\code{\link{uni_berkeley_secondary_2}},
\code{\link{uni_bonn_1}},
\code{\link{uni_bonn_2}},
\code{\link{uni_chicago_primary_1}},
\code{\link{uni_chicago_secondary_2}},
\code{\link{uni_columbia}},
\code{\link{uni_cornell_accent_web_3}},
\code{\link{uni_cornell_primary_web_1}},
\code{\link{uni_cornell_secondary_web_2}},
\code{\link{uni_freiburg_0}},
\code{\link{uni_freiburg_1}},
\code{\link{uni_freiburg_2}},
\code{\link{uni_freiburg_blue}},
\code{\link{uni_freiburg_br}},
\code{\link{uni_freiburg_grey}},
\code{\link{uni_freiburg_info}},
\code{\link{uni_goettingen_1}},
\code{\link{uni_goettingen_2}},
\code{\link{uni_goettingen_3}},
\code{\link{uni_hamburg_1}},
\code{\link{uni_hamburg_2}},
\code{\link{uni_harvard_accent_3}},
\code{\link{uni_harvard_primary_1}},
\code{\link{uni_harvard_secondary_2}},
\code{\link{uni_hopkins_accent_4}},
\code{\link{uni_hopkins_grayscale_5}},
\code{\link{uni_hopkins_identity_1}},
\code{\link{uni_hopkins_primary_2}},
\code{\link{uni_hopkins_secondary_3}},
\code{\link{uni_jena_1}},
\code{\link{uni_jena_2}},
\code{\link{uni_kiel_1}},
\code{\link{uni_kiel_2}},
\code{\link{uni_koeln_1}},
\code{\link{uni_koeln_2}},
\code{\link{uni_konstanz_1}},
\code{\link{uni_konstanz_2}},
\code{\link{uni_mannheim_1}},
\code{\link{uni_mannheim_2}},
\code{\link{uni_michigan_primary_1}},
\code{\link{uni_michigan_secondary_2}},
\code{\link{uni_northwestern_primary_1}},
\code{\link{uni_northwestern_secondary_2}},
\code{\link{uni_pal_1}},
\code{\link{uni_penn_logo_1}},
\code{\link{uni_penn_web_2}},
\code{\link{uni_princeton_0}},
\code{\link{uni_princeton_1}},
\code{\link{uni_princeton_2}},
\code{\link{uni_regensburg_1}},
\code{\link{uni_regensburg_3}},
\code{\link{uni_stanford_accent_2}},
\code{\link{uni_stanford_primary_1}},
\code{\link{uni_stanford_web_3}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_2_print}},
\code{\link{uni_ulm_1}},
\code{\link{uni_ulm_2}},
\code{\link{uni_yale_primary}},
\code{\link{uni_york_accent_4}},
\code{\link{uni_york_neutral_3}},
\code{\link{uni_york_primary_1}},
\code{\link{uni_york_secondary_2}}
}
\author{
\strong{unicol}, 2022-11-30.
}
\concept{university color palettes}
\keyword{datasets}
