% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UWaterloo.R
\docType{data}
\name{uwaterloo_main}
\alias{uwaterloo_main}
\title{Main colors of the University of Waterloo, CA}
\format{
An object of class \code{character} of length 9.
}
\source{
Color definitions are based on the University of Waterloo's 
\href{https://uwaterloo.ca/brand/}{Brand home}.
}
\usage{
uwaterloo_main
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uwaterloo_main} provides the main University colors 
of the \href{https://uwaterloo.ca/}{University of Waterloo}, Canada.
}
\details{
The main colors are gold/yellow, black/ grey and white at 4 Levels.
\code{"Gold/Yellow Level 1"} (defined as CMYK: 2 0 45 0, RGB: 255 255 170, HEX:	#FFFFAA), 
\code{"Gold/Yellow Level 2"} (defined as CMYK: 0 1 100 0, RGB:	255 234 41, HEX:	#FFEA3D), 
\code{"Gold/Yellow Level 3 (primary)"} (defined as CMYK: 0 14 80 0, RGB:	255 213 79, HEX:	#FDD54F), 
\code{"Gold/Yellow Level 4"} (defined as CMYK: 0 27 100 0, RGB:	228 180 41, HEX:	#E4B429), 
\code{"Black/Grey Level 1"} (defined as CMYK: 0 0 0 15, RGB:	223 223 223, HEX:	#DFDFDF), 
\code{"Black/Grey Level 2"} (defined as CMYK:0 0 0 35, RGB:	162 162 162, HEX:	#A2A2A2) 
\code{"Black/Grey Level 3"} (defined as CMYK:0 0 0 50, RGB:	120 120 120, HEX:	#787878), 
\code{"Black/Grey Level 4"} (defined as CMYK:0 0 0 100, RGB:	0 0 0, HEX:	#000000), and
\code{"White"} (defined as CMYK:0 0 0 0, RGB:	255 255 255, HEX:	#FFFFFF).


\code{uwaterloo_main} uses the HEX color definitions.
}
\examples{
uwaterloo_main
unikn::seecol(uwaterloo_main, main = "Main colors of the University of Waterloo") 
# view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{example_1}},
\code{\link{oxford_blog}},
\code{\link{oxford_brand}},
\code{\link{oxford_error}},
\code{\link{oxford_general}},
\code{\link{oxford_graduate}},
\code{\link{oxford_link}},
\code{\link{oxford_shades}},
\code{\link{oxford_socialmedia}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{uni_manchester}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_print}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{uwaterloo_arts}},
\code{\link{uwaterloo_engineering}},
\code{\link{uwaterloo_environment}},
\code{\link{uwaterloo_health}},
\code{\link{uwaterloo_math}},
\code{\link{uwaterloo_science}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{waikato}},
\code{\link{western_uni_1}},
\code{\link{yale_1}},
\code{\link{yeshiva}},
\code{\link{york_1}},
\code{\link{york_2}}
}
\author{
\strong{unicol}, 2023-06-29.
}
\concept{university color palettes}
\keyword{datasets}
