% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Waikato.R
\docType{data}
\name{waikato}
\alias{waikato}
\title{Colors of the University of Waikato, NZ}
\format{
An object of class \code{character} of length 8.
}
\source{
Color definitions are based on 
\href{https://www.waikato.ac.nz/web-guides/brand-standards/colours}{Waikato's brand standards}.
}
\usage{
waikato
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{waikato} provides the 8 official web colors 
of the \href{https://www.waikato.ac.nz/}{University of Waikato}, New Zealand.
}
\details{
The 8 official web colors are
\code{"waikato_light_red"} (defined as HEX #D40100),
\code{"waikato_medium_red"} (defined as HEX #BE0403),
\code{"waikato_dark_red"} (defined as HEX #AA0000),
\code{"waikato_light_yellow"} (defined as HEX #FAA61A),
\code{"waikato_darker_yellow"} (defined as HEX #E79000),
\code{"waikato_dark_grey"} (defined as HEX #353535),
\code{"waikato_blue_grey"} (defined as HEX #32373D), and
\code{"waikato_light_grey"} (defined as HEX #F2F2F2).

\code{waikato} uses the HEX color definitions.
}
\examples{
waikato
unikn::seecol(waikato, main = "University of Waikato") # view color palettes

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{example_1}},
\code{\link{oxford_error}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{uni_manchester}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_print}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{uwaterloo_arts}},
\code{\link{uwaterloo_engineering}},
\code{\link{uwaterloo_environment}},
\code{\link{uwaterloo_health}},
\code{\link{uwaterloo_main}},
\code{\link{uwaterloo_math}},
\code{\link{uwaterloo_science}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{western_uni_1}},
\code{\link{yale_1}},
\code{\link{yeshiva}},
\code{\link{york_1}},
\code{\link{york_2}}
}
\author{
\strong{unicol}, 2023-07-05.
}
\concept{university color palettes}
\keyword{datasets}
