% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/York.R
\docType{data}
\name{york_2}
\alias{york_2}
\title{Secondary colors of the York University, Canada}
\format{
An object of class \code{character} of length 8.
}
\source{
Color definitions are based on the
\href{https://www.yorku.ca/brand/wp-content/uploads/sites/18/2021/07/YORKU_Brand_Standards_July2021.pdf}{York Brand Standards (PDF)}.
}
\usage{
york_2
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{york_2} provides the five secondary colors 
of the \href{https://www.yorku.ca/}{York University}, Canada.
}
\details{
\code{"york_red_medium"}  (defined as RGB 175/13/26,   HEX #AF0D1A),
\code{"york_red_dark"}    (defined as RGB 129/0/1,     HEX #810001),
\code{"york_bright_blue"} (defined as RGB 58/194/239,  HEX #3AC2EF),
\code{"york_light_blue"}  (defined as RGB 172/230/248, HEX #ACE6F8),
\code{"york_light_grey"}  (defined as RGB 225/223/220, HEX #E1DFDC),
\code{"york_pewter"}      (defined as RGB 214/207/202, HEX #D6CFCA),
\code{"york_grey_medium"} (defined as RGB 183/174/169, HEX #B7AEA9), and
\code{"york_grey_dark"}   (defined as RGB 104/98/96,   HEX #686260).

\code{york_2} uses the HEX color definitions.
}
\examples{
york_2
unikn::seecol(york_2, main = "York University 2") # view color palette

}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.

Other university color palettes: 
\code{\link{example_1}},
\code{\link{oxford_error}},
\code{\link{pitt_1}},
\code{\link{pitt_2}},
\code{\link{uni_manchester}},
\code{\link{uni_stuttgart_1}},
\code{\link{uni_stuttgart_print}},
\code{\link{upenn_1}},
\code{\link{upenn_2}},
\code{\link{uwaterloo_arts}},
\code{\link{uwaterloo_engineering}},
\code{\link{uwaterloo_environment}},
\code{\link{uwaterloo_health}},
\code{\link{uwaterloo_main}},
\code{\link{uwaterloo_math}},
\code{\link{uwaterloo_science}},
\code{\link{vanderbilt_1}},
\code{\link{vanderbilt_2}},
\code{\link{vanderbilt_3}},
\code{\link{waikato}},
\code{\link{western_uni_1}},
\code{\link{yale_1}},
\code{\link{yeshiva}},
\code{\link{york_1}}
}
\author{
\strong{unicol}, 2023-06-24.
}
\concept{university color palettes}
\keyword{datasets}
